Fetch Download File Name

Gets either the file name specified by the user or the default value if that is null

Usage

ASM_fetch_dlfn(state, extension = ".zip")

Arguments

state

ASM state from ASM_fetch_state()

extension

File extension for the download (default: ".zip")

Value

character object with the download file name

Examples

# Creating a state object for testing
sess_res = ASM_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:631af84e28478adc1628dd2c9fb22662
#> → DW: State initialized
#> → DW: module checksum updated:f5866ed2a560dbc564dec9145314b519
#> → DW: module checksum updated:4911af9c225aac6361acea8f26bbfac9
#> → DW: module checksum updated:c9b0caeab52fb59e0dcbb88ee042729a
#> → DW: module checksum updated:b7f3e08819cabbd242f5bebe1290714b
#> → DW: module checksum updated:8d8944be2daf4ea8aa3b1c1c707cee45
#> → DW: module checksum updated:a6367d7e4e002eca11050774144105f8
#> → DW: module checksum updated:4a381cdf6502fe13dca9c009a25b4418
#> → DW: module checksum updated:1050496cc0eca17d51102e051220ec49
#> → DW: module checksum updated:df684a1d4fbbd8bd33749607f9b982c0
#> → DW: module checksum updated:e3054989800f1e8ca24386371b52accc
#> → DW: module checksum updated:3b43dbda29ded3a3a3ef514ccf0c0578
#> → DW: module checksum updated:550d10d9e424d549878ab4ba5343896e
#> → DW: module checksum updated:5d2a75ed77a22f4ff8cc2a0d2bf2e3e5
#> → DW: module checksum updated:0272b5040b80ff35adf3500fdb60d8c7
#> → DW: module checksum updated:173ecf7d96c0bbb0962c7846df0d4bca
#> → FG: State initialized
#> → FG: module checksum updated:91329389512102f6fab3f9986c1e5703
#> → FG: module checksum updated:fc57299d631b6abb1df8f9f86234f357
#> → FG: module checksum updated:5d6c6277e408e1fda1873c0b547cf090
#> → FG: module checksum updated:258f0401a5c4125fdeb85108bc0287f8
#> → FG: module checksum updated:064001e91fcb1cac1e84f2a3c7bf1365
#> → FG: module checksum updated:12bfdc81e8b855150c050744f7243575
#> → FG: module checksum updated:ac45899babd85840592d7d642a8135fb
#> → FG: module checksum updated:3fc56cc6900918b7df69bd76e824b206
#> → FG: module checksum updated:2de9b01542d545af1f34c56c9c84063c
#> → FG: module checksum updated:2eb604c0b43dbd893319a8efec046245
#> → FG: module checksum updated:d8d277d03b4a60fe64b04553101893b6
#> → FG: module checksum updated:993ea30d9e2bf7ff787dc7f0a5ce84f5
#> → FG: module checksum updated:fae1583ca0f27a62fbf0e208eb83ea34
#> → ASM: State initialized
state = sess_res$state
dlfn = ASM_fetch_dlfn(state)
dlfn
#> [1] "Save Analysis.zip"