Updates DM Module Checksum

Takes a DM state and updates the checksum used to trigger downstream updates

Usage

DM_update_checksum(state)

Arguments

state

DM state from DM_fetch_state()

Value

DM state object with the checksum updated

Examples

# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = DM_test_mksession(session=list())
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: setting word placeholders: 
#> → ASM:   -> setting docx ph: HEADERLEFT = left header
#> → ASM:   -> setting docx ph: HEADERRIGHT = right header
#> → ASM:   -> setting docx ph: FOOTERLEFT = left footer
#> → ASM: module isgood: TRUE
#> → DM: including file
#> → DM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DM:   dest:   file.path("config","report.docx")
#> → DM: including file
#> → DM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DM:   dest:   file.path("config","report.pptx")
#> → DM: including file
#> → DM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DM:   dest:   file.path("config","report.yaml")
#> → DM: module checksum updated: f280116ab6b4488df6de76aff513a83f
#> → DM: State initialized
#> → DM: ds checksum updated: 641fcbc6616d24d05acded4317a3bb12
#> → DM: module checksum updated: 78099100dae67cc0fee50517ea72a920
#> → DM: adding preload sources
#> → DM:  - file added: TEST_DATA.xlsx
#> → DM: added: TEST_DATA.xlsx
#> → DM:  - url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_PC.csv
#> → DM: url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_PC.csv
#> → DM:  - url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_EX.csv
#> → DM: url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_EX.csv
#> → DM:  - url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_DM.csv
#> → DM: url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_DM.csv
#> → DM: loading element idx: 1
#> → DM: ds checksum updated: dc8dcc9aaa810b766bbbb147ddcc5253
#> → DM: module checksum updated: acb6c1e1b6b307803f3393d7f8a8f13e
#> → DM: added element idx: 1
#> → DM: loading element idx: 2
#> → DM: ds checksum updated: e1abbd14bebe9ce1e8b08c8c38d112ed
#> → DM: module checksum updated: efbab51bb324955dcf1881d5863956f6
#> → DM: added element idx: 2
#> → DM: loading element idx: 3
#> → DM: ds checksum updated: 2f61e29a3ee91102fd41cb0fd2efcdf3
#> → DM: module checksum updated: 6a5d70cf870d24d84315d55d934ad07a
#> → DM: added element idx: 3
#> → DM: loading element idx: 4
#> → DM: ds checksum updated: 02b5018d85b363d7c93e1fba6631037b
#> → DM: module checksum updated: 1e20d2ed454e1e1dfb4a04fcec07ca88
#> → DM: added element idx: 4
#> → DM: module isgood: TRUE
#> → # preload_complete is good: TRUE ------------------------------------------
state   = sess_res$state
session = sess_res$session
input   = sess_res$input

# Creates a new empty element
state = DM_new_element(state)

# Delete the current element
state = DM_del_current_element(state)

# Fetch a list of the current element
element = DM_fetch_current_element(state)

# You can modify the element
element[["name"]] = "A more descriptive name"

# You can now place element back in the state
state = DM_set_current_element(state, element)
#> → DM: ds checksum updated: e8099603e42798f5a5f4e80a8d12c8e8
#> → DM: module checksum updated: 3b6251b0c5c688d7d26f2011981d4ffa

# This will pull the portion of the code associated with this module.
code = DM_fetch_code(state)

cat(code)
#> # Excel (file) ============================================================
#> DM_obj_1 <- rio::import(file="data/DM/TEST_DATA.xlsx", which ="DATA")
#> DM_obj_1 <- janitor::clean_names(DM_obj_1, case="none")
#> # STDM PC (url) ===========================================================
#> DM_obj_2 <- rio::import(file="https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_PC.csv")
#> DM_obj_2 <- janitor::clean_names(DM_obj_2, case="none")
#> # STDM EX (url) ===========================================================
#> DM_obj_3 <- rio::import(file="https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_EX.csv")
#> DM_obj_3 <- janitor::clean_names(DM_obj_3, case="none")
#> # STDM DM (url) ===========================================================
#> DM_obj_5 <- rio::import(file="https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_DM.csv")
#> DM_obj_5 <- janitor::clean_names(DM_obj_5, case="none")

# This forces and update of the module checksum
state = DM_update_checksum(state)