Append Report Elements

Takes the current state of the app and appends data views to an xlsx report object.

Usage

DW_append_report(state, rpt, rpttype, gen_code_only = FALSE)

Arguments

state

DW state from DW_fetch_state()

rpt

Report with the current content of the report which will be appended to in this function. For details on the structure see the documentation for FM_generate_report.

rpttype

Type of report to generate (supported "xlsx").

gen_code_only

Boolean value indicating that only code should be generated (FALSE).

Value

list containing the following elements

isgood: Return status of the function.

hasrptele: Boolean indicator if the module has any reportable elements.

code: Data wrangling R command.

msgs: Messages to be passed back to the user.

rpt: Report with any additions passed back to the user.

See also

Examples

# We need a state object to use below
sess_res = DW_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:b4bd907af4f3e3b571027d1230bf99f1
#> → DW: State initialized
#> → DW: module checksum updated:a9c58c8823c19140e4608cc9a0d95cc1
#> → DW: module checksum updated:ed1b4d28dcc64d59a538273bfd8cbfe7
#> → DW: module checksum updated:517754706893fd756b06d3b79bac8d5b
#> → DW: module checksum updated:7a46321aac57ea8514ea07caf4ccb4e7
#> → DW: module checksum updated:9929ec9ea2bc5d28342a19889261e65f
#> → DW: module checksum updated:c2a8ef415bd36e981fe4f88e5b499ba3
#> → DW: module checksum updated:0c32255ec1c5482a60938f787fdbcb17
#> → DW: module checksum updated:f096133b286d340e723bdda9ed0c9172
#> → DW: module checksum updated:d7df0ea03ec446323da2696b703c60b1
#> → DW: module checksum updated:87a8e10e49e67fb0b21a0bf96b398b6c
#> → DW: module checksum updated:209c3bcc8497eb25c2f3301a3d76006f
#> → DW: module checksum updated:5924e7a6a637d1bd43d099d6db3f4939
#> → DW: module checksum updated:f6e09337e2c4eedbbe705f0bc0b784d5
#> → DW: module checksum updated:cc00c0564573ad88f31beb55262f5443
#> → DW: module checksum updated:23b114d90922065a80aa8e665a158e6e
#> → DW: module checksum updated:db71e5e51679cd7673d2d951d6fda71e
#> → DW: module checksum updated:8111c3e76b1f5aabf6e2ebf0da2acf1e
#> → DW: module checksum updated:7dace004a38bcfba67f7d38deede9af8
#> → DW: module checksum updated:699a9d95559cf9e6d041b4901dff275c
#> → DW: module checksum updated:41d02c9ea6c806fecf527817498ec1c0
#> → DW: module checksum updated:35889d32f3526013f0f32b6268c94921
#> → DW: module checksum updated:69b9a3a363278ccde7439cf078d9eb6e
#> → DW: module checksum updated:a55b96d282553362ecf1cf8761be2261
#> → DW: module checksum updated:cae93dfa234c69be5467d9ec44bf54da
#> → DW: module checksum updated:9b88e64e8ec46ae2734b51432a29e6f1
state = sess_res$state

rpt = list(summary = list(), sheets=list())

rpt_res = DW_append_report(state,
 rpt     = rpt,
 rpttype = "xlsx")

# Shows if report elements are present
rpt_res$hasrptele
#> [1] TRUE

# Code chunk to generate report element
cat(paste(rpt_res$code, collapse="\n"))
#> rpt[["sheets"]][["DW_myDS_1"]]=DW_myDS_1
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_1",
#>     Description="Observations"
#>   )
#> )
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_1",
#>     Description="Observations"
#>   )
#> )
#> rpt[["sheets"]][["DW_myDS_2"]]=DW_myDS_2
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_2",
#>     Description="PK 3mg SD IV"
#>   )
#> )
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_2",
#>     Description="PK 3mg SD IV"
#>   )
#> )
#> rpt[["sheets"]][["DW_myDS_3"]]=DW_myDS_3
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_3",
#>     Description="PK 3mg MD SC (first dose)"
#>   )
#> )
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_3",
#>     Description="PK 3mg MD SC (first dose)"
#>   )
#> )
#> rpt[["sheets"]][["DW_myDS_4"]]=DW_myDS_4
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_4",
#>     Description="Parameters"
#>   )
#> )
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_4",
#>     Description="Parameters"
#>   )
#> )
#> rpt[["sheets"]][["DW_myDS_5"]]=DW_myDS_5
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_5",
#>     Description="PK/PD 3mg SD IV"
#>   )
#> )
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_5",
#>     Description="PK/PD 3mg SD IV"
#>   )
#> )
#> rpt[["sheets"]][["DW_myDS_6"]]=DW_myDS_6
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_6",
#>     Description="PK/PD 3mg MD SC (first dose)"
#>   )
#> )
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="DW_myDS_6",
#>     Description="PK/PD 3mg MD SC (first dose)"
#>   )
#> )

# Tabular summary of data views
rpt_res$rpt$summary
#>   Sheet_Name                  Description
#> 1  DW_myDS_1                 Observations
#> 2  DW_myDS_2                 PK 3mg SD IV
#> 3  DW_myDS_3    PK 3mg MD SC (first dose)
#> 4  DW_myDS_4                   Parameters
#> 5  DW_myDS_5              PK/PD 3mg SD IV
#> 6  DW_myDS_6 PK/PD 3mg MD SC (first dose)