Fetch Module Code

Fetches the code to generate results seen in the app

Usage

DW_fetch_code(state)

Arguments

state

DW state from DW_fetch_state()

Value

Character object vector with the lines of code and isgood)

Examples

# This will create a formods DW state object for the example
sess_res = DW_test_mksession(session=list())
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:48559d397570afc9fba8ad9b252fa75a
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DW:   dest:   file.path("config","report.docx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DW:   dest:   file.path("config","report.pptx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DW:   dest:   file.path("config","report.yaml")
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85
#> → DW: module checksum updated:a5d1d3b8a436befcfa119b5d72d3697b
#> → DW: module checksum updated:48d9e35c3f568c7a6f72914538d4ecd4
#> → DW: module checksum updated:680bde3fe602684945b93dda69abc020
#> → DW: module checksum updated:1f571c6eabb372eff6c1ef806344b2e4
#> → DW: module checksum updated:a1632ef431da47f98334a9281270a3fd
#> → DW: module checksum updated:518e56af743d086ea4b1b6070bf70472
#> → DW: module checksum updated:460fd6e3a4a46f56f17bde8ff4ef5c01
#> → DW: module checksum updated:12335ca5ced95462c2ed1df919a72d15
#> → DW: module checksum updated:0794933dac491313b32ca4bbfd996641
#> → DW: module checksum updated:75abe1a1cfc787fc68049b40c9e5749e
#> → DW: module checksum updated:070868f0125fd16b3e648dc05be84cd0
#> → DW: module checksum updated:1da012ee43c59bf8bedc7376511d179e
#> → DW: module checksum updated:6dfd06414929a762cc4f07d6d428289b
#> → DW: module checksum updated:83f7106b33c98e7650432749c16d7a36
#> → DW: module checksum updated:32652abd1996af932cbdf304cc509bd0
#> → DW: module checksum updated:dfa171056875bc474e03e0c911ace2b6
#> → DW: module checksum updated:5d4dfb7affd116c593a856034804aa04
#> → DW: module checksum updated:f47016aa2630055e8a1ae44ab79cabd2
#> → DW: module checksum updated:46cc2c9503359a39be98c3c0b889851a
#> → DW: module checksum updated:58ed3ad9e797e77eba05a2576b78e50f
#> → DW: module checksum updated:aaf6ebcd2be24f05cd9b4776ef042986
#> → DW: module checksum updated:afbbf99a8ed5f01191217b8370603a9d
#> → DW: module checksum updated:086f760681a2b47a9a1dfd572f876fcd
#> → DW: module checksum updated:8c045cb5e0fe9bd74e84d87d3c6725b5
#> → DW: module checksum updated:779a7d3c81e58ea4c8a756c2e422fee0
#> → DW: module checksum updated:104438a337e6617c4504477506e90863
#> → DW: module checksum updated:1bd57823c76715ccceab0be02ed7a79e
#> → DW: module checksum updated:8b901fc11d897dd27c774accf0eb7097
#> → DW: module checksum updated:7d55b640df1207dcdc56a10c9b8d8b87
#> → DW: module checksum updated:8d91f192ea994a133c8e2fbb7ce7c11a
#> → DW: module checksum updated:d5c81022154ffb7347d7a30d66a38f60
state   = sess_res$state
code = DW_fetch_code(state)
cat(code)
#> # Observations ============================================================
#> DW_myDS_1 = myDS
#> DW_myDS_1 = dplyr::filter(DW_myDS_1,EVID == 0)
#> DW_myDS_1 = dplyr::filter(DW_myDS_1,DV != 0)
#> DW_myDS_1 = dplyr::mutate(DW_myDS_1,IDCMT = paste0(ID, ', ', CMT))
#> 
#> # PK 3mg SD IV ============================================================
#> DW_myDS_2 = myDS
#> DW_myDS_2 = dplyr::filter(DW_myDS_2,EVID == 0)
#> DW_myDS_2 = dplyr::filter(DW_myDS_2,DV != 0)
#> DW_myDS_2 = dplyr::filter(DW_myDS_2,Cohort %in% c("SD 3 mg IV"))
#> DW_myDS_2 = dplyr::filter(DW_myDS_2,CMT %in% c("C_ng_ml"))
#> 
#> # PK 3mg MD SC (first dose) ===============================================
#> DW_myDS_3 = myDS
#> DW_myDS_3 = dplyr::filter(DW_myDS_3,EVID == 0)
#> DW_myDS_3 = dplyr::filter(DW_myDS_3,DV != 0)
#> DW_myDS_3 = dplyr::filter(DW_myDS_3,Cohort %in% c("MD 3 mg SC"))
#> DW_myDS_3 = dplyr::filter(DW_myDS_3,DOSE_NUM == 1)
#> DW_myDS_3 = dplyr::filter(DW_myDS_3,CMT %in% c("C_ng_ml"))
#> 
#> # Parameters ==============================================================
#> DW_myDS_4 = myDS
#> DW_myDS_4 = dplyr::group_by(DW_myDS_4,ID)
#> DW_myDS_4 = dplyr::filter(DW_myDS_4,row_number()==1)
#> DW_myDS_4 = dplyr::select(DW_myDS_4,ID, DOSE, DOSE_STR, Cohort, ROUTE, ka, CL, Vc, Vp, Q)
#> DW_myDS_4 = tidyr::pivot_longer(DW_myDS_4, cols      = c("ka", "CL", "Vc", "Vp", "Q"), names_to  = "parameter", values_to = "values")
#> 
#> # PK/PD 3mg SD IV w/BQL ===================================================
#> DW_myDS_5 = myDS
#> DW_myDS_5 = dplyr::filter(DW_myDS_5,EVID == 0)
#> DW_myDS_5 = dplyr::filter(DW_myDS_5,Cohort %in% c("SD 3 mg IV"))
#> 
#> # PK/PD 3mg MD SC (first dose) w/BQL ======================================
#> DW_myDS_6 = myDS
#> DW_myDS_6 = dplyr::filter(DW_myDS_6,EVID == 0)
#> DW_myDS_6 = dplyr::filter(DW_myDS_6,Cohort %in% c("MD 3 mg SC"))
#> DW_myDS_6 = dplyr::filter(DW_myDS_6,DOSE_NUM == 1)
#> 
#> # PK 3mg SD IV w/BQL ======================================================
#> DW_myDS_7 = myDS
#> DW_myDS_7 = dplyr::filter(DW_myDS_7,EVID == 0)
#> DW_myDS_7 = dplyr::filter(DW_myDS_7,Cohort %in% c("SD 3 mg IV"))
#> DW_myDS_7 = dplyr::filter(DW_myDS_7,CMT %in% c("C_ng_ml"))