Fetch Data Wrangling State

Merges default app options with the changes made in the UI

Usage

DW_fetch_state(
  id,
  input,
  session,
  FM_yaml_file,
  MOD_yaml_file,
  id_UD,
  react_state
)

Arguments

id

Shiny module ID

input

Shiny input variable

session

Shiny session variable

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

id_UD

ID string for the upload data module used to handle uploads or the name of the list element in react_state where the data set is stored.

react_state

Variable passed to server to allow reaction outside of module (NULL)

Value

List containing the current state of the DM module including default values from the yaml file as well as any changes made by the user. The structure of the list is defined below.

yaml: Contents of the yaml file.

MC: Module components of the yaml file.

DW: Data wrangling state

isgood: Boolean status of the state. FALSE if the dataset identified by id_UD is bad.

checksum: MD5 sum indicating if there was a change in the datasets within the view. Use this to trigger updates in respose to changes in this module.

button_counters: List of counters to detect button clicks.

code_previous: Loading code from the UD field.

current_view: View id of the current active data wrangling view.

UD: Copy of the "UD" field of the id_UD from the react_state input.

ui: Current value of form elements in the UI

ui_hold: List of hold elements to disable updates before a full ui referesh is complete.

view_cntr: Counter for tracking view ids, value contains the id of the last view created.

views: List of data wrangling views. Each view has the following structure:

isgood: Boolean status of the data view. False if evaluation fails

id: Character id (view_idx)

idx: Numeric id (1)

view_ds_object_name: Object name for this data view

code_previous: Code to load data and assign to view object.

key: User key (short description)

WDS: Current value of the data view with all of the successful commands in elements_table evaluated.

elements_table: Table of data wrangling elements.

checksum: MD5 sum of WDS

code: Code to generate WDS from start to finish

code_dw_only: Code for jsut the wrangling portion.

MOD_TYPE: Character data containing the type of module "DW"

id: Character data containing the module id

FM_yaml_file: App configuration file with FM as main section.

MOD_yaml_file: Module configuration file with MC as main section. module in the session variable.

Examples

# Within shiny both session and input variables will exist, 
# this creates examples here for testing purposes:
sess_res = DW_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:631af84e28478adc1628dd2c9fb22662
#> → DW: State initialized
#> → DW: module checksum updated:f5866ed2a560dbc564dec9145314b519
#> → DW: module checksum updated:4911af9c225aac6361acea8f26bbfac9
#> → DW: module checksum updated:c9b0caeab52fb59e0dcbb88ee042729a
#> → DW: module checksum updated:d53e4eeadacffd272f5e584700d4e5c6
#> → DW: module checksum updated:25e606a4b68a27b1b111d4ce7d6ed944
#> → DW: module checksum updated:cf7553b85f514e80b0495bac005e30a3
#> → DW: module checksum updated:9237881640f3d22346fbce5f152d9e69
#> → DW: module checksum updated:928a817e1c9205fbbe273de91af26c6a
session = sess_res$session
input   = sess_res$input

# Configuration files
FM_yaml_file  = system.file(package = "formods", "templates", "formods.yaml")
MOD_yaml_file = system.file(package = "formods", "templates", "DW.yaml")

# We need to specify both the DW module id as well as the
# id of the UD module that feeds into it.
id    = "DW"
id_UD = "UD"

# Creating an empty state object
state = DW_fetch_state(id              = id,           
                       input           = input, 
                       session         = session,
                       FM_yaml_file    = FM_yaml_file, 
                       MOD_yaml_file   = MOD_yaml_file,
                       id_UD           = "UD",
                       react_state     = NULL)