Initialize DW Module State

Creates a list of the initialized module state

Usage

DW_init_state(FM_yaml_file, MOD_yaml_file, id, id_UD, session)

Arguments

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

id

Shiny module ID

id_UD

ID string for the upload data module used to handle uploads or the name of the list element in react_state where the data set is stored.

session

Shiny session variable module (NULL)

Value

list containing an empty DW state

Examples

# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = DW_test_mksession()
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: setting word placeholders: 
#> → ASM:   -> setting docx ph: HEADERLEFT = left header
#> → ASM:   -> setting docx ph: HEADERRIGHT = right header
#> → ASM:   -> setting docx ph: FOOTERLEFT = left footer
#> → ASM: module isgood: TRUE
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
#> → UD: module isgood: TRUE
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DW:   dest:   file.path("config","report.docx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DW:   dest:   file.path("config","report.pptx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DW:   dest:   file.path("config","report.yaml")
#> → DW: State initialized
#> → DW: module checksum updated:5b0f0b05ee3ac7336a74c564bb6efdad
#> → DW: loading data view idx: 1
#> → DW: setting name: Observations
#> → DW: module checksum updated:1ac3e0afcc601f848943f92b854b3830
#> → DW:   -> filter
#> → DW: module checksum updated:67aff6e926eba73b3ecb361d26624844
#> → DW:   -> filter
#> → DW: module checksum updated:0234f6d458ef7487a3b2e991b2d0957b
#> → DW:   -> mutate
#> → DW: module checksum updated:37d0958f042076b974fdd3f894157822
#> → DW: loading data view idx: 2
#> → DW: setting name: PK 3mg SD IV
#> → DW: module checksum updated:4e7ca05728d66df3adfcf87387f9543a
#> → DW:   -> filter
#> → DW: module checksum updated:2d95bc56262f46b7f135dd33fc4722b3
#> → DW:   -> filter
#> → DW: module checksum updated:b8a30837145926d2d9cc6f788c10b7ed
#> → DW:   -> filter
#> → DW: module checksum updated:24b554ee8736ebc84979c896ac2e93c3
#> → DW:   -> filter
#> → DW: module checksum updated:daad88263faf7cbf90871ed00a0dd275
#> → DW: loading data view idx: 3
#> → DW: setting name: PK 3mg SD IV (NCA)
#> → DW: module checksum updated:a7b3ea35cdba272682716aa4619d3983
#> → DW:   -> filter
#> → DW: module checksum updated:e8d2f77c0bd731995d845ee8be623e22
#> → DW:   -> filter
#> → DW: module checksum updated:24cc7b591ca756a0abf560fb5295e04d
#> → DW:   -> filter
#> → DW: module checksum updated:f79ef26bdb4dac420c24650b9e96f721
#> → DW: loading data view idx: 4
#> → DW: setting name: PKPD 3mg SD IV (NCA)
#> → DW: module checksum updated:c737e8b3a28ba600a83ddf58a969c90a
#> → DW:   -> filter
#> → DW: module checksum updated:4c0d054f8ec47ae97fd1c650d8ddad55
#> → DW:   -> filter
#> → DW: module checksum updated:1e3d65ec1bd6ee5850d67bd2250e3223
#> → DW:   -> filter
#> → DW: module checksum updated:1e3d65ec1bd6ee5850d67bd2250e3223
#> → DW: module isgood: TRUE
session = sess_res$session
input   = sess_res$input

state = DW_init_state(
   FM_yaml_file  = system.file(package = "formods",
                               "templates",
                               "formods.yaml"),
   MOD_yaml_file = system.file(package = "formods",
                               "templates",
                               "DW.yaml"),
   id              = "DW",
   id_UD           = "UD",
   session         = session)
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DW:   dest:   file.path("config","report.docx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DW:   dest:   file.path("config","report.pptx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DW:   dest:   file.path("config","report.yaml")
#> → DW: State initialized
#> → DW: module checksum updated:5b0f0b05ee3ac7336a74c564bb6efdad

state
#> $yaml
#> $yaml$FM
#> $yaml$FM$include
#> $yaml$FM$include$files
#> $yaml$FM$include$files[[1]]
#> $yaml$FM$include$files[[1]]$file
#> $yaml$FM$include$files[[1]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\")"
#> 
#> $yaml$FM$include$files[[1]]$file$dest
#> [1] "file.path(\"config\",\"report.docx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[2]]
#> $yaml$FM$include$files[[2]]$file
#> $yaml$FM$include$files[[2]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\")"
#> 
#> $yaml$FM$include$files[[2]]$file$dest
#> [1] "file.path(\"config\",\"report.pptx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[3]]
#> $yaml$FM$include$files[[3]]$file
#> $yaml$FM$include$files[[3]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\")"
#> 
#> $yaml$FM$include$files[[3]]$file$dest
#> [1] "file.path(\"config\",\"report.yaml\")"
#> 
#> 
#> 
#> 
#> 
#> $yaml$FM$deployed
#> [1] FALSE
#> 
#> $yaml$FM$code
#> $yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $yaml$FM$code$gen_file
#> [1] "run_analysis.R"
#> 
#> $yaml$FM$code$gen_preamble
#> [1] "# formods automated output ------------------------------------------------\n# https://formods.ubiquity.tools/\nrm(list=ls())"
#> 
#> $yaml$FM$code$packages
#> [1] "onbrand" "writexl"
#> 
#> 
#> $yaml$FM$notifications
#> $yaml$FM$notifications$config
#> $yaml$FM$notifications$config$success
#> $yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $yaml$FM$notifications$config$failure
#> $yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $yaml$FM$notifications$config$info
#> $yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $yaml$FM$notifications$config$warning
#> $yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $yaml$FM$reporting
#> $yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $yaml$FM$reporting$content_init
#> $yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.docx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> $yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.pptx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> 
#> $yaml$FM$reporting$phs
#> $yaml$FM$reporting$phs[[1]]
#> $yaml$FM$reporting$phs[[1]]$name
#> [1] "HEADERLEFT"
#> 
#> $yaml$FM$reporting$phs[[1]]$location
#> [1] "header"
#> 
#> $yaml$FM$reporting$phs[[1]]$value
#> [1] ""
#> 
#> $yaml$FM$reporting$phs[[1]]$tooltip
#> [1] "left header text"
#> 
#> 
#> $yaml$FM$reporting$phs[[2]]
#> $yaml$FM$reporting$phs[[2]]$name
#> [1] "HEADERRIGHT"
#> 
#> $yaml$FM$reporting$phs[[2]]$location
#> [1] "header"
#> 
#> $yaml$FM$reporting$phs[[2]]$value
#> [1] ""
#> 
#> $yaml$FM$reporting$phs[[2]]$tooltip
#> [1] "right header text"
#> 
#> 
#> $yaml$FM$reporting$phs[[3]]
#> $yaml$FM$reporting$phs[[3]]$name
#> [1] "FOOTERLEFT"
#> 
#> $yaml$FM$reporting$phs[[3]]$location
#> [1] "footer"
#> 
#> $yaml$FM$reporting$phs[[3]]$value
#> [1] ""
#> 
#> $yaml$FM$reporting$phs[[3]]$tooltip
#> [1] "left footer text"
#> 
#> 
#> 
#> $yaml$FM$reporting$phs_formatting
#> $yaml$FM$reporting$phs_formatting$width
#> [1] "100%"
#> 
#> $yaml$FM$reporting$phs_formatting$tt_position
#> [1] "left"
#> 
#> $yaml$FM$reporting$phs_formatting$tt_size
#> [1] "medium"
#> 
#> 
#> 
#> $yaml$FM$ui
#> $yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $yaml$FM$ui$select_size
#> [1] 10
#> 
#> $yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $yaml$FM$data_meta
#> $yaml$FM$data_meta$data_header
#> [1] "<span style='color:===COLOR==='><b>===NAME===</b><br/><font size='-3'>===LABEL===</font></span>"
#> 
#> $yaml$FM$data_meta$subtext
#> [1] "===LABEL===: ===RANGE==="
#> 
#> $yaml$FM$data_meta$many_sep
#> [1] ",...,"
#> 
#> $yaml$FM$data_meta$data_types
#> $yaml$FM$data_meta$data_types$character
#> $yaml$FM$data_meta$data_types$character$color
#> [1] "#DD4B39"
#> 
#> $yaml$FM$data_meta$data_types$character$label
#> [1] "text"
#> 
#> 
#> $yaml$FM$data_meta$data_types$double
#> $yaml$FM$data_meta$data_types$double$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$double$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$integer
#> $yaml$FM$data_meta$data_types$integer$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$integer$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$other
#> $yaml$FM$data_meta$data_types$other$color
#> [1] "black"
#> 
#> $yaml$FM$data_meta$data_types$other$label
#> [1] "other"
#> 
#> 
#> 
#> 
#> $yaml$FM$workflows
#> $yaml$FM$workflows$example
#> $yaml$FM$workflows$example$group
#> [1] "Examples"
#> 
#> $yaml$FM$workflows$example$desc
#> [1] "Example Workflow"
#> 
#> $yaml$FM$workflows$example$require_ds
#> [1] TRUE
#> 
#> $yaml$FM$workflows$example$preload
#> [1] "file.path('.', 'example.yaml')"
#> 
#> 
#> 
#> $yaml$FM$labels
#> $yaml$FM$labels$ui_label
#> [1] "put labels here"
#> 
#> 
#> $yaml$FM$user_files
#> $yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $yaml$FM$logging
#> $yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $MC
#> $MC$module
#> $MC$module$type
#> [1] "DW"
#> 
#> $MC$module$depends
#> $MC$module$depends$id_ASM
#> [1] "ASM"
#> 
#> $MC$module$depends$id_UD
#> [1] "UD"
#> 
#> 
#> 
#> $MC$compact
#> $MC$compact$code
#> [1] TRUE
#> 
#> $MC$compact$preview
#> [1] TRUE
#> 
#> $MC$compact$clip
#> [1] TRUE
#> 
#> 
#> $MC$reporting
#> $MC$reporting$enabled
#> [1] TRUE
#> 
#> $MC$reporting$priority
#> [1] 1
#> 
#> 
#> $MC$notifications
#> $MC$notifications$new_dw_element
#> [1] "Added element: ===ACTION=== (===DESC===)"
#> 
#> $MC$notifications$del_dw_element
#> [1] "Removed element: ===ACTION=== (===DESC===)"
#> 
#> 
#> $MC$tooltips
#> $MC$tooltips$include
#> [1] TRUE
#> 
#> $MC$tooltips$show_code
#> [1] "Show data wrangling code"
#> 
#> $MC$tooltips$dw_elements
#> [1] "Current wrangling elements"
#> 
#> 
#> $MC$labels
#> $MC$labels$tab_main
#> [1] "Wrangle"
#> 
#> $MC$labels$tab_code
#> [1] "Code"
#> 
#> $MC$labels$add_element
#> [1] "Add Operation/Action"
#> 
#> $MC$labels$select_dw_views
#> [1] "Current data view"
#> 
#> $MC$labels$fds_filter_column
#> [1] "Filter by Column"
#> 
#> $MC$labels$fds_filter_operator
#> [1] "Condition"
#> 
#> $MC$labels$fds_filter_rhs
#> [1] "Matching"
#> 
#> $MC$labels$fds_mutate_column
#> [1] "New or current column to alter"
#> 
#> $MC$labels$fds_mutate_rhs
#> [1] "New Value"
#> 
#> $MC$labels$fds_rename_column
#> [1] "Curret Column"
#> 
#> $MC$labels$fds_rename_rhs
#> [1] "New Name"
#> 
#> $MC$labels$fds_group_column
#> [1] "Columns to group"
#> 
#> $MC$labels$fds_longer_column
#> [1] "Columns to make longer"
#> 
#> $MC$labels$fds_longer_names
#> [1] "New column to contain names"
#> 
#> $MC$labels$fds_longer_values
#> [1] "New column to contain values"
#> 
#> $MC$labels$fds_wider_names
#> [1] "Column with names"
#> 
#> $MC$labels$fds_wider_values
#> [1] "Column with values"
#> 
#> $MC$labels$new_dw
#> [1] "New"
#> 
#> $MC$labels$del_dw
#> [1] "Delete"
#> 
#> $MC$labels$save_dw
#> [1] "Save"
#> 
#> $MC$labels$copy_dw
#> [1] "Copy"
#> 
#> $MC$labels$clip_dw
#> [1] "Code"
#> 
#> $MC$labels$ungroup_data
#> [1] "Removing current grouping"
#> 
#> $MC$labels$keep_onerow
#> [1] "Keep first row in current grouping"
#> 
#> $MC$labels$no_dataset
#> [1] "No dataset available to wrangle. You need to load a dataset to use this module."
#> 
#> $MC$labels$no_dw_elements
#> [1] "No data wrangling elements defined yet"
#> 
#> $MC$labels$busy
#> $MC$labels$busy$dv_update
#> [1] "Updating data views. Please be patient..."
#> 
#> 
#> 
#> $MC$errors
#> $MC$errors$fds_filter_column
#> [1] "Filter column not specified."
#> 
#> $MC$errors$fds_filter_operator
#> [1] "Filter condition not specified."
#> 
#> $MC$errors$fds_filter_rhs
#> [1] "Filter RHS not specified."
#> 
#> $MC$errors$fds_mutate_column
#> [1] "Mutate column not specified."
#> 
#> $MC$errors$fds_mutate_rhs
#> [1] "Mutate RHS not specified."
#> 
#> $MC$errors$fds_rename_column
#> [1] "Column to rename not specified."
#> 
#> $MC$errors$fds_rename_rhs
#> [1] "New column name not specified."
#> 
#> $MC$errors$fds_group_column
#> [1] "No columns to group specified."
#> 
#> $MC$errors$fds_select_column
#> [1] "You must select at least one column to keep."
#> 
#> $MC$errors$fds_longer_column
#> [1] "You must select at least one column to make longer."
#> 
#> $MC$errors$fds_longer_names
#> [1] "You must supply a column for the names of the new longer dataset"
#> 
#> $MC$errors$fds_longer_values
#> [1] "You must supply a column for the values of the new longer dataset"
#> 
#> $MC$errors$fds_wider_names
#> [1] "You must select a column with the names."
#> 
#> $MC$errors$`fds_wider_value s`
#> [1] "You must select a column with the values."
#> 
#> $MC$errors$current_key
#> [1] "Short description"
#> 
#> $MC$errors$current_key_empty
#> [1] "The data view description field cannot be empty"
#> 
#> $MC$errors$unknown_action
#> [1] "The selection action is not supported."
#> 
#> $MC$errors$no_rows
#> [1] "There are no rows in the data set."
#> 
#> $MC$errors$element_not_added
#> [1] "Unable to add the data wrangling element:"
#> 
#> $MC$errors$fix_bad_element
#> [1] "There is a bad element in the data wrangling chain,\nyou need to fix that before adding new elements"
#> 
#> 
#> $MC$operators
#> $MC$operators[[1]]
#> $MC$operators[[1]]$type
#> [1] "factor"
#> 
#> $MC$operators[[1]]$text
#> [1] "is one of"
#> 
#> $MC$operators[[1]]$rop
#> [1] "%in%"
#> 
#> 
#> $MC$operators[[2]]
#> $MC$operators[[2]]$type
#> [1] "factor"
#> 
#> $MC$operators[[2]]$text
#> [1] "is not"
#> 
#> $MC$operators[[2]]$rop
#> [1] "!%in%"
#> 
#> 
#> $MC$operators[[3]]
#> $MC$operators[[3]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[3]]$text
#> [1] "is within the range"
#> 
#> $MC$operators[[3]]$rop
#> [1] "within"
#> 
#> 
#> $MC$operators[[4]]
#> $MC$operators[[4]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[4]]$text
#> [1] "is equal to"
#> 
#> $MC$operators[[4]]$rop
#> [1] "=="
#> 
#> 
#> $MC$operators[[5]]
#> $MC$operators[[5]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[5]]$text
#> [1] "is not equal to"
#> 
#> $MC$operators[[5]]$rop
#> [1] "!="
#> 
#> 
#> 
#> $MC$actions
#> $MC$actions$filter
#> $MC$actions$filter$choice
#> [1] "Filter"
#> 
#> $MC$actions$filter$subtext
#> [1] "Filter data"
#> 
#> 
#> $MC$actions$mutate
#> $MC$actions$mutate$choice
#> [1] "Alter"
#> 
#> $MC$actions$mutate$subtext
#> [1] "Create new column or change existing column"
#> 
#> 
#> $MC$actions$rename
#> $MC$actions$rename$choice
#> [1] "Rename"
#> 
#> $MC$actions$rename$subtext
#> [1] "Rename column"
#> 
#> 
#> $MC$actions$group
#> $MC$actions$group$choice
#> [1] "Group"
#> 
#> $MC$actions$group$subtext
#> [1] "Group subsequent operations by columns"
#> 
#> 
#> $MC$actions$ungroup
#> $MC$actions$ungroup$choice
#> [1] "Ungroup"
#> 
#> $MC$actions$ungroup$subtext
#> [1] "Removing groping"
#> 
#> 
#> $MC$actions$longer
#> $MC$actions$longer$choice
#> [1] "Elongate"
#> 
#> $MC$actions$longer$subtext
#> [1] "Make Longer: rows to columns"
#> 
#> 
#> $MC$actions$wider
#> $MC$actions$wider$choice
#> [1] "Widen"
#> 
#> $MC$actions$wider$subtext
#> [1] "Make Wider: columns to rows"
#> 
#> 
#> $MC$actions$onerow
#> $MC$actions$onerow$choice
#> [1] "First Row"
#> 
#> $MC$actions$onerow$subtext
#> [1] "Keep only first row of grouped data"
#> 
#> 
#> $MC$actions$select
#> $MC$actions$select$choice
#> [1] "Select"
#> 
#> $MC$actions$select$subtext
#> [1] "Choose columns to keep"
#> 
#> 
#> 
#> $MC$formatting
#> $MC$formatting$preview
#> $MC$formatting$preview$width
#> [1] 800
#> 
#> $MC$formatting$preview$height
#> [1] 400
#> 
#> 
#> $MC$formatting$dw_elements
#> $MC$formatting$dw_elements$width
#> [1] 600
#> 
#> $MC$formatting$dw_elements$height
#> [1] 200
#> 
#> 
#> $MC$formatting$current_key
#> $MC$formatting$current_key$width
#> [1] 300
#> 
#> 
#> $MC$formatting$code
#> $MC$formatting$code$width
#> [1] 800
#> 
#> $MC$formatting$code$height
#> [1] 300
#> 
#> 
#> $MC$formatting$button_dw_del
#> $MC$formatting$button_dw_del$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_del$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_del$tooltip
#> [1] "Delete the current data view."
#> 
#> $MC$formatting$button_dw_del$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_clip
#> $MC$formatting$button_dw_clip$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_clip$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_clip$tooltip
#> [1] "Copy code to generate current data view to the clipboard."
#> 
#> $MC$formatting$button_dw_clip$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$select_dw_views
#> $MC$formatting$select_dw_views$width
#> [1] 200
#> 
#> 
#> $MC$formatting$button_dw_save
#> $MC$formatting$button_dw_save$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_save$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_save$tooltip
#> [1] "Save changes to data view name"
#> 
#> $MC$formatting$button_dw_save$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_copy
#> $MC$formatting$button_dw_copy$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_copy$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_copy$tooltip
#> [1] "Copy current data view to a new data view."
#> 
#> $MC$formatting$button_dw_copy$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_new
#> $MC$formatting$button_dw_new$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_new$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_new$tooltip
#> [1] "Create a new empty data view."
#> 
#> $MC$formatting$button_dw_new$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_add_element
#> $MC$formatting$button_dw_add_element$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_add_element$block
#> [1] TRUE
#> 
#> 
#> 
#> $MC$code
#> $MC$code$readOnly
#> [1] TRUE
#> 
#> $MC$code$mode
#> [1] "r"
#> 
#> $MC$code$packages
#> [1] "dplyr" "tidyr"
#> 
#> $MC$code$preamble
#> [1] "# Data wrangling ----------------------------------------------------------"
#> 
#> 
#> $MC$ds_object_name
#> [1] "DW_myDS"
#> 
#> 
#> $DW
#> $DW$button_counters
#> $DW$button_counters$button_dw_add_element
#> [1] 0
#> 
#> $DW$button_counters$button_dw_new
#> [1] 0
#> 
#> $DW$button_counters$button_dw_del
#> [1] 0
#> 
#> $DW$button_counters$button_dw_save
#> [1] 0
#> 
#> $DW$button_counters$button_dw_copy
#> [1] 0
#> 
#> 
#> $DW$ui_hold
#> $DW$ui_hold$hot_dw_elements
#> [1] FALSE
#> 
#> $DW$ui_hold$current_key
#> [1] FALSE
#> 
#> $DW$ui_hold$select_dw_views
#> [1] FALSE
#> 
#> 
#> $DW$ui_ids
#>  [1] "hot_dw_elements"            "button_dw_add_element"     
#>  [3] "button_dw_new"              "button_dw_del"             
#>  [5] "button_dw_copy"             "button_dw_save"            
#>  [7] "select_fds_filter_column"   "select_fds_filter_operator"
#>  [9] "fds_filter_rhs"             "select_fds_mutate_column"  
#> [11] "select_fds_mutate_rhs"      "select_fds_rename_column"  
#> [13] "fds_rename_rhs"             "current_key"               
#> [15] "select_fds_group_column"    "select_fds_select_column"  
#> [17] "select_fds_longer_column"   "select_fds_longer_values"  
#> [19] "select_fds_longer_names"    "select_fds_wider_values"   
#> [21] "select_fds_wider_names"     "select_dw_views"           
#> [23] "select_dw_element"         
#> 
#> $DW$isgood
#> [1] TRUE
#> 
#> $DW$op_choices
#> $DW$op_choices$factor
#> is one of    is not 
#>    "%in%"   "!%in%" 
#> 
#> $DW$op_choices$not_factor
#> is within the range         is equal to     is not equal to 
#>            "within"                "=="                "!=" 
#> 
#> 
#> $DW$view_cntr
#> [1] 1
#> 
#> $DW$UD
#> $DW$UD$isgood
#> [1] TRUE
#> 
#> $DW$UD$checksum
#> [1] "897d952fecbc804999396a96f9df4b20"
#> 
#> $DW$UD$contents
#> # A tibble: 10,680 × 33
#>       ID TIME_DY TIME_HR NTIME_DY NTIME_HR  TIME     AMT    DV RECDESC     DOSE
#>    <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl>   <dbl> <dbl> <chr>      <dbl>
#>  1     1       0       0        0        0     0 3000000    0  dose     3000000
#>  2     1       0       0        0        0     0       0 1250  BM_ng_ml 3000000
#>  3     1       0       0        0        0     0       0    0  C_ng_ml  3000000
#>  4     1       1      24        1       24    24       0 1647  BM_ng_ml 3000000
#>  5     1       1      24        1       24    24       0  690. C_ng_ml  3000000
#>  6     1       2      48        2       48    48       0 1947  BM_ng_ml 3000000
#>  7     1       2      48        2       48    48       0  588. C_ng_ml  3000000
#>  8     1       3      72        3       72    72       0 2170  BM_ng_ml 3000000
#>  9     1       3      72        3       72    72       0  506. C_ng_ml  3000000
#> 10     1       4      96        4       96    96       0 2332  BM_ng_ml 3000000
#> # ℹ 10,670 more rows
#> # ℹ 23 more variables: DOSE_STR <chr>, Cohort <chr>, ROUTE <chr>, WT <dbl>,
#> #   SUBTYPE <chr>, SEX <chr>, F1 <dbl>, ka <dbl>, CL <dbl>, Vc <dbl>, Vp <dbl>,
#> #   Q <dbl>, BM_IC <dbl>, kdeg_BM <dbl>, Emax <dbl>, EC50 <dbl>, CMT <chr>,
#> #   DUR <dbl>, EVID <dbl>, NDose <chr>, DOSE_NUM <dbl>, SUBTYPE_ID <dbl>,
#> #   SEX_ID <dbl>
#> 
#> $DW$UD$object_name
#> [1] "myDS"
#> 
#> 
#> $DW$code_previous
#> [1] "myDS = readxl::read_excel(path=\"TEST_DATA.xlsx\", sheet=\"DATA\")\nmyDS = janitor::clean_names(myDS, case=\"none\")"
#> 
#> $DW$views
#> $DW$views$view_1
#> $DW$views$view_1$isgood
#> [1] TRUE
#> 
#> $DW$views$view_1$id
#> [1] "view_1"
#> 
#> $DW$views$view_1$idx
#> [1] 1
#> 
#> $DW$views$view_1$view_ds_object_name
#> [1] "DW_myDS_1"
#> 
#> $DW$views$view_1$code_previous
#> [1] "DW_myDS_1 = myDS"
#> 
#> $DW$views$view_1$key
#> [1] "data_view_1"
#> 
#> $DW$views$view_1$WDS
#> # A tibble: 10,680 × 33
#>       ID TIME_DY TIME_HR NTIME_DY NTIME_HR  TIME     AMT    DV RECDESC     DOSE
#>    <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl>   <dbl> <dbl> <chr>      <dbl>
#>  1     1       0       0        0        0     0 3000000    0  dose     3000000
#>  2     1       0       0        0        0     0       0 1250  BM_ng_ml 3000000
#>  3     1       0       0        0        0     0       0    0  C_ng_ml  3000000
#>  4     1       1      24        1       24    24       0 1647  BM_ng_ml 3000000
#>  5     1       1      24        1       24    24       0  690. C_ng_ml  3000000
#>  6     1       2      48        2       48    48       0 1947  BM_ng_ml 3000000
#>  7     1       2      48        2       48    48       0  588. C_ng_ml  3000000
#>  8     1       3      72        3       72    72       0 2170  BM_ng_ml 3000000
#>  9     1       3      72        3       72    72       0  506. C_ng_ml  3000000
#> 10     1       4      96        4       96    96       0 2332  BM_ng_ml 3000000
#> # ℹ 10,670 more rows
#> # ℹ 23 more variables: DOSE_STR <chr>, Cohort <chr>, ROUTE <chr>, WT <dbl>,
#> #   SUBTYPE <chr>, SEX <chr>, F1 <dbl>, ka <dbl>, CL <dbl>, Vc <dbl>, Vp <dbl>,
#> #   Q <dbl>, BM_IC <dbl>, kdeg_BM <dbl>, Emax <dbl>, EC50 <dbl>, CMT <chr>,
#> #   DUR <dbl>, EVID <dbl>, NDose <chr>, DOSE_NUM <dbl>, SUBTYPE_ID <dbl>,
#> #   SEX_ID <dbl>
#> 
#> $DW$views$view_1$elements_table
#> NULL
#> 
#> $DW$views$view_1$elements_list
#> list()
#> 
#> $DW$views$view_1$dwe_cntr
#> [1] 1
#> 
#> $DW$views$view_1$checksum
#> [1] "b18a369b1e4d93010a04a56c7cde00c8"
#> 
#> $DW$views$view_1$code
#> NULL
#> 
#> $DW$views$view_1$code_dw_only
#> NULL
#> 
#> 
#> 
#> $DW$current_view
#> [1] "view_1"
#> 
#> $DW$checksum
#> [1] "5b0f0b05ee3ac7336a74c564bb6efdad"
#> 
#> 
#> $MOD_TYPE
#> [1] "DW"
#> 
#> $id
#> [1] "DW"
#> 
#> $dep_mod_ids
#> [1] "UD"
#> 
#> $FM_yaml_file
#> [1] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpcUKOgR/temp_libpath5bb2519acd9d/formods/templates/formods.yaml"
#> 
#> $MOD_yaml_file
#> [1] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpcUKOgR/temp_libpath5bb2519acd9d/formods/templates/DW.yaml"
#> 
#> $shiny_token
#> [1] "non_shiny"
#>