Preload Data for DW Module

Populates the supplied session variable with information from list of sources.

Usage

DW_preload(
  session,
  src_list,
  yaml_res,
  mod_ID = NULL,
  react_state = list(),
  quickload = FALSE
)

Arguments

session

Shiny session variable (in app) or a list (outside of app)

src_list

List of preload data (all read together with module IDs at the top level)

mod_ID

Module ID of the module being loaded.

react_state

Reactive shiny object (in app) or a list (outside of app) used to trigger reactions.

quickload

Logical TRUE to load reduced analysis FALSE to load the full analysis

Value

list with the following elements

isgood: Boolean indicating the exit status of the function.

msgs: Messages to be passed back to the user.

session: Session object

input: The value of the shiny input at the end of the session initialization.

state: App state.

react_state: The react_state components.