Evaluates Figure Generation Code

Takes the current state and rebuilds the active figure. If the elements table has a row flagged for deletion, it will be deleted. If the cmd input is not NULL it will attempt to append that element to the figure.

Usage

FG_build(
  state,
  del_row = NULL,
  cmd = NULL,
  element = "unknown",
  desc = "unknown"
)

Arguments

state

FG state from FG_fetch_state()

del_row

Row number to be deleted (NULL if no rows need to be deleted)

cmd

String containing the plotting command. Set to NULL to initialize a new figure or force a rebuild after a dataset update.

element

Short name for the figure element being performed, eg. point

desc

Verbose description for the action being performed

Value

list with the following elements

isgood: Return status of the function.

msgs: Messages to be passed back to the user.

pages: List with each element containing a ggplot object (p) and the code to generate that object (code)

Examples

library(formods)
# Within shiny both session and input variables will exist, 
# this creates examples here for testing purposes:
sess_res = FG_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:b4bd907af4f3e3b571027d1230bf99f1
#> → DW: State initialized
#> → DW: module checksum updated:a9c58c8823c19140e4608cc9a0d95cc1
#> → DW: module checksum updated:ed1b4d28dcc64d59a538273bfd8cbfe7
#> → DW: module checksum updated:517754706893fd756b06d3b79bac8d5b
#> → DW: module checksum updated:7a46321aac57ea8514ea07caf4ccb4e7
#> → DW: module checksum updated:9929ec9ea2bc5d28342a19889261e65f
#> → DW: module checksum updated:c2a8ef415bd36e981fe4f88e5b499ba3
#> → DW: module checksum updated:0c32255ec1c5482a60938f787fdbcb17
#> → DW: module checksum updated:f096133b286d340e723bdda9ed0c9172
#> → DW: module checksum updated:d7df0ea03ec446323da2696b703c60b1
#> → DW: module checksum updated:87a8e10e49e67fb0b21a0bf96b398b6c
#> → DW: module checksum updated:209c3bcc8497eb25c2f3301a3d76006f
#> → DW: module checksum updated:5924e7a6a637d1bd43d099d6db3f4939
#> → DW: module checksum updated:f6e09337e2c4eedbbe705f0bc0b784d5
#> → DW: module checksum updated:cc00c0564573ad88f31beb55262f5443
#> → DW: module checksum updated:23b114d90922065a80aa8e665a158e6e
#> → DW: module checksum updated:db71e5e51679cd7673d2d951d6fda71e
#> → DW: module checksum updated:8111c3e76b1f5aabf6e2ebf0da2acf1e
#> → DW: module checksum updated:7dace004a38bcfba67f7d38deede9af8
#> → DW: module checksum updated:699a9d95559cf9e6d041b4901dff275c
#> → DW: module checksum updated:41d02c9ea6c806fecf527817498ec1c0
#> → DW: module checksum updated:35889d32f3526013f0f32b6268c94921
#> → DW: module checksum updated:69b9a3a363278ccde7439cf078d9eb6e
#> → DW: module checksum updated:a55b96d282553362ecf1cf8761be2261
#> → DW: module checksum updated:cae93dfa234c69be5467d9ec44bf54da
#> → DW: module checksum updated:9b88e64e8ec46ae2734b51432a29e6f1
#> → FG: State initialized
#> → FG: module checksum updated:91329389512102f6fab3f9986c1e5703
#> → FG: module checksum updated:0d23c8025661426bba48ab9e598cab08
#> → FG: module checksum updated:526e61b790dc54f85c2590fbda7a88e0
#> → FG: module checksum updated:7a4bf1a647ba47f817db673198dd3416
#> → FG: module checksum updated:7aed1b869dcd297ac9c2b612e3afe65e
#> → FG: module checksum updated:534317a2633e6bf3d791b741a514b338
#> → FG: module checksum updated:4f0f2f7282e300304a1fa139a796e7af
#> → FG: module checksum updated:08270b4145cc06b1fa62ec62a35aa3db
#> → FG: module checksum updated:26a46eed167abb5c71b0e38b49e60a8b
#> → FG: module checksum updated:6235ef131cffae9be88d1cbd030b6ed9
#> → FG: module checksum updated:e6ec7cc5193e451166f88222b9ea5353
#> → FG: module checksum updated:b382604475a80b1cb3d504d5b980facd
#> → FG: module checksum updated:d5b568929302ff8076e5d767c25dd9fb
session = sess_res$session
input   = sess_res$input

# This will create a populated FG state object:
state   = sess_res$state

# This sets the current active figure to Fig_1
state[["FG"]][["current_fig"]]  =  "Fig_1"

# This is a paginated figure, and we can access a specific
# figure using the following:
pg_1 = FG_extract_page(state, 1)
pg_2 = FG_extract_page(state, 2)

# This will give you access to the current figure directly:
current_fig = FG_fetch_current_fig(state)

# For example this will set the key for that figure:
current_fig$key = "Individual profiles by cohort (multiple pages)"

# Once you're done you can put it back into the state:
state = FG_set_current_fig(state, current_fig)

# If you made any changes to the actual figure, this will 
# force a rebuild of the current figure:
state = FG_build( state=state, del_row = NULL, cmd = NULL)
#> → FG: module checksum updated:77a5e9273edd8417f59d34314d243be4

# To create a new empty figure you can do this:
state = FG_new_fig(state)
#> → FG: module checksum updated:838205c14fe9d103c345cb969de12d41