Initialize New Figure

Creates a new figure in a FG module

Usage

FG_new_fig(state)

Arguments

state

FG state from FG_fetch_state()

Value

FG state object containing a new empty figure and that figure set as the current active figure

Examples

library(formods)
# Within shiny both session and input variables will exist, 
# this creates examples here for testing purposes:
sess_res = FG_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:48559d397570afc9fba8ad9b252fa75a
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85
#> → DW: module checksum updated:a5d1d3b8a436befcfa119b5d72d3697b
#> → DW: module checksum updated:48d9e35c3f568c7a6f72914538d4ecd4
#> → DW: module checksum updated:680bde3fe602684945b93dda69abc020
#> → DW: module checksum updated:1f571c6eabb372eff6c1ef806344b2e4
#> → DW: module checksum updated:a1632ef431da47f98334a9281270a3fd
#> → DW: module checksum updated:518e56af743d086ea4b1b6070bf70472
#> → DW: module checksum updated:460fd6e3a4a46f56f17bde8ff4ef5c01
#> → DW: module checksum updated:12335ca5ced95462c2ed1df919a72d15
#> → DW: module checksum updated:0794933dac491313b32ca4bbfd996641
#> → DW: module checksum updated:75abe1a1cfc787fc68049b40c9e5749e
#> → DW: module checksum updated:070868f0125fd16b3e648dc05be84cd0
#> → DW: module checksum updated:1da012ee43c59bf8bedc7376511d179e
#> → DW: module checksum updated:6dfd06414929a762cc4f07d6d428289b
#> → DW: module checksum updated:83f7106b33c98e7650432749c16d7a36
#> → DW: module checksum updated:32652abd1996af932cbdf304cc509bd0
#> → DW: module checksum updated:dfa171056875bc474e03e0c911ace2b6
#> → DW: module checksum updated:5d4dfb7affd116c593a856034804aa04
#> → DW: module checksum updated:f47016aa2630055e8a1ae44ab79cabd2
#> → DW: module checksum updated:46cc2c9503359a39be98c3c0b889851a
#> → DW: module checksum updated:58ed3ad9e797e77eba05a2576b78e50f
#> → DW: module checksum updated:aaf6ebcd2be24f05cd9b4776ef042986
#> → DW: module checksum updated:afbbf99a8ed5f01191217b8370603a9d
#> → DW: module checksum updated:086f760681a2b47a9a1dfd572f876fcd
#> → DW: module checksum updated:8c045cb5e0fe9bd74e84d87d3c6725b5
#> → DW: module checksum updated:779a7d3c81e58ea4c8a756c2e422fee0
#> → DW: module checksum updated:104438a337e6617c4504477506e90863
#> → DW: module checksum updated:1bd57823c76715ccceab0be02ed7a79e
#> → DW: module checksum updated:8b901fc11d897dd27c774accf0eb7097
#> → DW: module checksum updated:7d55b640df1207dcdc56a10c9b8d8b87
#> → DW: module checksum updated:8d91f192ea994a133c8e2fbb7ce7c11a
#> → DW: module checksum updated:d5c81022154ffb7347d7a30d66a38f60
#> → FG: State initialized
#> → FG: module checksum updated:a77ad23e318409eb0b536b8e2bd9f54b
#> → FG: module checksum updated:3475cb480f9580919aada696ba54e7cd
#> → FG: module checksum updated:1528962c088fc3f5ee3eb49442038da0
#> → FG: module checksum updated:b482ea5c3b56a99d0dd6e8317fc49461
#> → FG: module checksum updated:629384957a547ad8e2c6689b9e65cdcb
#> → FG: module checksum updated:6563018430fb1c957909aa0187c9bf27
#> → FG: module checksum updated:dccd2ffd46784d65fb6d93c4befc641f
#> → FG: module checksum updated:f9575335d95fa0be5243a0dca91b19df
#> → FG: module checksum updated:2dc6a2a9c78f6ae3470a43e84f609a60
#> → FG: module checksum updated:c3eb7cd2bbde19854f4afade90d697c6
#> → FG: module checksum updated:f57e5310b00967309c18a0315b0e149a
#> → FG: module checksum updated:89233d9b5aa642aafb5958d4d6ff9866
#> → FG: module checksum updated:a630f784cba02353ef380371cb818f95
session = sess_res$session
input   = sess_res$input

# This will create a populated FG state object:
state   = sess_res$state

# This sets the current active figure to Fig_1
state[["FG"]][["current_fig"]]  =  "Fig_1"

# This is a paginated figure, and we can access a specific
# figure using the following:
pg_1 = FG_extract_page(state, 1)
pg_2 = FG_extract_page(state, 2)

# This will give you access to the current figure directly:
current_fig = FG_fetch_current_fig(state)

# For example this will set the key for that figure:
current_fig$key = "Individual profiles by cohort (multiple pages)"

# Once you're done you can put it back into the state:
state = FG_set_current_fig(state, current_fig)

# If you made any changes to the actual figure, this will 
# force a rebuild of the current figure:
state = FG_build( state=state, del_row = NULL, cmd = NULL)
#> → FG: module checksum updated:4e8cc1a6b15c7b9b7a8af40c52b0e8c2

# To create a new empty figure you can do this:
state = FG_new_fig(state)
#> → FG: module checksum updated:eab38ee50b70d716692150558c0d7ec7