Fetches Informaiton About the App

Returns diagnostic information about the app

Usage

FM_fetch_app_info(session)

Arguments

session

Shiny session variable.

Value

List with information about the app with the following structure

uiele: All system information as UI elements to be used in shiny apps.

uiele_packages: UI element for installed packages to be used in shiny apps.

uiele_options: UI element for current options.

uiele_modules: UI element for loaded formods modules to be used in shiny apps.

modules: List with formods module IDs used in the app as the element names. Each contains details about that module.

msgs: System information as text to be used in a report/terminal.

si_packages Dataframe with currently used packages.

si_options Dataframe with current options

Examples

# We need a Shiny session object to use this function:
sess_res = UD_test_mksession()
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: setting word placeholders: 
#> → ASM:   -> setting docx ph: HEADERLEFT = left header
#> → ASM:   -> setting docx ph: HEADERRIGHT = right header
#> → ASM:   -> setting docx ph: FOOTERLEFT = left footer
#> → ASM: module isgood: TRUE
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
#> → UD: module isgood: TRUE
session = sess_res$session
app_info  = FM_fetch_app_info(session)
app_info$msgs
#>   [1] "Modules"                                                                                                                                  
#>   [2] "ID: ASM"                                                                                                                                  
#>   [3] "type: ASM"                                                                                                                                
#>   [4] "FM_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpLfv7kX/temp_libpath689b4b58ac21/formods/templates/formods.yaml"
#>   [5] "MOD_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpLfv7kX/temp_libpath689b4b58ac21/formods/templates/ASM.yaml"   
#>   [6] "User files: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//RtmpxrOFHp/non_shiny/FM"                                                    
#>   [7] "Log file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//RtmpxrOFHp/non_shiny/FM/formods_log.txt"                                      
#>   [8] "ID: UD"                                                                                                                                   
#>   [9] "type: UD"                                                                                                                                 
#>  [10] "FM_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpLfv7kX/temp_libpath689b4b58ac21/formods/templates/formods.yaml"
#>  [11] "MOD_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpLfv7kX/temp_libpath689b4b58ac21/formods/templates/UD.yaml"    
#>  [12] "User files: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//RtmpxrOFHp/non_shiny/FM"                                                    
#>  [13] "Log file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//RtmpxrOFHp/non_shiny/FM/formods_log.txt"                                      
#>  [14] "Package dependencies: janitor, readxl, rio"                                                                                               
#>  [15] " package           * version date (UTC) lib source"                                                                                       
#>  [16] " abind               1.4-8   2024-09-12 [2] CRAN (R 4.4.1)"                                                                               
#>  [17] " askpass             1.2.1   2024-10-04 [2] CRAN (R 4.4.1)"                                                                               
#>  [18] " backports           1.5.0   2024-05-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [19] " brio                1.1.5   2024-04-24 [2] CRAN (R 4.4.0)"                                                                               
#>  [20] " broom               1.0.7   2024-09-26 [2] CRAN (R 4.4.1)"                                                                               
#>  [21] " bslib               0.9.0   2025-01-30 [2] CRAN (R 4.4.1)"                                                                               
#>  [22] " cachem              1.1.0   2024-05-16 [2] CRAN (R 4.4.0)"                                                                               
#>  [23] " car                 3.1-3   2024-09-27 [2] CRAN (R 4.4.1)"                                                                               
#>  [24] " carData             3.0-5   2022-01-06 [2] CRAN (R 4.4.0)"                                                                               
#>  [25] " cellranger          1.1.0   2016-07-27 [2] CRAN (R 4.4.0)"                                                                               
#>  [26] " cli                 3.6.3   2024-06-21 [2] CRAN (R 4.4.0)"                                                                               
#>  [27] " clipr               0.8.0   2022-02-22 [2] CRAN (R 4.4.0)"                                                                               
#>  [28] " colorspace          2.1-1   2024-07-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [29] " covr                3.6.4   2023-11-09 [2] CRAN (R 4.4.0)"                                                                               
#>  [30] " crosstalk           1.2.1   2023-11-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [31] " curl                6.2.0   2025-01-23 [2] CRAN (R 4.4.1)"                                                                               
#>  [32] " data.table          1.16.4  2024-12-06 [2] CRAN (R 4.4.1)"                                                                               
#>  [33] " desc                1.4.3   2023-12-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [34] " devtools            2.4.5   2022-10-11 [2] CRAN (R 4.4.0)"                                                                               
#>  [35] " digest              0.6.37  2024-08-19 [2] CRAN (R 4.4.1)"                                                                               
#>  [36] " downlit             0.4.4   2024-06-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [37] " dplyr               1.1.4   2023-11-17 [2] CRAN (R 4.4.0)"                                                                               
#>  [38] " DT                  0.33    2024-04-04 [2] CRAN (R 4.4.0)"                                                                               
#>  [39] " ellipsis            0.3.2   2021-04-29 [2] CRAN (R 4.4.0)"                                                                               
#>  [40] " evaluate            1.0.3   2025-01-10 [2] CRAN (R 4.4.1)"                                                                               
#>  [41] " fansi               1.0.6   2023-12-08 [2] CRAN (R 4.4.0)"                                                                               
#>  [42] " farver              2.1.2   2024-05-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [43] " fastmap             1.2.0   2024-05-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [44] " flextable           0.9.7   2024-10-27 [2] CRAN (R 4.4.1)"                                                                               
#>  [45] " fontawesome         0.5.3   2024-11-16 [2] CRAN (R 4.4.1)"                                                                               
#>  [46] " fontBitstreamVera   0.1.1   2017-02-01 [2] CRAN (R 4.4.0)"                                                                               
#>  [47] " fontLiberation      0.1.0   2016-10-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [48] " fontquiver          0.2.1   2017-02-01 [2] CRAN (R 4.4.0)"                                                                               
#>  [49] " formods           * 0.2.1   2025-03-04 [1] local"                                                                                        
#>  [50] " Formula             1.2-5   2023-02-24 [2] CRAN (R 4.4.0)"                                                                               
#>  [51] " fs                  1.6.5   2024-10-30 [2] CRAN (R 4.4.1)"                                                                               
#>  [52] " gdtools             0.4.1   2024-11-04 [2] CRAN (R 4.4.1)"                                                                               
#>  [53] " generics            0.1.3   2022-07-05 [2] CRAN (R 4.4.0)"                                                                               
#>  [54] " ggforce             0.4.2   2024-02-19 [2] CRAN (R 4.4.0)"                                                                               
#>  [55] " ggplot2             3.5.1   2024-04-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [56] " ggpubr              0.6.0   2023-02-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [57] " ggsignif            0.6.4   2022-10-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [58] " glue                1.8.0   2024-09-30 [2] CRAN (R 4.4.1)"                                                                               
#>  [59] " gtable              0.3.6   2024-10-25 [2] CRAN (R 4.4.1)"                                                                               
#>  [60] " gtools              3.9.5   2023-11-20 [2] CRAN (R 4.4.0)"                                                                               
#>  [61] " here                1.0.1   2020-12-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [62] " htmltools           0.5.8.1 2024-04-04 [2] CRAN (R 4.4.0)"                                                                               
#>  [63] " htmlwidgets         1.6.4   2023-12-06 [2] CRAN (R 4.4.0)"                                                                               
#>  [64] " httpuv              1.6.15  2024-03-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [65] " httr                1.4.7   2023-08-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [66] " httr2               1.1.0   2025-01-18 [2] CRAN (R 4.4.1)"                                                                               
#>  [67] " janitor             2.2.1   2024-12-22 [2] CRAN (R 4.4.1)"                                                                               
#>  [68] " jquerylib           0.1.4   2021-04-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [69] " jsonlite            1.8.9   2024-09-20 [2] CRAN (R 4.4.1)"                                                                               
#>  [70] " knitr               1.49    2024-11-08 [2] CRAN (R 4.4.1)"                                                                               
#>  [71] " labeling            0.4.3   2023-08-29 [2] CRAN (R 4.4.0)"                                                                               
#>  [72] " later               1.4.1   2024-11-27 [2] CRAN (R 4.4.1)"                                                                               
#>  [73] " lazyeval            0.2.2   2019-03-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [74] " lifecycle           1.0.4   2023-11-07 [2] CRAN (R 4.4.0)"                                                                               
#>  [75] " lubridate           1.9.4   2024-12-08 [2] CRAN (R 4.4.1)"                                                                               
#>  [76] " magrittr            2.0.3   2022-03-30 [2] CRAN (R 4.4.0)"                                                                               
#>  [77] " MASS                7.3-64  2025-01-04 [2] CRAN (R 4.4.1)"                                                                               
#>  [78] " memoise             2.0.1   2021-11-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [79] " mime                0.12    2021-09-28 [2] CRAN (R 4.4.0)"                                                                               
#>  [80] " miniUI              0.1.1.1 2018-05-18 [2] CRAN (R 4.4.0)"                                                                               
#>  [81] " munsell             0.5.1   2024-04-01 [2] CRAN (R 4.4.0)"                                                                               
#>  [82] " officer             0.6.7   2024-10-09 [2] CRAN (R 4.4.1)"                                                                               
#>  [83] " onbrand             1.0.7   2025-02-07 [2] Github (john-harrold/onbrand@884bbd8)"                                                        
#>  [84] " openssl             2.3.2   2025-02-03 [2] CRAN (R 4.4.1)"                                                                               
#>  [85] " pillar              1.10.1  2025-01-07 [2] CRAN (R 4.4.1)"                                                                               
#>  [86] " pkgbuild            1.4.6   2025-01-16 [2] CRAN (R 4.4.1)"                                                                               
#>  [87] " pkgconfig           2.0.3   2019-09-22 [2] CRAN (R 4.4.0)"                                                                               
#>  [88] " pkgdown             2.1.1   2024-09-17 [2] CRAN (R 4.4.1)"                                                                               
#>  [89] " pkgload             1.4.0   2024-06-28 [2] CRAN (R 4.4.0)"                                                                               
#>  [90] " plotly              4.10.4  2024-01-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [91] " polyclip            1.10-7  2024-07-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [92] " profvis             0.4.0   2024-09-20 [2] CRAN (R 4.4.1)"                                                                               
#>  [93] " promises            1.3.2   2024-11-28 [2] CRAN (R 4.4.1)"                                                                               
#>  [94] " prompter            1.2.0   2024-04-08 [2] CRAN (R 4.4.0)"                                                                               
#>  [95] " purrr               1.0.4   2025-02-05 [2] CRAN (R 4.4.1)"                                                                               
#>  [96] " R.methodsS3         1.8.2   2022-06-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [97] " R.oo                1.27.0  2024-11-01 [2] CRAN (R 4.4.1)"                                                                               
#>  [98] " R.utils             2.12.3  2023-11-18 [2] CRAN (R 4.4.0)"                                                                               
#>  [99] " R6                  2.5.1   2021-08-19 [2] CRAN (R 4.4.1)"                                                                               
#> [100] " ragg                1.3.3   2024-09-11 [2] CRAN (R 4.4.1)"                                                                               
#> [101] " rappdirs            0.3.3   2021-01-31 [2] CRAN (R 4.4.0)"                                                                               
#> [102] " Rcpp                1.0.14  2025-01-12 [2] CRAN (R 4.4.1)"                                                                               
#> [103] " readxl              1.4.3   2023-07-06 [2] CRAN (R 4.4.0)"                                                                               
#> [104] " remotes             2.5.0   2024-03-17 [2] CRAN (R 4.4.1)"                                                                               
#> [105] " rex                 1.2.1   2021-11-26 [2] CRAN (R 4.4.0)"                                                                               
#> [106] " rhandsontable       0.3.8   2021-05-27 [2] CRAN (R 4.4.0)"                                                                               
#> [107] " rio                 1.2.3   2024-09-25 [2] CRAN (R 4.4.1)"                                                                               
#> [108] " rlang               1.1.5   2025-01-17 [2] CRAN (R 4.4.1)"                                                                               
#> [109] " rmarkdown           2.29    2024-11-04 [2] CRAN (R 4.4.1)"                                                                               
#> [110] " rprojroot           2.0.4   2023-11-05 [2] CRAN (R 4.4.0)"                                                                               
#> [111] " rstatix             0.7.2   2023-02-01 [2] CRAN (R 4.4.0)"                                                                               
#> [112] " rstudioapi          0.17.1  2024-10-22 [2] CRAN (R 4.4.1)"                                                                               
#> [113] " sass                0.4.9   2024-03-15 [2] CRAN (R 4.4.0)"                                                                               
#> [114] " scales              1.3.0   2023-11-28 [2] CRAN (R 4.4.0)"                                                                               
#> [115] " sessioninfo         1.2.3   2025-02-05 [2] CRAN (R 4.4.1)"                                                                               
#> [116] " shiny               1.10.0  2024-12-14 [2] CRAN (R 4.4.1)"                                                                               
#> [117] " shinyAce            0.4.4   2025-02-03 [2] CRAN (R 4.4.1)"                                                                               
#> [118] " shinybusy           0.3.3   2024-03-09 [2] CRAN (R 4.4.0)"                                                                               
#> [119] " shinydashboard      0.7.2   2021-09-30 [2] CRAN (R 4.4.0)"                                                                               
#> [120] " shinyWidgets        0.8.7   2024-09-23 [2] CRAN (R 4.4.1)"                                                                               
#> [121] " snakecase           0.11.1  2023-08-27 [2] CRAN (R 4.4.0)"                                                                               
#> [122] " stringi             1.8.4   2024-05-06 [2] CRAN (R 4.4.0)"                                                                               
#> [123] " stringr             1.5.1   2023-11-14 [2] CRAN (R 4.4.0)"                                                                               
#> [124] " systemfonts         1.2.1   2025-01-20 [2] CRAN (R 4.4.1)"                                                                               
#> [125] " testthat            3.2.3   2025-01-13 [2] CRAN (R 4.4.1)"                                                                               
#> [126] " textshaping         1.0.0   2025-01-20 [2] CRAN (R 4.4.1)"                                                                               
#> [127] " tibble              3.2.1   2023-03-20 [2] CRAN (R 4.4.0)"                                                                               
#> [128] " tidyr               1.3.1   2024-01-24 [2] CRAN (R 4.4.0)"                                                                               
#> [129] " tidyselect          1.2.1   2024-03-11 [2] CRAN (R 4.4.0)"                                                                               
#> [130] " timechange          0.3.0   2024-01-18 [2] CRAN (R 4.4.0)"                                                                               
#> [131] " tweenr              2.0.3   2024-02-26 [2] CRAN (R 4.4.0)"                                                                               
#> [132] " urlchecker          1.0.1   2021-11-30 [2] CRAN (R 4.4.0)"                                                                               
#> [133] " usethis             3.1.0   2024-11-26 [2] CRAN (R 4.4.1)"                                                                               
#> [134] " uuid                1.2-1   2024-07-29 [2] CRAN (R 4.4.0)"                                                                               
#> [135] " vctrs               0.6.5   2023-12-01 [2] CRAN (R 4.4.0)"                                                                               
#> [136] " viridisLite         0.4.2   2023-05-02 [2] CRAN (R 4.4.0)"                                                                               
#> [137] " whisker             0.4.1   2022-12-05 [2] CRAN (R 4.4.0)"                                                                               
#> [138] " withr               3.0.2   2024-10-28 [2] CRAN (R 4.4.1)"                                                                               
#> [139] " xfun                0.50    2025-01-07 [2] CRAN (R 4.4.1)"                                                                               
#> [140] " xml2                1.3.6   2023-12-04 [2] CRAN (R 4.4.0)"                                                                               
#> [141] " xtable              1.8-4   2019-04-21 [2] CRAN (R 4.4.0)"                                                                               
#> [142] " yaml                2.3.10  2024-07-26 [2] CRAN (R 4.4.0)"                                                                               
#> [143] " zip                 2.3.2   2025-02-01 [2] CRAN (R 4.4.1)"                                                                               
#> [144] ""                                                                                                                                         
#> [145] " [1] /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpLfv7kX/temp_libpath689b4b58ac21"                                        
#> [146] " [2] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library"                                                                
#> [147] " * ── Packages attached to the search path."