Fetches the App State

Returns the entire state of the App

Usage

FM_fetch_app_state(session)

Arguments

session

Shiny session variable.

Value

App state or NULL if it's not defined.

Examples

# We need a Shiny session object to use this function:
sess_res = UD_test_mksession()
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: setting word placeholders: 
#> → ASM:   -> setting docx ph: HEADERLEFT = left header
#> → ASM:   -> setting docx ph: HEADERRIGHT = right header
#> → ASM:   -> setting docx ph: FOOTERLEFT = left footer
#> → ASM: module isgood: TRUE
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
#> → UD: module isgood: TRUE
session = sess_res$session
app_state = FM_fetch_app_state(session)
app_state
#> $FM_ASM
#> $FM_ASM$yaml
#> $FM_ASM$yaml$FM
#> $FM_ASM$yaml$FM$include
#> $FM_ASM$yaml$FM$include$files
#> $FM_ASM$yaml$FM$include$files[[1]]
#> $FM_ASM$yaml$FM$include$files[[1]]$file
#> $FM_ASM$yaml$FM$include$files[[1]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\")"
#> 
#> $FM_ASM$yaml$FM$include$files[[1]]$file$dest
#> [1] "file.path(\"config\",\"report.docx\")"
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$include$files[[2]]
#> $FM_ASM$yaml$FM$include$files[[2]]$file
#> $FM_ASM$yaml$FM$include$files[[2]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\")"
#> 
#> $FM_ASM$yaml$FM$include$files[[2]]$file$dest
#> [1] "file.path(\"config\",\"report.pptx\")"
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$include$files[[3]]
#> $FM_ASM$yaml$FM$include$files[[3]]$file
#> $FM_ASM$yaml$FM$include$files[[3]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\")"
#> 
#> $FM_ASM$yaml$FM$include$files[[3]]$file$dest
#> [1] "file.path(\"config\",\"report.yaml\")"
#> 
#> 
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$deployed
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$code
#> $FM_ASM$yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $FM_ASM$yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $FM_ASM$yaml$FM$code$gen_file
#> [1] "run_analysis.R"
#> 
#> $FM_ASM$yaml$FM$code$gen_preamble
#> [1] "# formods automated output ------------------------------------------------\n# https://formods.ubiquity.tools/\nrm(list=ls())"
#> 
#> $FM_ASM$yaml$FM$code$packages
#> [1] "onbrand" "writexl"
#> 
#> 
#> $FM_ASM$yaml$FM$notifications
#> $FM_ASM$yaml$FM$notifications$config
#> $FM_ASM$yaml$FM$notifications$config$success
#> $FM_ASM$yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $FM_ASM$yaml$FM$notifications$config$failure
#> $FM_ASM$yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $FM_ASM$yaml$FM$notifications$config$info
#> $FM_ASM$yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $FM_ASM$yaml$FM$notifications$config$warning
#> $FM_ASM$yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $FM_ASM$yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$reporting
#> $FM_ASM$yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $FM_ASM$yaml$FM$reporting$content_init
#> $FM_ASM$yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $FM_ASM$yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.docx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> $FM_ASM$yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.pptx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> 
#> $FM_ASM$yaml$FM$reporting$phs
#> $FM_ASM$yaml$FM$reporting$phs[[1]]
#> $FM_ASM$yaml$FM$reporting$phs[[1]]$name
#> [1] "HEADERLEFT"
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[1]]$location
#> [1] "header"
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[1]]$value
#> [1] ""
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[1]]$tooltip
#> [1] "left header text"
#> 
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[2]]
#> $FM_ASM$yaml$FM$reporting$phs[[2]]$name
#> [1] "HEADERRIGHT"
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[2]]$location
#> [1] "header"
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[2]]$value
#> [1] ""
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[2]]$tooltip
#> [1] "right header text"
#> 
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[3]]
#> $FM_ASM$yaml$FM$reporting$phs[[3]]$name
#> [1] "FOOTERLEFT"
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[3]]$location
#> [1] "footer"
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[3]]$value
#> [1] ""
#> 
#> $FM_ASM$yaml$FM$reporting$phs[[3]]$tooltip
#> [1] "left footer text"
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$reporting$phs_formatting
#> $FM_ASM$yaml$FM$reporting$phs_formatting$width
#> [1] "100%"
#> 
#> $FM_ASM$yaml$FM$reporting$phs_formatting$tt_position
#> [1] "left"
#> 
#> $FM_ASM$yaml$FM$reporting$phs_formatting$tt_size
#> [1] "medium"
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$ui
#> $FM_ASM$yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $FM_ASM$yaml$FM$ui$select_size
#> [1] 10
#> 
#> $FM_ASM$yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $FM_ASM$yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $FM_ASM$yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $FM_ASM$yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $FM_ASM$yaml$FM$data_meta
#> $FM_ASM$yaml$FM$data_meta$data_header
#> [1] "<span style='color:===COLOR==='><b>===NAME===</b><br/><font size='-3'>===LABEL===</font></span>"
#> 
#> $FM_ASM$yaml$FM$data_meta$subtext
#> [1] "===LABEL===: ===RANGE==="
#> 
#> $FM_ASM$yaml$FM$data_meta$many_sep
#> [1] ",...,"
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types
#> $FM_ASM$yaml$FM$data_meta$data_types$character
#> $FM_ASM$yaml$FM$data_meta$data_types$character$color
#> [1] "#DD4B39"
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types$character$label
#> [1] "text"
#> 
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types$double
#> $FM_ASM$yaml$FM$data_meta$data_types$double$color
#> [1] "#3C8DBC"
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types$double$label
#> [1] "num"
#> 
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types$integer
#> $FM_ASM$yaml$FM$data_meta$data_types$integer$color
#> [1] "#3C8DBC"
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types$integer$label
#> [1] "num"
#> 
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types$other
#> $FM_ASM$yaml$FM$data_meta$data_types$other$color
#> [1] "black"
#> 
#> $FM_ASM$yaml$FM$data_meta$data_types$other$label
#> [1] "other"
#> 
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$workflows
#> $FM_ASM$yaml$FM$workflows$example
#> $FM_ASM$yaml$FM$workflows$example$group
#> [1] "Examples"
#> 
#> $FM_ASM$yaml$FM$workflows$example$desc
#> [1] "Example Workflow"
#> 
#> $FM_ASM$yaml$FM$workflows$example$require_ds
#> [1] TRUE
#> 
#> $FM_ASM$yaml$FM$workflows$example$preload
#> [1] "file.path('.', 'example.yaml')"
#> 
#> 
#> 
#> $FM_ASM$yaml$FM$labels
#> $FM_ASM$yaml$FM$labels$ui_label
#> [1] "put labels here"
#> 
#> 
#> $FM_ASM$yaml$FM$user_files
#> $FM_ASM$yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $FM_ASM$yaml$FM$logging
#> $FM_ASM$yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $FM_ASM$yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $FM_ASM$yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $FM_ASM$yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $FM_ASM$yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $FM_ASM$MC
#> $FM_ASM$MC$module
#> $FM_ASM$MC$module$type
#> [1] "ASM"
#> 
#> $FM_ASM$MC$module$depends
#> NULL
#> 
#> 
#> $FM_ASM$MC$code
#> $FM_ASM$MC$code$readOnly
#> [1] TRUE
#> 
#> $FM_ASM$MC$code$mode
#> [1] "r"
#> 
#> 
#> $FM_ASM$MC$show_log
#> [1] 50
#> 
#> $FM_ASM$MC$reporting
#> $FM_ASM$MC$reporting$enabled
#> [1] FALSE
#> 
#> $FM_ASM$MC$reporting$priority
#> [1] 1
#> 
#> 
#> $FM_ASM$MC$formatting
#> $FM_ASM$MC$formatting$input_load_state
#> $FM_ASM$MC$formatting$input_load_state$width
#> [1] "100%"
#> 
#> 
#> $FM_ASM$MC$formatting$ui_asm_save_name
#> $FM_ASM$MC$formatting$ui_asm_save_name$width
#> [1] "100%"
#> 
#> 
#> $FM_ASM$MC$formatting$button_state_save
#> $FM_ASM$MC$formatting$button_state_save$size
#> [1] "sm"
#> 
#> $FM_ASM$MC$formatting$button_state_save$block
#> [1] TRUE
#> 
#> 
#> $FM_ASM$MC$formatting$button_rpt_xlsx
#> $FM_ASM$MC$formatting$button_rpt_xlsx$size
#> [1] "sm"
#> 
#> $FM_ASM$MC$formatting$button_rpt_xlsx$block
#> [1] TRUE
#> 
#> 
#> $FM_ASM$MC$formatting$button_rpt_pptx
#> $FM_ASM$MC$formatting$button_rpt_pptx$size
#> [1] "sm"
#> 
#> $FM_ASM$MC$formatting$button_rpt_pptx$block
#> [1] TRUE
#> 
#> 
#> $FM_ASM$MC$formatting$button_rpt_docx
#> $FM_ASM$MC$formatting$button_rpt_docx$size
#> [1] "sm"
#> 
#> $FM_ASM$MC$formatting$button_rpt_docx$block
#> [1] TRUE
#> 
#> 
#> $FM_ASM$MC$formatting$switch_gen_reports
#> $FM_ASM$MC$formatting$switch_gen_reports$default
#> [1] FALSE
#> 
#> $FM_ASM$MC$formatting$switch_gen_reports$right
#> [1] TRUE
#> 
#> $FM_ASM$MC$formatting$switch_gen_reports$inline
#> [1] TRUE
#> 
#> $FM_ASM$MC$formatting$switch_gen_reports$width
#> NULL
#> 
#> 
#> 
#> $FM_ASM$MC$labels
#> $FM_ASM$MC$labels$ui_asm_save_name
#> [1] "Save Analysis"
#> 
#> $FM_ASM$MC$labels$ui_asm_save_name_ph
#> [1] "my_analysis"
#> 
#> $FM_ASM$MC$labels$ui_asm_switch_gen_rpts
#> [1] "Generate reports when saving."
#> 
#> $FM_ASM$MC$labels$save_state
#> [1] "Save"
#> 
#> $FM_ASM$MC$labels$gen_rpt_header
#> [1] "Generate Reports"
#> 
#> $FM_ASM$MC$labels$gen_rpt_docx_ph
#> [1] "Word Report Paceholders"
#> 
#> $FM_ASM$MC$labels$input_load_state
#> [1] "Load Analysis"
#> 
#> $FM_ASM$MC$labels$upload_button
#> [1] "Load"
#> 
#> $FM_ASM$MC$labels$upload_placeholder
#> [1] "Select analysis file"
#> 
#> $FM_ASM$MC$labels$ui_asm_rpt_xlsx
#> [1] "xlsx"
#> 
#> $FM_ASM$MC$labels$ui_asm_rpt_pptx
#> [1] "pptx"
#> 
#> $FM_ASM$MC$labels$ui_asm_rpt_docx
#> [1] "docx"
#> 
#> $FM_ASM$MC$labels$load_success
#> [1] "Analysis loaded"
#> 
#> $FM_ASM$MC$labels$busy
#> $FM_ASM$MC$labels$busy$loading_state
#> [1] "Loading analysis"
#> 
#> $FM_ASM$MC$labels$busy$saving_state
#> [1] "Saving analysis"
#> 
#> $FM_ASM$MC$labels$busy$xlsx
#> [1] "Generating report: Excel"
#> 
#> $FM_ASM$MC$labels$busy$docx
#> [1] "Generating report: Word"
#> 
#> $FM_ASM$MC$labels$busy$pptx
#> [1] "Generating report: PowerPoint"
#> 
#> 
#> 
#> $FM_ASM$MC$errors
#> $FM_ASM$MC$errors$bad_input
#> [1] "Input error message"
#> 
#> $FM_ASM$MC$errors$load_failed
#> [1] "Failed to load analysis"
#> 
#> 
#> $FM_ASM$MC$tooltips
#> $FM_ASM$MC$tooltips$include
#> [1] TRUE
#> 
#> $FM_ASM$MC$tooltips$components
#> $FM_ASM$MC$tooltips$components$input_tip
#> [1] "My tool tip"
#> 
#> 
#> 
#> 
#> $FM_ASM$ASM
#> $FM_ASM$ASM$button_counters
#> $FM_ASM$ASM$button_counters$button_state_save
#> [1] 0
#> 
#> $FM_ASM$ASM$button_counters$button_rpt_xlsx
#> [1] 0
#> 
#> $FM_ASM$ASM$button_counters$button_rpt_docx
#> [1] 0
#> 
#> $FM_ASM$ASM$button_counters$button_rpt_pptx
#> [1] 0
#> 
#> 
#> $FM_ASM$ASM$ui_hold
#> list()
#> 
#> $FM_ASM$ASM$ui_ids
#> [1] "button_state_save"          "button_rpt_xlsx"           
#> [3] "button_rpt_docx"            "button_rpt_pptx"           
#> [5] "ui_asm_save_name"           "switch_gen_rpts"           
#> [7] "ui_asm_docx_ph_HEADERLEFT"  "ui_asm_docx_ph_HEADERRIGHT"
#> [9] "ui_asm_docx_ph_FOOTERLEFT" 
#> 
#> $FM_ASM$ASM$isgood
#> [1] TRUE
#> 
#> $FM_ASM$ASM$checksum
#> [1] "f9e884084b84794d762a535f3facec85"
#> 
#> $FM_ASM$ASM$ph_uis
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERLEFT
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERLEFT$name
#> [1] "HEADERLEFT"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERLEFT$location
#> [1] "header"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERLEFT$value
#> [1] "left header"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERLEFT$tooltip
#> [1] "left header text"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERLEFT$ph_idx
#> [1] 1
#> 
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERRIGHT
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERRIGHT$name
#> [1] "HEADERRIGHT"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERRIGHT$location
#> [1] "header"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERRIGHT$value
#> [1] "right header"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERRIGHT$tooltip
#> [1] "right header text"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_HEADERRIGHT$ph_idx
#> [1] 2
#> 
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_FOOTERLEFT
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_FOOTERLEFT$name
#> [1] "FOOTERLEFT"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_FOOTERLEFT$location
#> [1] "footer"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_FOOTERLEFT$value
#> [1] "left footer"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_FOOTERLEFT$tooltip
#> [1] "left footer text"
#> 
#> $FM_ASM$ASM$ph_uis$ui_asm_docx_ph_FOOTERLEFT$ph_idx
#> [1] 3
#> 
#> 
#> 
#> $FM_ASM$ASM$ui
#> $FM_ASM$ASM$ui$ui_asm_docx_ph_HEADERLEFT
#> [1] "left header"
#> 
#> $FM_ASM$ASM$ui$ui_asm_docx_ph_HEADERRIGHT
#> [1] "right header"
#> 
#> $FM_ASM$ASM$ui$ui_asm_docx_ph_FOOTERLEFT
#> [1] "left footer"
#> 
#> $FM_ASM$ASM$ui$button_state_save
#> [1] ""
#> 
#> $FM_ASM$ASM$ui$button_rpt_xlsx
#> [1] ""
#> 
#> $FM_ASM$ASM$ui$button_rpt_docx
#> [1] ""
#> 
#> $FM_ASM$ASM$ui$button_rpt_pptx
#> [1] ""
#> 
#> $FM_ASM$ASM$ui$ui_asm_save_name
#> [1] ""
#> 
#> $FM_ASM$ASM$ui$switch_gen_rpts
#> [1] ""
#> 
#> 
#> 
#> $FM_ASM$MOD_TYPE
#> [1] "ASM"
#> 
#> $FM_ASM$id
#> [1] "ASM"
#> 
#> $FM_ASM$FM_yaml_file
#> [1] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpSYRoUO/temp_libpath13c2570e2c951/formods/templates/formods.yaml"
#> 
#> $FM_ASM$MOD_yaml_file
#> [1] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpSYRoUO/temp_libpath13c2570e2c951/formods/templates/ASM.yaml"
#> 
#> $FM_ASM$shiny_token
#> [1] "non_shiny"
#> 
#> 
#> $FM_UD
#> $FM_UD$yaml
#> $FM_UD$yaml$FM
#> $FM_UD$yaml$FM$include
#> $FM_UD$yaml$FM$include$files
#> $FM_UD$yaml$FM$include$files[[1]]
#> $FM_UD$yaml$FM$include$files[[1]]$file
#> $FM_UD$yaml$FM$include$files[[1]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\")"
#> 
#> $FM_UD$yaml$FM$include$files[[1]]$file$dest
#> [1] "file.path(\"config\",\"report.docx\")"
#> 
#> 
#> 
#> $FM_UD$yaml$FM$include$files[[2]]
#> $FM_UD$yaml$FM$include$files[[2]]$file
#> $FM_UD$yaml$FM$include$files[[2]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\")"
#> 
#> $FM_UD$yaml$FM$include$files[[2]]$file$dest
#> [1] "file.path(\"config\",\"report.pptx\")"
#> 
#> 
#> 
#> $FM_UD$yaml$FM$include$files[[3]]
#> $FM_UD$yaml$FM$include$files[[3]]$file
#> $FM_UD$yaml$FM$include$files[[3]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\")"
#> 
#> $FM_UD$yaml$FM$include$files[[3]]$file$dest
#> [1] "file.path(\"config\",\"report.yaml\")"
#> 
#> 
#> 
#> 
#> 
#> $FM_UD$yaml$FM$deployed
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$code
#> $FM_UD$yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $FM_UD$yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $FM_UD$yaml$FM$code$gen_file
#> [1] "run_analysis.R"
#> 
#> $FM_UD$yaml$FM$code$gen_preamble
#> [1] "# formods automated output ------------------------------------------------\n# https://formods.ubiquity.tools/\nrm(list=ls())"
#> 
#> $FM_UD$yaml$FM$code$packages
#> [1] "onbrand" "writexl"
#> 
#> 
#> $FM_UD$yaml$FM$notifications
#> $FM_UD$yaml$FM$notifications$config
#> $FM_UD$yaml$FM$notifications$config$success
#> $FM_UD$yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $FM_UD$yaml$FM$notifications$config$failure
#> $FM_UD$yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $FM_UD$yaml$FM$notifications$config$info
#> $FM_UD$yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $FM_UD$yaml$FM$notifications$config$warning
#> $FM_UD$yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $FM_UD$yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $FM_UD$yaml$FM$reporting
#> $FM_UD$yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $FM_UD$yaml$FM$reporting$content_init
#> $FM_UD$yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $FM_UD$yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.docx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> $FM_UD$yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.pptx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> 
#> $FM_UD$yaml$FM$reporting$phs
#> $FM_UD$yaml$FM$reporting$phs[[1]]
#> $FM_UD$yaml$FM$reporting$phs[[1]]$name
#> [1] "HEADERLEFT"
#> 
#> $FM_UD$yaml$FM$reporting$phs[[1]]$location
#> [1] "header"
#> 
#> $FM_UD$yaml$FM$reporting$phs[[1]]$value
#> [1] ""
#> 
#> $FM_UD$yaml$FM$reporting$phs[[1]]$tooltip
#> [1] "left header text"
#> 
#> 
#> $FM_UD$yaml$FM$reporting$phs[[2]]
#> $FM_UD$yaml$FM$reporting$phs[[2]]$name
#> [1] "HEADERRIGHT"
#> 
#> $FM_UD$yaml$FM$reporting$phs[[2]]$location
#> [1] "header"
#> 
#> $FM_UD$yaml$FM$reporting$phs[[2]]$value
#> [1] ""
#> 
#> $FM_UD$yaml$FM$reporting$phs[[2]]$tooltip
#> [1] "right header text"
#> 
#> 
#> $FM_UD$yaml$FM$reporting$phs[[3]]
#> $FM_UD$yaml$FM$reporting$phs[[3]]$name
#> [1] "FOOTERLEFT"
#> 
#> $FM_UD$yaml$FM$reporting$phs[[3]]$location
#> [1] "footer"
#> 
#> $FM_UD$yaml$FM$reporting$phs[[3]]$value
#> [1] ""
#> 
#> $FM_UD$yaml$FM$reporting$phs[[3]]$tooltip
#> [1] "left footer text"
#> 
#> 
#> 
#> $FM_UD$yaml$FM$reporting$phs_formatting
#> $FM_UD$yaml$FM$reporting$phs_formatting$width
#> [1] "100%"
#> 
#> $FM_UD$yaml$FM$reporting$phs_formatting$tt_position
#> [1] "left"
#> 
#> $FM_UD$yaml$FM$reporting$phs_formatting$tt_size
#> [1] "medium"
#> 
#> 
#> 
#> $FM_UD$yaml$FM$ui
#> $FM_UD$yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $FM_UD$yaml$FM$ui$select_size
#> [1] 10
#> 
#> $FM_UD$yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $FM_UD$yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $FM_UD$yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $FM_UD$yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $FM_UD$yaml$FM$data_meta
#> $FM_UD$yaml$FM$data_meta$data_header
#> [1] "<span style='color:===COLOR==='><b>===NAME===</b><br/><font size='-3'>===LABEL===</font></span>"
#> 
#> $FM_UD$yaml$FM$data_meta$subtext
#> [1] "===LABEL===: ===RANGE==="
#> 
#> $FM_UD$yaml$FM$data_meta$many_sep
#> [1] ",...,"
#> 
#> $FM_UD$yaml$FM$data_meta$data_types
#> $FM_UD$yaml$FM$data_meta$data_types$character
#> $FM_UD$yaml$FM$data_meta$data_types$character$color
#> [1] "#DD4B39"
#> 
#> $FM_UD$yaml$FM$data_meta$data_types$character$label
#> [1] "text"
#> 
#> 
#> $FM_UD$yaml$FM$data_meta$data_types$double
#> $FM_UD$yaml$FM$data_meta$data_types$double$color
#> [1] "#3C8DBC"
#> 
#> $FM_UD$yaml$FM$data_meta$data_types$double$label
#> [1] "num"
#> 
#> 
#> $FM_UD$yaml$FM$data_meta$data_types$integer
#> $FM_UD$yaml$FM$data_meta$data_types$integer$color
#> [1] "#3C8DBC"
#> 
#> $FM_UD$yaml$FM$data_meta$data_types$integer$label
#> [1] "num"
#> 
#> 
#> $FM_UD$yaml$FM$data_meta$data_types$other
#> $FM_UD$yaml$FM$data_meta$data_types$other$color
#> [1] "black"
#> 
#> $FM_UD$yaml$FM$data_meta$data_types$other$label
#> [1] "other"
#> 
#> 
#> 
#> 
#> $FM_UD$yaml$FM$workflows
#> $FM_UD$yaml$FM$workflows$example
#> $FM_UD$yaml$FM$workflows$example$group
#> [1] "Examples"
#> 
#> $FM_UD$yaml$FM$workflows$example$desc
#> [1] "Example Workflow"
#> 
#> $FM_UD$yaml$FM$workflows$example$require_ds
#> [1] TRUE
#> 
#> $FM_UD$yaml$FM$workflows$example$preload
#> [1] "file.path('.', 'example.yaml')"
#> 
#> 
#> 
#> $FM_UD$yaml$FM$labels
#> $FM_UD$yaml$FM$labels$ui_label
#> [1] "put labels here"
#> 
#> 
#> $FM_UD$yaml$FM$user_files
#> $FM_UD$yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $FM_UD$yaml$FM$logging
#> $FM_UD$yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $FM_UD$yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $FM_UD$yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $FM_UD$yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $FM_UD$yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $FM_UD$MC
#> $FM_UD$MC$module
#> $FM_UD$MC$module$type
#> [1] "UD"
#> 
#> $FM_UD$MC$module$depends
#> $FM_UD$MC$module$depends$id_ASM
#> [1] "ASM"
#> 
#> 
#> 
#> $FM_UD$MC$compact
#> $FM_UD$MC$compact$code
#> [1] FALSE
#> 
#> $FM_UD$MC$compact$preview
#> [1] TRUE
#> 
#> 
#> $FM_UD$MC$reporting
#> $FM_UD$MC$reporting$enabled
#> [1] FALSE
#> 
#> $FM_UD$MC$reporting$priority
#> [1] 1
#> 
#> 
#> $FM_UD$MC$labels
#> $FM_UD$MC$labels$default_ds
#> [1] "Original data set"
#> 
#> $FM_UD$MC$labels$workflow
#> [1] "Select workflow to run"
#> 
#> $FM_UD$MC$labels$run_wf
#> [1] "Run Workflow"
#> 
#> $FM_UD$MC$labels$upload_button
#> [1] "Select file"
#> 
#> $FM_UD$MC$labels$tab_main
#> [1] "Upload file"
#> 
#> $FM_UD$MC$labels$tab_code
#> [1] "Code"
#> 
#> $FM_UD$MC$labels$tab_preview
#> [1] "Data preview"
#> 
#> $FM_UD$MC$labels$switch_clean
#> [1] "Clean column names on upload"
#> 
#> $FM_UD$MC$labels$rwf_success
#> [1] "Workflow complete."
#> 
#> $FM_UD$MC$labels$busy
#> $FM_UD$MC$labels$busy$rwf
#> [1] "Running workflow be patient."
#> 
#> 
#> $FM_UD$MC$labels$msg_bad_extension
#> [1] "Incorrect file extension (===EXT===). Unable to load file  ===FILE===.\nOnly the following extesions are allowed: ===ALLOWEDEXT===."
#> 
#> 
#> $FM_UD$MC$errors
#> $FM_UD$MC$errors$no_workflows_found
#> [1] "No valid workflows have been found"
#> 
#> $FM_UD$MC$errors$no_ds_for_workflow
#> [1] "The selected workflow requires a dataset. Please upload a dataset first then run the workflow."
#> 
#> $FM_UD$MC$errors$no_ds_for_workflow_short
#> [1] "Workflow not run, no dataset."
#> 
#> $FM_UD$MC$errors$no_asm_state
#> [1] "Unable to read ASM state."
#> 
#> $FM_UD$MC$errors$ss_failed
#> [1] "Unable to save workflow file."
#> 
#> $FM_UD$MC$errors$ls_failed
#> [1] "Unable to load workflow file."
#> 
#> $FM_UD$MC$errors$rwf_failed
#> [1] "Workflow failed."
#> 
#> 
#> $FM_UD$MC$tooltips
#> $FM_UD$MC$tooltips$include
#> [1] TRUE
#> 
#> $FM_UD$MC$tooltips$show_code
#> [1] "Show data reading code"
#> 
#> 
#> $FM_UD$MC$formatting
#> $FM_UD$MC$formatting$preview
#> $FM_UD$MC$formatting$preview$width
#> [1] 800
#> 
#> $FM_UD$MC$formatting$preview$height
#> [1] 500
#> 
#> 
#> $FM_UD$MC$formatting$code
#> $FM_UD$MC$formatting$code$width
#> [1] 500
#> 
#> $FM_UD$MC$formatting$code$height
#> [1] 200
#> 
#> 
#> $FM_UD$MC$formatting$switch_clean
#> $FM_UD$MC$formatting$switch_clean$tooltip
#> [1] "Strange characters in column names can cause problems. When switched on, this will clean up the column headers and fix those that might break things."
#> 
#> $FM_UD$MC$formatting$switch_clean$tooltip_position
#> [1] "bottom"
#> 
#> 
#> $FM_UD$MC$formatting$workflow
#> $FM_UD$MC$formatting$workflow$width
#> [1] "200px"
#> 
#> $FM_UD$MC$formatting$workflow$liveSearch
#> [1] TRUE
#> 
#> $FM_UD$MC$formatting$workflow$size
#> [1] 10
#> 
#> 
#> $FM_UD$MC$formatting$btn_run_wf
#> $FM_UD$MC$formatting$btn_run_wf$size
#> [1] "sm"
#> 
#> $FM_UD$MC$formatting$btn_run_wf$width
#> [1] "200px"
#> 
#> $FM_UD$MC$formatting$btn_run_wf$block
#> [1] TRUE
#> 
#> $FM_UD$MC$formatting$btn_run_wf$tooltip
#> [1] "Run selected workflow"
#> 
#> $FM_UD$MC$formatting$btn_run_wf$tooltip_position
#> [1] "right"
#> 
#> 
#> 
#> $FM_UD$MC$allowed_extensions
#> [1] "csv"  "tsv"  "xls"  "xlsx"
#> 
#> $FM_UD$MC$code
#> $FM_UD$MC$code$packages
#> [1] "readxl"  "readr"   "janitor"
#> 
#> $FM_UD$MC$code$readOnly
#> [1] TRUE
#> 
#> $FM_UD$MC$code$mode
#> [1] "r"
#> 
#> $FM_UD$MC$code$preamble
#> [1] "# Loading data ------------------------------------------------------------"
#> 
#> 
#> $FM_UD$MC$ds_object_name
#> [1] "myDS"
#> 
#> $FM_UD$MC$clean_data
#> $FM_UD$MC$clean_data$enabled
#> [1] TRUE
#> 
#> $FM_UD$MC$clean_data$default
#> [1] TRUE
#> 
#> $FM_UD$MC$clean_data$code
#> [1] "# Make sure you put whatever packages you use here in the packages section\n# above.\n#\n# contents - data frame containing the loaded dataset.\n# code     - code used to load the dataset you need to append any cleaning\n#            code you generate  here.\n# object_name - Name of the dataset object used in generated code.\nif(system.file(package=\"janitor\") != \"\"){\n  contents = janitor::clean_names(contents, case=\"none\")\n  code = paste0(code, '\\n', object_name, ' = janitor::clean_names(', object_name,', case=\"none\")')\n}"
#> 
#> 
#> 
#> $FM_UD$UD
#> $FM_UD$UD$button_counters
#> $FM_UD$UD$button_counters$btn_run_wf
#> [1] 0
#> 
#> 
#> $FM_UD$UD$ui_hold
#> $FM_UD$UD$ui_hold$workflow
#> [1] FALSE
#> 
#> $FM_UD$UD$ui_hold$btn_run_wf
#> [1] FALSE
#> 
#> 
#> $FM_UD$UD$ui_ids
#> [1] "workflow"   "btn_run_wf"
#> 
#> $FM_UD$UD$isgood
#> [1] TRUE
#> 
#> $FM_UD$UD$clean
#> [1] TRUE
#> 
#> $FM_UD$UD$code
#> [1] "myDS = readxl::read_excel(path=\"TEST_DATA.xlsx\", sheet=\"DATA\")\nmyDS = janitor::clean_names(myDS, case=\"none\")"
#> 
#> $FM_UD$UD$object_name
#> [1] "myDS"
#> 
#> $FM_UD$UD$checksum
#> [1] "897d952fecbc804999396a96f9df4b20"
#> 
#> $FM_UD$UD$ui
#> $FM_UD$UD$ui$workflow
#> [1] ""
#> 
#> $FM_UD$UD$ui$btn_run_wf
#> [1] 0
#> 
#> 
#> $FM_UD$UD$ui_old
#> $FM_UD$UD$ui_old$workflow
#> [1] ""
#> 
#> $FM_UD$UD$ui_old$btn_run_wf
#> [1] 0
#> 
#> 
#> $FM_UD$UD$data_file_local
#> [1] "/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmpml8hrr/non_shiny/FM/TEST_DATA.xlsx"
#> 
#> $FM_UD$UD$data_file_ext
#> [1] "xlsx"
#> 
#> $FM_UD$UD$data_file
#> [1] "TEST_DATA.xlsx"
#> 
#> $FM_UD$UD$sheet
#> [1] "DATA"
#> 
#> $FM_UD$UD$sheets
#> [1] "DATA" "DDT" 
#> 
#> $FM_UD$UD$contents
#> # A tibble: 10,680 × 33
#>       ID TIME_DY TIME_HR NTIME_DY NTIME_HR  TIME     AMT    DV RECDESC     DOSE
#>    <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl>   <dbl> <dbl> <chr>      <dbl>
#>  1     1       0       0        0        0     0 3000000    0  dose     3000000
#>  2     1       0       0        0        0     0       0 1250  BM_ng_ml 3000000
#>  3     1       0       0        0        0     0       0    0  C_ng_ml  3000000
#>  4     1       1      24        1       24    24       0 1647  BM_ng_ml 3000000
#>  5     1       1      24        1       24    24       0  690. C_ng_ml  3000000
#>  6     1       2      48        2       48    48       0 1947  BM_ng_ml 3000000
#>  7     1       2      48        2       48    48       0  588. C_ng_ml  3000000
#>  8     1       3      72        3       72    72       0 2170  BM_ng_ml 3000000
#>  9     1       3      72        3       72    72       0  506. C_ng_ml  3000000
#> 10     1       4      96        4       96    96       0 2332  BM_ng_ml 3000000
#> # ℹ 10,670 more rows
#> # ℹ 23 more variables: DOSE_STR <chr>, Cohort <chr>, ROUTE <chr>, WT <dbl>,
#> #   SUBTYPE <chr>, SEX <chr>, F1 <dbl>, ka <dbl>, CL <dbl>, Vc <dbl>, Vp <dbl>,
#> #   Q <dbl>, BM_IC <dbl>, kdeg_BM <dbl>, Emax <dbl>, EC50 <dbl>, CMT <chr>,
#> #   DUR <dbl>, EVID <dbl>, NDose <chr>, DOSE_NUM <dbl>, SUBTYPE_ID <dbl>,
#> #   SEX_ID <dbl>
#> 
#> $FM_UD$UD$load_msg
#> <em>File loaded.</em>
#> 
#> 
#> $FM_UD$MOD_TYPE
#> [1] "UD"
#> 
#> $FM_UD$id
#> [1] "UD"
#> 
#> $FM_UD$FM_yaml_file
#> [1] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpSYRoUO/temp_libpath13c2570e2c951/formods/templates/formods.yaml"
#> 
#> $FM_UD$MOD_yaml_file
#> [1] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpSYRoUO/temp_libpath13c2570e2c951/formods/templates/UD.yaml"
#> 
#> $FM_UD$shiny_token
#> [1] "non_shiny"
#> 
#>