Creates Formatting for Datasets

Takes a data frame and information in the site configureation to produce formatting information to make it easier for the user to see data type information.

Usage

FM_fetch_data_format(df, state)

Arguments

df

Raw dataframe to be built into an rhandsontable.

state

Current module state after yaml file has been read.

Value

list with the following elements:

col_heads: Vector of formatting information for column headers to be use with rhandsontable.

Examples

# We need a module state object to use this function:
sess_res = UD_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:48559d397570afc9fba8ad9b252fa75a
state = sess_res$state

data_file_local =  system.file(package="formods", "test_data", "TEST_DATA.xlsx")
sheet           = "DATA"

df = readxl::read_excel(path=data_file_local, sheet=sheet)

FM_fetch_data_format(df, state)
#> $col_heads
#> $col_heads$ID
#> [1] "<span style='color:#3C8DBC'><b>ID</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$TIME_DY
#> [1] "<span style='color:#3C8DBC'><b>TIME_DY</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$TIME_HR
#> [1] "<span style='color:#3C8DBC'><b>TIME_HR</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$NTIME_DY
#> [1] "<span style='color:#3C8DBC'><b>NTIME_DY</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$NTIME_HR
#> [1] "<span style='color:#3C8DBC'><b>NTIME_HR</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$TIME
#> [1] "<span style='color:#3C8DBC'><b>TIME</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$AMT
#> [1] "<span style='color:#3C8DBC'><b>AMT</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$DV
#> [1] "<span style='color:#3C8DBC'><b>DV</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$RECDESC
#> [1] "<span style='color:#DD4B39'><b>RECDESC</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$DOSE
#> [1] "<span style='color:#3C8DBC'><b>DOSE</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$DOSE_STR
#> [1] "<span style='color:#DD4B39'><b>DOSE_STR</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$Cohort
#> [1] "<span style='color:#DD4B39'><b>Cohort</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$ROUTE
#> [1] "<span style='color:#DD4B39'><b>ROUTE</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$WT
#> [1] "<span style='color:#3C8DBC'><b>WT</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$SUBTYPE
#> [1] "<span style='color:#DD4B39'><b>SUBTYPE</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$SEX
#> [1] "<span style='color:#DD4B39'><b>SEX</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$F1
#> [1] "<span style='color:#3C8DBC'><b>F1</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$ka
#> [1] "<span style='color:#3C8DBC'><b>ka</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$CL
#> [1] "<span style='color:#3C8DBC'><b>CL</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$Vc
#> [1] "<span style='color:#3C8DBC'><b>Vc</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$Vp
#> [1] "<span style='color:#3C8DBC'><b>Vp</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$Q
#> [1] "<span style='color:#3C8DBC'><b>Q</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$BM_IC
#> [1] "<span style='color:#3C8DBC'><b>BM_IC</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$kdeg_BM
#> [1] "<span style='color:#3C8DBC'><b>kdeg_BM</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$Emax
#> [1] "<span style='color:#3C8DBC'><b>Emax</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$EC50
#> [1] "<span style='color:#3C8DBC'><b>EC50</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$CMT
#> [1] "<span style='color:#DD4B39'><b>CMT</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$DUR
#> [1] "<span style='color:#3C8DBC'><b>DUR</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$EVID
#> [1] "<span style='color:#3C8DBC'><b>EVID</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$NDose
#> [1] "<span style='color:#DD4B39'><b>NDose</b><br/><font size='-3'>text</font></span>"
#> 
#> $col_heads$DOSE_NUM
#> [1] "<span style='color:#3C8DBC'><b>DOSE_NUM</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$SUBTYPE_ID
#> [1] "<span style='color:#3C8DBC'><b>SUBTYPE_ID</b><br/><font size='-3'>num</font></span>"
#> 
#> $col_heads$SEX_ID
#> [1] "<span style='color:#3C8DBC'><b>SEX_ID</b><br/><font size='-3'>num</font></span>"
#> 
#> 
#> $col_subtext
#> $col_subtext$ID
#> [1] "num: 1,⋅⋅⋅,360"
#> 
#> $col_subtext$TIME_DY
#> [1] "num: 0,⋅⋅⋅,84"
#> 
#> $col_subtext$TIME_HR
#> [1] "num: 0,⋅⋅⋅,2016"
#> 
#> $col_subtext$NTIME_DY
#> [1] "num: 0,⋅⋅⋅,42"
#> 
#> $col_subtext$NTIME_HR
#> [1] "num: 0,⋅⋅⋅,1008"
#> 
#> $col_subtext$TIME
#> [1] "num: 0,⋅⋅⋅,2016"
#> 
#> $col_subtext$AMT
#> [1] "num: 0,⋅⋅⋅,5e+08"
#> 
#> $col_subtext$DV
#> [1] "num: 0,⋅⋅⋅,473500"
#> 
#> $col_subtext$RECDESC
#> [1] "text: BM_ng_ml, C_ng_ml, dose"
#> 
#> $col_subtext$DOSE
#> [1] "num: 3,⋅⋅⋅,500"
#> 
#> $col_subtext$DOSE_STR
#> [1] "text: 10 mg,⋅⋅⋅,500 mg"
#> 
#> $col_subtext$Cohort
#> [1] "text: MD 10 mg IV,⋅⋅⋅,SD 500 mg IV"
#> 
#> $col_subtext$ROUTE
#> [1] "text: IV, SC"
#> 
#> $col_subtext$WT
#> [1] "num: 49.6,⋅⋅⋅,109"
#> 
#> $col_subtext$SUBTYPE
#> [1] "text: HV, PT"
#> 
#> $col_subtext$SEX
#> [1] "text: Female, Male"
#> 
#> $col_subtext$F1
#> [1] "num: 0.744"
#> 
#> $col_subtext$ka
#> [1] "num: 0.0349,⋅⋅⋅,2.55"
#> 
#> $col_subtext$CL
#> [1] "num: 0.0758,⋅⋅⋅,0.454"
#> 
#> $col_subtext$Vc
#> [1] "num: 1.41,⋅⋅⋅,12.7"
#> 
#> $col_subtext$Vp
#> [1] "num: 1.23,⋅⋅⋅,7.44"
#> 
#> $col_subtext$Q
#> [1] "num: 0.0402,⋅⋅⋅,7.53"
#> 
#> $col_subtext$BM_IC
#> [1] "num: 543,⋅⋅⋅,2190"
#> 
#> $col_subtext$kdeg_BM
#> [1] "num: 0.1"
#> 
#> $col_subtext$Emax
#> [1] "num: 44,⋅⋅⋅,287"
#> 
#> $col_subtext$EC50
#> [1] "num: 300"
#> 
#> $col_subtext$CMT
#> [1] "text: Ac,⋅⋅⋅,C_ng_ml"
#> 
#> $col_subtext$DUR
#> [1] "num: 0"
#> 
#> $col_subtext$EVID
#> [1] "num: 0, 1"
#> 
#> $col_subtext$NDose
#> [1] "text: MD, SD"
#> 
#> $col_subtext$DOSE_NUM
#> [1] "num: 1,⋅⋅⋅,4"
#> 
#> $col_subtext$SUBTYPE_ID
#> [1] "num: 0, 1"
#> 
#> $col_subtext$SEX_ID
#> [1] "num: 0, 1"
#> 
#> 
#> $col_info
#> $col_info$ID
#> $col_info$ID$color
#> [1] "#3C8DBC"
#> 
#> $col_info$ID$label
#> [1] "num"
#> 
#> $col_info$ID$factor
#> [1] FALSE
#> 
#> 
#> $col_info$TIME_DY
#> $col_info$TIME_DY$color
#> [1] "#3C8DBC"
#> 
#> $col_info$TIME_DY$label
#> [1] "num"
#> 
#> $col_info$TIME_DY$factor
#> [1] FALSE
#> 
#> 
#> $col_info$TIME_HR
#> $col_info$TIME_HR$color
#> [1] "#3C8DBC"
#> 
#> $col_info$TIME_HR$label
#> [1] "num"
#> 
#> $col_info$TIME_HR$factor
#> [1] FALSE
#> 
#> 
#> $col_info$NTIME_DY
#> $col_info$NTIME_DY$color
#> [1] "#3C8DBC"
#> 
#> $col_info$NTIME_DY$label
#> [1] "num"
#> 
#> $col_info$NTIME_DY$factor
#> [1] FALSE
#> 
#> 
#> $col_info$NTIME_HR
#> $col_info$NTIME_HR$color
#> [1] "#3C8DBC"
#> 
#> $col_info$NTIME_HR$label
#> [1] "num"
#> 
#> $col_info$NTIME_HR$factor
#> [1] FALSE
#> 
#> 
#> $col_info$TIME
#> $col_info$TIME$color
#> [1] "#3C8DBC"
#> 
#> $col_info$TIME$label
#> [1] "num"
#> 
#> $col_info$TIME$factor
#> [1] FALSE
#> 
#> 
#> $col_info$AMT
#> $col_info$AMT$color
#> [1] "#3C8DBC"
#> 
#> $col_info$AMT$label
#> [1] "num"
#> 
#> $col_info$AMT$factor
#> [1] FALSE
#> 
#> 
#> $col_info$DV
#> $col_info$DV$color
#> [1] "#3C8DBC"
#> 
#> $col_info$DV$label
#> [1] "num"
#> 
#> $col_info$DV$factor
#> [1] FALSE
#> 
#> 
#> $col_info$RECDESC
#> $col_info$RECDESC$color
#> [1] "#DD4B39"
#> 
#> $col_info$RECDESC$label
#> [1] "text"
#> 
#> $col_info$RECDESC$factor
#> [1] FALSE
#> 
#> 
#> $col_info$DOSE
#> $col_info$DOSE$color
#> [1] "#3C8DBC"
#> 
#> $col_info$DOSE$label
#> [1] "num"
#> 
#> $col_info$DOSE$factor
#> [1] FALSE
#> 
#> 
#> $col_info$DOSE_STR
#> $col_info$DOSE_STR$color
#> [1] "#DD4B39"
#> 
#> $col_info$DOSE_STR$label
#> [1] "text"
#> 
#> $col_info$DOSE_STR$factor
#> [1] FALSE
#> 
#> 
#> $col_info$Cohort
#> $col_info$Cohort$color
#> [1] "#DD4B39"
#> 
#> $col_info$Cohort$label
#> [1] "text"
#> 
#> $col_info$Cohort$factor
#> [1] FALSE
#> 
#> 
#> $col_info$ROUTE
#> $col_info$ROUTE$color
#> [1] "#DD4B39"
#> 
#> $col_info$ROUTE$label
#> [1] "text"
#> 
#> $col_info$ROUTE$factor
#> [1] FALSE
#> 
#> 
#> $col_info$WT
#> $col_info$WT$color
#> [1] "#3C8DBC"
#> 
#> $col_info$WT$label
#> [1] "num"
#> 
#> $col_info$WT$factor
#> [1] FALSE
#> 
#> 
#> $col_info$SUBTYPE
#> $col_info$SUBTYPE$color
#> [1] "#DD4B39"
#> 
#> $col_info$SUBTYPE$label
#> [1] "text"
#> 
#> $col_info$SUBTYPE$factor
#> [1] FALSE
#> 
#> 
#> $col_info$SEX
#> $col_info$SEX$color
#> [1] "#DD4B39"
#> 
#> $col_info$SEX$label
#> [1] "text"
#> 
#> $col_info$SEX$factor
#> [1] FALSE
#> 
#> 
#> $col_info$F1
#> $col_info$F1$color
#> [1] "#3C8DBC"
#> 
#> $col_info$F1$label
#> [1] "num"
#> 
#> $col_info$F1$factor
#> [1] FALSE
#> 
#> 
#> $col_info$ka
#> $col_info$ka$color
#> [1] "#3C8DBC"
#> 
#> $col_info$ka$label
#> [1] "num"
#> 
#> $col_info$ka$factor
#> [1] FALSE
#> 
#> 
#> $col_info$CL
#> $col_info$CL$color
#> [1] "#3C8DBC"
#> 
#> $col_info$CL$label
#> [1] "num"
#> 
#> $col_info$CL$factor
#> [1] FALSE
#> 
#> 
#> $col_info$Vc
#> $col_info$Vc$color
#> [1] "#3C8DBC"
#> 
#> $col_info$Vc$label
#> [1] "num"
#> 
#> $col_info$Vc$factor
#> [1] FALSE
#> 
#> 
#> $col_info$Vp
#> $col_info$Vp$color
#> [1] "#3C8DBC"
#> 
#> $col_info$Vp$label
#> [1] "num"
#> 
#> $col_info$Vp$factor
#> [1] FALSE
#> 
#> 
#> $col_info$Q
#> $col_info$Q$color
#> [1] "#3C8DBC"
#> 
#> $col_info$Q$label
#> [1] "num"
#> 
#> $col_info$Q$factor
#> [1] FALSE
#> 
#> 
#> $col_info$BM_IC
#> $col_info$BM_IC$color
#> [1] "#3C8DBC"
#> 
#> $col_info$BM_IC$label
#> [1] "num"
#> 
#> $col_info$BM_IC$factor
#> [1] FALSE
#> 
#> 
#> $col_info$kdeg_BM
#> $col_info$kdeg_BM$color
#> [1] "#3C8DBC"
#> 
#> $col_info$kdeg_BM$label
#> [1] "num"
#> 
#> $col_info$kdeg_BM$factor
#> [1] FALSE
#> 
#> 
#> $col_info$Emax
#> $col_info$Emax$color
#> [1] "#3C8DBC"
#> 
#> $col_info$Emax$label
#> [1] "num"
#> 
#> $col_info$Emax$factor
#> [1] FALSE
#> 
#> 
#> $col_info$EC50
#> $col_info$EC50$color
#> [1] "#3C8DBC"
#> 
#> $col_info$EC50$label
#> [1] "num"
#> 
#> $col_info$EC50$factor
#> [1] FALSE
#> 
#> 
#> $col_info$CMT
#> $col_info$CMT$color
#> [1] "#DD4B39"
#> 
#> $col_info$CMT$label
#> [1] "text"
#> 
#> $col_info$CMT$factor
#> [1] FALSE
#> 
#> 
#> $col_info$DUR
#> $col_info$DUR$color
#> [1] "#3C8DBC"
#> 
#> $col_info$DUR$label
#> [1] "num"
#> 
#> $col_info$DUR$factor
#> [1] FALSE
#> 
#> 
#> $col_info$EVID
#> $col_info$EVID$color
#> [1] "#3C8DBC"
#> 
#> $col_info$EVID$label
#> [1] "num"
#> 
#> $col_info$EVID$factor
#> [1] FALSE
#> 
#> 
#> $col_info$NDose
#> $col_info$NDose$color
#> [1] "#DD4B39"
#> 
#> $col_info$NDose$label
#> [1] "text"
#> 
#> $col_info$NDose$factor
#> [1] FALSE
#> 
#> 
#> $col_info$DOSE_NUM
#> $col_info$DOSE_NUM$color
#> [1] "#3C8DBC"
#> 
#> $col_info$DOSE_NUM$label
#> [1] "num"
#> 
#> $col_info$DOSE_NUM$factor
#> [1] FALSE
#> 
#> 
#> $col_info$SUBTYPE_ID
#> $col_info$SUBTYPE_ID$color
#> [1] "#3C8DBC"
#> 
#> $col_info$SUBTYPE_ID$label
#> [1] "num"
#> 
#> $col_info$SUBTYPE_ID$factor
#> [1] FALSE
#> 
#> 
#> $col_info$SEX_ID
#> $col_info$SEX_ID$color
#> [1] "#3C8DBC"
#> 
#> $col_info$SEX_ID$label
#> [1] "num"
#> 
#> $col_info$SEX_ID$factor
#> [1] FALSE
#> 
#> 
#>