Generate Code and Load DS

Generates the code for loading a dataset and returns both the code and the contents

Usage

UD_ds_read(
  state,
  data_file_ext = NULL,
  data_file_local = NULL,
  data_file = NULL,
  sheets = NULL,
  sheet = NULL
)

Arguments

state

UD state from UD_fetch_state()

data_file_ext

File extension of the uploaded file (e.g. "xlsx", "csv", etc).

data_file_local

Full path to the data file on the server.

data_file

Dataset file name without the path.

sheets

If the uploaded file is an excel file, this is a character vector of the sheets present in that file.

sheet

If the uploaded file is an excel file, this is the currently selected sheet.

Value

list with the elements of the dataset (contents, object_name, code, and isgood)

Examples

# We need a module state object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:b4bd907af4f3e3b571027d1230bf99f1
state = sess_res$state

# This is the full path to a test data file:
data_file_local  =  system.file(package="formods", "test_data", "TEST_DATA.xlsx")

# Excel file extension
data_file_ext    = "xlsx"

# Base file name
data_file        = "TEST_DATA.xlsx"

# Excel files need a sheet specification:
sheet            = "DATA"

# We will also attach the sheets along with it
sheets = readxl::excel_sheets(data_file_local)

ds_read_res = UD_ds_read(state, 
  data_file_ext   = data_file_ext,
  data_file_local = data_file_local,
  data_file       = data_file,
  sheets          = sheets,
  sheet          = sheet)

ds_read_res
#> $contents
#> # A tibble: 18,360 × 33
#>       ID  TIME NTIME…¹ NTIME…² TIME_DY TIME_HR   AMT    DV RECDESC  DOSE DOSE_…³
#>    <dbl> <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl> <chr>   <dbl> <chr>  
#>  1     1  0     0         0     0         0      3e6    0  dose        3 3 mg   
#>  2     1  0.25  0.0104    0.25  0.0104    0.25   0   1255  BM_ng_…     3 3 mg   
#>  3     1  0.25  0.0104    0.25  0.0104    0.25   0    816  C_ng_ml     3 3 mg   
#>  4     1  0.5   0.0208    0.5   0.0208    0.5    0   1259  BM_ng_…     3 3 mg   
#>  5     1  0.5   0.0208    0.5   0.0208    0.5    0    814. C_ng_ml     3 3 mg   
#>  6     1  1     0.0417    1     0.0417    1      0   1269  BM_ng_…     3 3 mg   
#>  7     1  1     0.0417    1     0.0417    1      0    812. C_ng_ml     3 3 mg   
#>  8     1  2     0.0833    2     0.0833    2      0   1287  BM_ng_…     3 3 mg   
#>  9     1  2     0.0833    2     0.0833    2      0    806. C_ng_ml     3 3 mg   
#> 10     1  4     0.167     4     0.167     4      0   1324  BM_ng_…     3 3 mg   
#> # … with 18,350 more rows, 22 more variables: Cohort <chr>, ROUTE <chr>,
#> #   WT <dbl>, SUBTYPE <chr>, SEX <chr>, F1 <dbl>, ka <dbl>, CL <dbl>, Vc <dbl>,
#> #   Vp <dbl>, Q <dbl>, BM_IC <dbl>, kdeg_BM <dbl>, Emax <dbl>, EC50 <dbl>,
#> #   CMT <chr>, DUR <dbl>, EVID <dbl>, NDose <chr>, DOSE_NUM <dbl>,
#> #   SUBTYPE_ID <dbl>, SEX_ID <dbl>, and abbreviated variable names ¹​NTIME_DY,
#> #   ²​NTIME_HR, ³​DOSE_STR
#> 
#> $object_name
#> [1] "myDS"
#> 
#> $isgood
#> [1] TRUE
#> 
#> $code
#> [1] "myDS = readxl::read_excel(path=\"TEST_DATA.xlsx\", sheet=\"DATA\")"
#>