Fetch State Manager State

Merges default app options with the changes made in the UI

Usage

ASM_fetch_state(id, input, session, FM_yaml_file, MOD_yaml_file)

Arguments

id

Shiny module ID

input

Shiny input variable

session

Shiny session variable

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

Value

list containing the current state of the app including default values from the yaml file as well as any changes made by the user. The list has the following structure:

yaml: Full contents of the supplied yaml file.

MC: Module components of the yaml file.

ASM:

isgood: Boolean object indicating if the file was successfully loaded.

checksum: This is an MD5 sum of the loaded state file

MOD_TYPE: Character data containing the type of module "ASM"

id: Character data containing the module id module in the session variable.

FM_yaml_file: App configuration file with FM as main section.

MOD_yaml_file: Module configuration file with MC as main section.

Examples

# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = ASM_test_mksession(session=list())
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:d93bd3619945bd9ec762101a2a286346
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DW:   dest:   file.path("config","report.docx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DW:   dest:   file.path("config","report.pptx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DW:   dest:   file.path("config","report.yaml")
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85
#> → DW: module checksum updated:a5d1d3b8a436befcfa119b5d72d3697b
#> → DW: module checksum updated:48d9e35c3f568c7a6f72914538d4ecd4
#> → DW: module checksum updated:680bde3fe602684945b93dda69abc020
#> → DW: module checksum updated:1f571c6eabb372eff6c1ef806344b2e4
#> → DW: module checksum updated:a1632ef431da47f98334a9281270a3fd
#> → DW: module checksum updated:518e56af743d086ea4b1b6070bf70472
#> → DW: module checksum updated:460fd6e3a4a46f56f17bde8ff4ef5c01
#> → DW: module checksum updated:12335ca5ced95462c2ed1df919a72d15
#> → DW: module checksum updated:0794933dac491313b32ca4bbfd996641
#> → DW: module checksum updated:75abe1a1cfc787fc68049b40c9e5749e
#> → DW: module checksum updated:070868f0125fd16b3e648dc05be84cd0
#> → DW: module checksum updated:1da012ee43c59bf8bedc7376511d179e
#> → DW: module checksum updated:6dfd06414929a762cc4f07d6d428289b
#> → DW: module checksum updated:83f7106b33c98e7650432749c16d7a36
#> → DW: module checksum updated:32652abd1996af932cbdf304cc509bd0
#> → DW: module checksum updated:dfa171056875bc474e03e0c911ace2b6
#> → DW: module checksum updated:5d4dfb7affd116c593a856034804aa04
#> → DW: module checksum updated:f47016aa2630055e8a1ae44ab79cabd2
#> → DW: module checksum updated:46cc2c9503359a39be98c3c0b889851a
#> → DW: module checksum updated:58ed3ad9e797e77eba05a2576b78e50f
#> → DW: module checksum updated:aaf6ebcd2be24f05cd9b4776ef042986
#> → DW: module checksum updated:afbbf99a8ed5f01191217b8370603a9d
#> → DW: module checksum updated:086f760681a2b47a9a1dfd572f876fcd
#> → DW: module checksum updated:8c045cb5e0fe9bd74e84d87d3c6725b5
#> → DW: module checksum updated:779a7d3c81e58ea4c8a756c2e422fee0
#> → DW: module checksum updated:104438a337e6617c4504477506e90863
#> → DW: module checksum updated:1bd57823c76715ccceab0be02ed7a79e
#> → DW: module checksum updated:8b901fc11d897dd27c774accf0eb7097
#> → DW: module checksum updated:7d55b640df1207dcdc56a10c9b8d8b87
#> → DW: module checksum updated:8d91f192ea994a133c8e2fbb7ce7c11a
#> → DW: module checksum updated:d5c81022154ffb7347d7a30d66a38f60
#> → FG: including file
#> → FG:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → FG:   dest:   file.path("config","report.docx")
#> → FG: including file
#> → FG:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → FG:   dest:   file.path("config","report.pptx")
#> → FG: including file
#> → FG:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → FG:   dest:   file.path("config","report.yaml")
#> → FG: State initialized
#> → FG: module checksum updated:a77ad23e318409eb0b536b8e2bd9f54b
#> → FG: module checksum updated:03d39b733e5a2351212a4a33afe671d6
#> → FG: module checksum updated:81c30e5c56b085fd7c79dc71bff5c525
#> → FG: module checksum updated:244255d313153e3a5fbd9f8ecbe52a69
#> → FG: module checksum updated:4f64fcf4f901b96851bcb46f1353bc27
#> → FG: module checksum updated:d4f020216762801f697ece28ab861b5a
#> → FG: module checksum updated:51acbb56d0debac8a2c2b00db3b0a218
#> → FG: module checksum updated:f931de076b31f85955362233caafa163
#> → FG: module checksum updated:935b207dc73012251d694a06baf2ba7f
#> → FG: module checksum updated:4c4b2dc4f041cedeecdfb2d26398b135
#> → FG: module checksum updated:70a618a5d20c3695ea2ac50b0322e853
#> → FG: module checksum updated:7fccdf48d5c56256cbc7dbd172a7b918
#> → FG: module checksum updated:ce1c73f8ea3eac8db989b4c08f0800f9
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
session = sess_res$session
input   = sess_res$input

# Configuration files
FM_yaml_file  = system.file(package = "formods", "templates", "formods.yaml")
MOD_yaml_file = system.file(package = "formods", "templates", "ASM.yaml")

# We need to specify the ID of the ASM module
id = "ASM"

state = ASM_fetch_state(id           = id,
                        input        = input,
                        session      = session,
                        FM_yaml_file = FM_yaml_file,
                        MOD_yaml_file = MOD_yaml_file)

state
#> $yaml
#> $yaml$FM
#> $yaml$FM$include
#> $yaml$FM$include$files
#> $yaml$FM$include$files[[1]]
#> $yaml$FM$include$files[[1]]$file
#> $yaml$FM$include$files[[1]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\")"
#> 
#> $yaml$FM$include$files[[1]]$file$dest
#> [1] "file.path(\"config\",\"report.docx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[2]]
#> $yaml$FM$include$files[[2]]$file
#> $yaml$FM$include$files[[2]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\")"
#> 
#> $yaml$FM$include$files[[2]]$file$dest
#> [1] "file.path(\"config\",\"report.pptx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[3]]
#> $yaml$FM$include$files[[3]]$file
#> $yaml$FM$include$files[[3]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\")"
#> 
#> $yaml$FM$include$files[[3]]$file$dest
#> [1] "file.path(\"config\",\"report.yaml\")"
#> 
#> 
#> 
#> 
#> 
#> $yaml$FM$deployed
#> [1] FALSE
#> 
#> $yaml$FM$code
#> $yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $yaml$FM$code$gen_file
#> [1] "run_analysis.R"
#> 
#> $yaml$FM$code$gen_preamble
#> [1] "# formods automated output ------------------------------------------------\n# https://formods.ubiquity.tools/\nrm(list=ls())"
#> 
#> $yaml$FM$code$packages
#> [1] "onbrand" "writexl"
#> 
#> 
#> $yaml$FM$notifications
#> $yaml$FM$notifications$config
#> $yaml$FM$notifications$config$success
#> $yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $yaml$FM$notifications$config$failure
#> $yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $yaml$FM$notifications$config$info
#> $yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $yaml$FM$notifications$config$warning
#> $yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $yaml$FM$reporting
#> $yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $yaml$FM$reporting$content_init
#> $yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.docx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> $yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.pptx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> 
#> 
#> $yaml$FM$ui
#> $yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $yaml$FM$ui$select_size
#> [1] 10
#> 
#> $yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $yaml$FM$data_meta
#> $yaml$FM$data_meta$data_header
#> [1] "<span style='color:===COLOR==='><b>===NAME===</b><br/><font size='-3'>===LABEL===</font></span>"
#> 
#> $yaml$FM$data_meta$subtext
#> [1] "===LABEL===: ===RANGE==="
#> 
#> $yaml$FM$data_meta$many_sep
#> [1] ",⋅⋅⋅,"
#> 
#> $yaml$FM$data_meta$data_types
#> $yaml$FM$data_meta$data_types$character
#> $yaml$FM$data_meta$data_types$character$color
#> [1] "#DD4B39"
#> 
#> $yaml$FM$data_meta$data_types$character$label
#> [1] "text"
#> 
#> 
#> $yaml$FM$data_meta$data_types$double
#> $yaml$FM$data_meta$data_types$double$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$double$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$integer
#> $yaml$FM$data_meta$data_types$integer$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$integer$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$other
#> $yaml$FM$data_meta$data_types$other$color
#> [1] "black"
#> 
#> $yaml$FM$data_meta$data_types$other$label
#> [1] "other"
#> 
#> 
#> 
#> 
#> $yaml$FM$labels
#> $yaml$FM$labels$default_ds
#> [1] "Original data set"
#> 
#> 
#> $yaml$FM$user_files
#> $yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $yaml$FM$logging
#> $yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $MC
#> $MC$code
#> $MC$code$readOnly
#> [1] TRUE
#> 
#> $MC$code$mode
#> [1] "r"
#> 
#> 
#> $MC$reporting
#> $MC$reporting$enabled
#> [1] FALSE
#> 
#> $MC$reporting$priority
#> [1] 1
#> 
#> 
#> $MC$formatting
#> $MC$formatting$input_load_state
#> $MC$formatting$input_load_state$width
#> [1] "100%"
#> 
#> 
#> $MC$formatting$ui_asm_save_name
#> $MC$formatting$ui_asm_save_name$width
#> [1] "100%"
#> 
#> 
#> $MC$formatting$button_state_save
#> $MC$formatting$button_state_save$size
#> [1] "sm"
#> 
#> $MC$formatting$button_state_save$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_rpt_xlsx
#> $MC$formatting$button_rpt_xlsx$size
#> [1] "sm"
#> 
#> $MC$formatting$button_rpt_xlsx$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_rpt_pptx
#> $MC$formatting$button_rpt_pptx$size
#> [1] "sm"
#> 
#> $MC$formatting$button_rpt_pptx$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_rpt_docx
#> $MC$formatting$button_rpt_docx$size
#> [1] "sm"
#> 
#> $MC$formatting$button_rpt_docx$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$switch_gen_reports
#> $MC$formatting$switch_gen_reports$default
#> [1] FALSE
#> 
#> $MC$formatting$switch_gen_reports$right
#> [1] TRUE
#> 
#> $MC$formatting$switch_gen_reports$inline
#> [1] TRUE
#> 
#> $MC$formatting$switch_gen_reports$width
#> NULL
#> 
#> 
#> 
#> $MC$labels
#> $MC$labels$ui_asm_save_name
#> [1] "Save Analysis"
#> 
#> $MC$labels$ui_asm_save_name_ph
#> [1] "my_analysis"
#> 
#> $MC$labels$ui_asm_switch_gen_rpts
#> [1] "Generate reports when saving."
#> 
#> $MC$labels$save_state
#> [1] "Save"
#> 
#> $MC$labels$input_load_state
#> [1] "Load Analysis"
#> 
#> $MC$labels$upload_button
#> [1] "Load"
#> 
#> $MC$labels$upload_placeholder
#> [1] "Select analysis file"
#> 
#> $MC$labels$ui_asm_rpt_xlsx
#> [1] "xlsx"
#> 
#> $MC$labels$ui_asm_rpt_pptx
#> [1] "pptx"
#> 
#> $MC$labels$ui_asm_rpt_docx
#> [1] "docx"
#> 
#> $MC$labels$busy
#> $MC$labels$busy$saving_state
#> [1] "Saving app state"
#> 
#> $MC$labels$busy$xlsx
#> [1] "Generating report: Excel"
#> 
#> $MC$labels$busy$docx
#> [1] "Generating report: Word"
#> 
#> $MC$labels$busy$pptx
#> [1] "Generating report: PowerPoint"
#> 
#> 
#> 
#> $MC$errors
#> $MC$errors$bad_input
#> [1] "Input error message"
#> 
#> 
#> $MC$tooltips
#> $MC$tooltips$include
#> [1] TRUE
#> 
#> $MC$tooltips$components
#> $MC$tooltips$components$input_tip
#> [1] "My tool tip"
#> 
#> 
#> 
#> 
#> $ASM
#> $ASM$button_counters
#> $ASM$button_counters$button_state_save
#> [1] 0
#> 
#> $ASM$button_counters$button_rpt_xlsx
#> [1] 0
#> 
#> $ASM$button_counters$button_rpt_docx
#> [1] 0
#> 
#> $ASM$button_counters$button_rpt_pptx
#> [1] 0
#> 
#> 
#> $ASM$ui_hold
#> list()
#> 
#> $ASM$ui_ids
#> [1] "button_state_save" "button_rpt_xlsx"   "button_rpt_docx"  
#> [4] "button_rpt_pptx"   "ui_asm_save_name"  "switch_gen_rpts"  
#> 
#> $ASM$isgood
#> [1] TRUE
#> 
#> $ASM$checksum
#> [1] "f9e884084b84794d762a535f3facec85"
#> 
#> $ASM$ui
#> $ASM$ui$button_state_save
#> [1] ""
#> 
#> $ASM$ui$button_rpt_xlsx
#> [1] ""
#> 
#> $ASM$ui$button_rpt_docx
#> [1] ""
#> 
#> $ASM$ui$button_rpt_pptx
#> [1] ""
#> 
#> $ASM$ui$ui_asm_save_name
#> [1] ""
#> 
#> $ASM$ui$switch_gen_rpts
#> [1] ""
#> 
#> 
#> 
#> $MOD_TYPE
#> [1] "ASM"
#> 
#> $id
#> [1] "ASM"
#> 
#> $FM_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpOBgZvG/temp_libpath1221e5b7d7/formods/templates/formods.yaml"
#> 
#> $MOD_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpOBgZvG/temp_libpath1221e5b7d7/formods/templates/ASM.yaml"
#>