Attach Data Set to DW State

Attaches a dataset to the DW state supplied.

Usage

DW_attach_ds(state, id_UD, session)

Arguments

state

DW state from DW_fetch_state()

id_UD

ID string for the upload data module used to handle uploads

session

Shiny session variable

Value

state with data set attached

Examples

# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = DW_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:631af84e28478adc1628dd2c9fb22662
#> → DW: State initialized
#> → DW: module checksum updated:f5866ed2a560dbc564dec9145314b519
#> → DW: module checksum updated:4911af9c225aac6361acea8f26bbfac9
#> → DW: module checksum updated:c9b0caeab52fb59e0dcbb88ee042729a
#> → DW: module checksum updated:b7f3e08819cabbd242f5bebe1290714b
#> → DW: module checksum updated:8d8944be2daf4ea8aa3b1c1c707cee45
#> → DW: module checksum updated:a6367d7e4e002eca11050774144105f8
#> → DW: module checksum updated:4a381cdf6502fe13dca9c009a25b4418
#> → DW: module checksum updated:1050496cc0eca17d51102e051220ec49
#> → DW: module checksum updated:df684a1d4fbbd8bd33749607f9b982c0
#> → DW: module checksum updated:e3054989800f1e8ca24386371b52accc
#> → DW: module checksum updated:3b43dbda29ded3a3a3ef514ccf0c0578
#> → DW: module checksum updated:550d10d9e424d549878ab4ba5343896e
#> → DW: module checksum updated:5d2a75ed77a22f4ff8cc2a0d2bf2e3e5
#> → DW: module checksum updated:0272b5040b80ff35adf3500fdb60d8c7
#> → DW: module checksum updated:173ecf7d96c0bbb0962c7846df0d4bca
session = sess_res$session
input   = sess_res$input

# We also need a state variable
state = sess_res$state

# We need to identify the UD module with the data
id_UD = "UD"
state = DW_attach_ds(state, id_UD, session)