Initialize DW Module State

Creates a list of the initialized module state

Usage

DW_init_state(FM_yaml_file, MOD_yaml_file, id, id_UD, react_state)

Arguments

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

id

Shiny module ID

id_UD

ID string for the upload data module used to handle uploads or the name of the list element in react_state where the data set is stored.

react_state

Variable passed to server to allow reaction outside of module (NULL)

Value

list containing an empty DW state

Examples

state = DW_init_state(
   FM_yaml_file  = system.file(package = "formods",
                               "templates",
                               "formods.yaml"),
   MOD_yaml_file = system.file(package = "formods",
                               "templates",
                               "DW.yaml"),
   id              = "DW",
   id_UD           = "UD",
   react_state     = NULL)
#> → DW: State initialized
#> → DW: module checksum updated:2f88e16c4f21bfcb290b2a8288c05bd0
state
#> $yaml
#> $yaml$FM
#> $yaml$FM$code
#> $yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $yaml$FM$code$gen_file
#> [1] "formods_export.R"
#> 
#> $yaml$FM$code$gen_preamble
#> [1] "# formods automated output\n# https://formods.ubiquity.tools/\nrm(list=ls())\nlibrary(ggplot2)\nlibrary(ggforce)\nlibrary(readxl)\nlibrary(readr)\nlibrary(dplyr)\nlibrary(tidyr)"
#> 
#> $yaml$FM$code$gen_mods
#> $yaml$FM$code$gen_mods$UD
#> [1] "# Loading data"
#> 
#> $yaml$FM$code$gen_mods$DW
#> [1] "# Wrangling data"
#> 
#> $yaml$FM$code$gen_mods$FG
#> [1] "# Generating figures"
#> 
#> 
#> 
#> $yaml$FM$notifications
#> $yaml$FM$notifications$config
#> $yaml$FM$notifications$config$success
#> $yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $yaml$FM$notifications$config$failure
#> $yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $yaml$FM$notifications$config$info
#> $yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $yaml$FM$notifications$config$warning
#> $yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $yaml$FM$reporting
#> $yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $yaml$FM$reporting$content_init
#> $yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\"),\n  mapping  = file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\"))"
#> 
#> $yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\"),\n  mapping  = file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\"))"
#> 
#> 
#> 
#> $yaml$FM$ui
#> $yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $yaml$FM$labels
#> $yaml$FM$labels$default_ds
#> [1] "Original data set"
#> 
#> 
#> $yaml$FM$user_files
#> $yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $yaml$FM$logging
#> $yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $MC
#> $MC$compact
#> $MC$compact$code
#> [1] TRUE
#> 
#> $MC$compact$preview
#> [1] TRUE
#> 
#> 
#> $MC$reporting
#> $MC$reporting$enabled
#> [1] TRUE
#> 
#> $MC$reporting$priority
#> [1] 1
#> 
#> 
#> $MC$notifications
#> $MC$notifications$new_dw_element
#> [1] "Added element: ===ACTION=== (===DESC===)"
#> 
#> $MC$notifications$del_dw_element
#> [1] "Removed element: ===ACTION=== (===DESC===)"
#> 
#> 
#> $MC$tooltips
#> $MC$tooltips$include
#> [1] TRUE
#> 
#> $MC$tooltips$show_code
#> [1] "Show data wrangling code"
#> 
#> $MC$tooltips$dw_elements
#> [1] "Current wrangling elements"
#> 
#> 
#> $MC$labels
#> $MC$labels$tab_main
#> [1] "Wrangle"
#> 
#> $MC$labels$tab_code
#> [1] "Code"
#> 
#> $MC$labels$add_element
#> [1] "Add Operation/Action"
#> 
#> $MC$labels$fds_filter_column
#> [1] "Filter by Column"
#> 
#> $MC$labels$fds_filter_operator
#> [1] "Condition"
#> 
#> $MC$labels$fds_filter_rhs
#> [1] "Matching"
#> 
#> $MC$labels$fds_mutate_column
#> [1] "New or current column to alter"
#> 
#> $MC$labels$fds_mutate_rhs
#> [1] "New Value"
#> 
#> $MC$labels$fds_rename_column
#> [1] "Curret Column"
#> 
#> $MC$labels$fds_rename_rhs
#> [1] "New Name"
#> 
#> $MC$labels$fds_group_column
#> [1] "Columns to group"
#> 
#> $MC$labels$fds_longer_column
#> [1] "Columns to make longer"
#> 
#> $MC$labels$fds_longer_names
#> [1] "New column to contain names"
#> 
#> $MC$labels$fds_longer_values
#> [1] "New column to contain values"
#> 
#> $MC$labels$fds_wider_names
#> [1] "Column with names"
#> 
#> $MC$labels$fds_wider_values
#> [1] "Column with values"
#> 
#> $MC$labels$new_dw
#> [1] "New"
#> 
#> $MC$labels$del_dw
#> [1] "Delete"
#> 
#> $MC$labels$copy_dw
#> [1] "Copy"
#> 
#> $MC$labels$save_dw
#> [1] "Save"
#> 
#> $MC$labels$ungroup_data
#> [1] "Removing current grouping"
#> 
#> $MC$labels$keep_onerow
#> [1] "Keep first row in current grouping"
#> 
#> $MC$labels$no_dataset
#> [1] "No dataset loaded!"
#> 
#> $MC$labels$no_dw_elements
#> [1] "No data wrangling elements defined yet"
#> 
#> $MC$labels$busy
#> $MC$labels$busy$dv_update
#> [1] "Updating data views. Please be patient..."
#> 
#> 
#> 
#> $MC$errors
#> $MC$errors$fds_filter_column
#> [1] "Filter column not specified."
#> 
#> $MC$errors$fds_filter_operator
#> [1] "Filter condition not specified."
#> 
#> $MC$errors$fds_filter_rhs
#> [1] "Filter RHS not specified."
#> 
#> $MC$errors$fds_mutate_column
#> [1] "Mutate column not specified."
#> 
#> $MC$errors$fds_mutate_rhs
#> [1] "Mutate RHS not specified."
#> 
#> $MC$errors$fds_rename_column
#> [1] "Column to rename not specified."
#> 
#> $MC$errors$fds_rename_rhs
#> [1] "New column name not specified."
#> 
#> $MC$errors$fds_group_column
#> [1] "No columns to group specified."
#> 
#> $MC$errors$fds_select_column
#> [1] "You must select at least one column to keep."
#> 
#> $MC$errors$fds_longer_column
#> [1] "You must select at least one column to make longer."
#> 
#> $MC$errors$fds_longer_names
#> [1] "You must supply a column for the names of the new longer dataset"
#> 
#> $MC$errors$fds_longer_values
#> [1] "You must supply a column for the values of the new longer dataset"
#> 
#> $MC$errors$fds_wider_names
#> [1] "You must select a column with the names."
#> 
#> $MC$errors$`fds_wider_value s`
#> [1] "You must select a column with the values."
#> 
#> $MC$errors$current_key
#> [1] "Short description"
#> 
#> $MC$errors$current_key_empty
#> [1] "The data view description field cannot be empty"
#> 
#> $MC$errors$unknown_action
#> [1] "The selection action is not supported."
#> 
#> $MC$errors$no_rows
#> [1] "There are no rows in the data set."
#> 
#> $MC$errors$element_not_added
#> [1] "Unable to add the data wrangling element:"
#> 
#> $MC$errors$fix_bad_element
#> [1] "There is a bad element in the data wrangling chain,\nyou need to fix that before adding new elements"
#> 
#> 
#> $MC$operators
#> $MC$operators[[1]]
#> $MC$operators[[1]]$type
#> [1] "factor"
#> 
#> $MC$operators[[1]]$text
#> [1] "is one of"
#> 
#> $MC$operators[[1]]$rop
#> [1] "%in%"
#> 
#> 
#> $MC$operators[[2]]
#> $MC$operators[[2]]$type
#> [1] "factor"
#> 
#> $MC$operators[[2]]$text
#> [1] "is not"
#> 
#> $MC$operators[[2]]$rop
#> [1] "!%in%"
#> 
#> 
#> $MC$operators[[3]]
#> $MC$operators[[3]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[3]]$text
#> [1] "is within the range"
#> 
#> $MC$operators[[3]]$rop
#> [1] "within"
#> 
#> 
#> $MC$operators[[4]]
#> $MC$operators[[4]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[4]]$text
#> [1] "is equal to"
#> 
#> $MC$operators[[4]]$rop
#> [1] "=="
#> 
#> 
#> $MC$operators[[5]]
#> $MC$operators[[5]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[5]]$text
#> [1] "is not equal to"
#> 
#> $MC$operators[[5]]$rop
#> [1] "!="
#> 
#> 
#> 
#> $MC$actions
#> $MC$actions$filter
#> $MC$actions$filter$choice
#> [1] "Filter"
#> 
#> $MC$actions$filter$subtext
#> [1] "Filter data"
#> 
#> 
#> $MC$actions$mutate
#> $MC$actions$mutate$choice
#> [1] "Alter"
#> 
#> $MC$actions$mutate$subtext
#> [1] "Create new column or change existing column"
#> 
#> 
#> $MC$actions$rename
#> $MC$actions$rename$choice
#> [1] "Rename"
#> 
#> $MC$actions$rename$subtext
#> [1] "Rename column"
#> 
#> 
#> $MC$actions$group
#> $MC$actions$group$choice
#> [1] "Group"
#> 
#> $MC$actions$group$subtext
#> [1] "Group subsequent operations by columns"
#> 
#> 
#> $MC$actions$ungroup
#> $MC$actions$ungroup$choice
#> [1] "Ungroup"
#> 
#> $MC$actions$ungroup$subtext
#> [1] "Removing groping"
#> 
#> 
#> $MC$actions$longer
#> $MC$actions$longer$choice
#> [1] "Elongate"
#> 
#> $MC$actions$longer$subtext
#> [1] "Make Longer: rows to columns"
#> 
#> 
#> $MC$actions$wider
#> $MC$actions$wider$choice
#> [1] "Widen"
#> 
#> $MC$actions$wider$subtext
#> [1] "Make Wider: columns to rows"
#> 
#> 
#> $MC$actions$onerow
#> $MC$actions$onerow$choice
#> [1] "First Row"
#> 
#> $MC$actions$onerow$subtext
#> [1] "Keep only first row of grouped data"
#> 
#> 
#> $MC$actions$select
#> $MC$actions$select$choice
#> [1] "Select"
#> 
#> $MC$actions$select$subtext
#> [1] "Choose columns to keep"
#> 
#> 
#> 
#> $MC$formatting
#> $MC$formatting$preview
#> $MC$formatting$preview$width
#> [1] 800
#> 
#> $MC$formatting$preview$height
#> [1] 400
#> 
#> 
#> $MC$formatting$dw_elements
#> $MC$formatting$dw_elements$width
#> [1] 600
#> 
#> $MC$formatting$dw_elements$height
#> [1] 200
#> 
#> 
#> $MC$formatting$current_key
#> $MC$formatting$current_key$width
#> [1] 300
#> 
#> 
#> $MC$formatting$code
#> $MC$formatting$code$width
#> [1] 600
#> 
#> $MC$formatting$code$height
#> [1] 300
#> 
#> 
#> $MC$formatting$button_dw_del
#> $MC$formatting$button_dw_del$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_del$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_dw_save
#> $MC$formatting$button_dw_save$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_save$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_dw_copy
#> $MC$formatting$button_dw_copy$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_copy$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_dw_new
#> $MC$formatting$button_dw_new$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_new$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_dw_add_element
#> $MC$formatting$button_dw_add_element$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_add_element$block
#> [1] TRUE
#> 
#> 
#> 
#> $MC$code
#> $MC$code$readOnly
#> [1] TRUE
#> 
#> $MC$code$mode
#> [1] "r"
#> 
#> 
#> $MC$ds_object_name
#> [1] "DW_myDS"
#> 
#> 
#> $DW
#> $DW$button_counters
#> $DW$button_counters$button_dw_add_element
#> [1] 0
#> 
#> $DW$button_counters$button_dw_new
#> [1] 0
#> 
#> $DW$button_counters$button_dw_del
#> [1] 0
#> 
#> $DW$button_counters$button_dw_save
#> [1] 0
#> 
#> $DW$button_counters$button_dw_copy
#> [1] 0
#> 
#> 
#> $DW$ui_hold
#> $DW$ui_hold$hot_dw_elements
#> [1] FALSE
#> 
#> $DW$ui_hold$current_key
#> [1] FALSE
#> 
#> $DW$ui_hold$select_dw_views
#> [1] FALSE
#> 
#> 
#> $DW$ui_ids
#>  [1] "hot_dw_elements"            "button_dw_add_element"     
#>  [3] "button_dw_new"              "button_dw_del"             
#>  [5] "button_dw_copy"             "button_dw_save"            
#>  [7] "select_fds_filter_column"   "select_fds_filter_operator"
#>  [9] "fds_filter_rhs"             "select_fds_mutate_column"  
#> [11] "select_fds_mutate_rhs"      "select_fds_rename_column"  
#> [13] "fds_rename_rhs"             "current_key"               
#> [15] "select_fds_group_column"    "select_fds_select_column"  
#> [17] "select_fds_longer_column"   "select_fds_longer_values"  
#> [19] "select_fds_longer_names"    "select_fds_wider_values"   
#> [21] "select_fds_wider_names"     "select_dw_views"           
#> [23] "select_dw_element"         
#> 
#> $DW$isgood
#> [1] FALSE
#> 
#> $DW$op_choices
#> $DW$op_choices$factor
#> is one of    is not 
#>    "%in%"   "!%in%" 
#> 
#> $DW$op_choices$not_factor
#> is within the range         is equal to     is not equal to 
#>            "within"                "=="                "!=" 
#> 
#> 
#> $DW$view_cntr
#> [1] 0
#> 
#> $DW$UD
#> $DW$UD$isgood
#> [1] FALSE
#> 
#> 
#> $DW$checksum
#> [1] "2f88e16c4f21bfcb290b2a8288c05bd0"
#> 
#> 
#> $MOD_TYPE
#> [1] "DW"
#> 
#> $id
#> [1] "DW"
#> 
#> $FM_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpER8QDu/temp_libpath11ad24b1b9b53/formods/templates/formods.yaml"
#> 
#> $MOD_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpER8QDu/temp_libpath11ad24b1b9b53/formods/templates/DW.yaml"
#>