Initialize DW Module State

Creates a list of the initialized module state

Usage

DW_init_state(FM_yaml_file, MOD_yaml_file, id, id_UD, session)

Arguments

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

id

Shiny module ID

id_UD

ID string for the upload data module used to handle uploads or the name of the list element in react_state where the data set is stored.

session

Shiny session variable module (NULL)

Value

list containing an empty DW state

Examples

# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = DW_test_mksession(session=list())
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:48559d397570afc9fba8ad9b252fa75a
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DW:   dest:   file.path("config","report.docx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DW:   dest:   file.path("config","report.pptx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DW:   dest:   file.path("config","report.yaml")
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85
#> → DW: module checksum updated:a5d1d3b8a436befcfa119b5d72d3697b
#> → DW: module checksum updated:48d9e35c3f568c7a6f72914538d4ecd4
#> → DW: module checksum updated:680bde3fe602684945b93dda69abc020
#> → DW: module checksum updated:1f571c6eabb372eff6c1ef806344b2e4
#> → DW: module checksum updated:a1632ef431da47f98334a9281270a3fd
#> → DW: module checksum updated:518e56af743d086ea4b1b6070bf70472
#> → DW: module checksum updated:460fd6e3a4a46f56f17bde8ff4ef5c01
#> → DW: module checksum updated:12335ca5ced95462c2ed1df919a72d15
#> → DW: module checksum updated:0794933dac491313b32ca4bbfd996641
#> → DW: module checksum updated:75abe1a1cfc787fc68049b40c9e5749e
#> → DW: module checksum updated:070868f0125fd16b3e648dc05be84cd0
#> → DW: module checksum updated:1da012ee43c59bf8bedc7376511d179e
#> → DW: module checksum updated:6dfd06414929a762cc4f07d6d428289b
#> → DW: module checksum updated:83f7106b33c98e7650432749c16d7a36
#> → DW: module checksum updated:32652abd1996af932cbdf304cc509bd0
#> → DW: module checksum updated:dfa171056875bc474e03e0c911ace2b6
#> → DW: module checksum updated:5d4dfb7affd116c593a856034804aa04
#> → DW: module checksum updated:f47016aa2630055e8a1ae44ab79cabd2
#> → DW: module checksum updated:46cc2c9503359a39be98c3c0b889851a
#> → DW: module checksum updated:58ed3ad9e797e77eba05a2576b78e50f
#> → DW: module checksum updated:aaf6ebcd2be24f05cd9b4776ef042986
#> → DW: module checksum updated:afbbf99a8ed5f01191217b8370603a9d
#> → DW: module checksum updated:086f760681a2b47a9a1dfd572f876fcd
#> → DW: module checksum updated:8c045cb5e0fe9bd74e84d87d3c6725b5
#> → DW: module checksum updated:779a7d3c81e58ea4c8a756c2e422fee0
#> → DW: module checksum updated:104438a337e6617c4504477506e90863
#> → DW: module checksum updated:1bd57823c76715ccceab0be02ed7a79e
#> → DW: module checksum updated:8b901fc11d897dd27c774accf0eb7097
#> → DW: module checksum updated:7d55b640df1207dcdc56a10c9b8d8b87
#> → DW: module checksum updated:8d91f192ea994a133c8e2fbb7ce7c11a
#> → DW: module checksum updated:d5c81022154ffb7347d7a30d66a38f60
session = sess_res$session
input   = sess_res$input

state = DW_init_state(
   FM_yaml_file  = system.file(package = "formods",
                               "templates",
                               "formods.yaml"),
   MOD_yaml_file = system.file(package = "formods",
                               "templates",
                               "DW.yaml"),
   id              = "DW",
   id_UD           = "UD",
   session         = session)
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DW:   dest:   file.path("config","report.docx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DW:   dest:   file.path("config","report.pptx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DW:   dest:   file.path("config","report.yaml")
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85

state
#> $yaml
#> $yaml$FM
#> $yaml$FM$include
#> $yaml$FM$include$files
#> $yaml$FM$include$files[[1]]
#> $yaml$FM$include$files[[1]]$file
#> $yaml$FM$include$files[[1]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\")"
#> 
#> $yaml$FM$include$files[[1]]$file$dest
#> [1] "file.path(\"config\",\"report.docx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[2]]
#> $yaml$FM$include$files[[2]]$file
#> $yaml$FM$include$files[[2]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\")"
#> 
#> $yaml$FM$include$files[[2]]$file$dest
#> [1] "file.path(\"config\",\"report.pptx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[3]]
#> $yaml$FM$include$files[[3]]$file
#> $yaml$FM$include$files[[3]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\")"
#> 
#> $yaml$FM$include$files[[3]]$file$dest
#> [1] "file.path(\"config\",\"report.yaml\")"
#> 
#> 
#> 
#> 
#> 
#> $yaml$FM$deployed
#> [1] FALSE
#> 
#> $yaml$FM$code
#> $yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $yaml$FM$code$gen_file
#> [1] "run_analysis.R"
#> 
#> $yaml$FM$code$gen_preamble
#> [1] "# formods automated output ------------------------------------------------\n# https://formods.ubiquity.tools/\nrm(list=ls())"
#> 
#> $yaml$FM$code$packages
#> [1] "onbrand" "writexl"
#> 
#> 
#> $yaml$FM$notifications
#> $yaml$FM$notifications$config
#> $yaml$FM$notifications$config$success
#> $yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $yaml$FM$notifications$config$failure
#> $yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $yaml$FM$notifications$config$info
#> $yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $yaml$FM$notifications$config$warning
#> $yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $yaml$FM$reporting
#> $yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $yaml$FM$reporting$content_init
#> $yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.docx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> $yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.pptx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> 
#> 
#> $yaml$FM$ui
#> $yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $yaml$FM$ui$select_size
#> [1] 10
#> 
#> $yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $yaml$FM$data_meta
#> $yaml$FM$data_meta$data_header
#> [1] "<span style='color:===COLOR==='><b>===NAME===</b><br/><font size='-3'>===LABEL===</font></span>"
#> 
#> $yaml$FM$data_meta$subtext
#> [1] "===LABEL===: ===RANGE==="
#> 
#> $yaml$FM$data_meta$many_sep
#> [1] ",⋅⋅⋅,"
#> 
#> $yaml$FM$data_meta$data_types
#> $yaml$FM$data_meta$data_types$character
#> $yaml$FM$data_meta$data_types$character$color
#> [1] "#DD4B39"
#> 
#> $yaml$FM$data_meta$data_types$character$label
#> [1] "text"
#> 
#> 
#> $yaml$FM$data_meta$data_types$double
#> $yaml$FM$data_meta$data_types$double$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$double$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$integer
#> $yaml$FM$data_meta$data_types$integer$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$integer$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$other
#> $yaml$FM$data_meta$data_types$other$color
#> [1] "black"
#> 
#> $yaml$FM$data_meta$data_types$other$label
#> [1] "other"
#> 
#> 
#> 
#> 
#> $yaml$FM$labels
#> $yaml$FM$labels$default_ds
#> [1] "Original data set"
#> 
#> 
#> $yaml$FM$user_files
#> $yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $yaml$FM$logging
#> $yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $MC
#> $MC$compact
#> $MC$compact$code
#> [1] TRUE
#> 
#> $MC$compact$preview
#> [1] TRUE
#> 
#> $MC$compact$clip
#> [1] TRUE
#> 
#> 
#> $MC$reporting
#> $MC$reporting$enabled
#> [1] TRUE
#> 
#> $MC$reporting$priority
#> [1] 1
#> 
#> 
#> $MC$notifications
#> $MC$notifications$new_dw_element
#> [1] "Added element: ===ACTION=== (===DESC===)"
#> 
#> $MC$notifications$del_dw_element
#> [1] "Removed element: ===ACTION=== (===DESC===)"
#> 
#> 
#> $MC$tooltips
#> $MC$tooltips$include
#> [1] TRUE
#> 
#> $MC$tooltips$show_code
#> [1] "Show data wrangling code"
#> 
#> $MC$tooltips$dw_elements
#> [1] "Current wrangling elements"
#> 
#> 
#> $MC$labels
#> $MC$labels$tab_main
#> [1] "Wrangle"
#> 
#> $MC$labels$tab_code
#> [1] "Code"
#> 
#> $MC$labels$add_element
#> [1] "Add Operation/Action"
#> 
#> $MC$labels$fds_filter_column
#> [1] "Filter by Column"
#> 
#> $MC$labels$fds_filter_operator
#> [1] "Condition"
#> 
#> $MC$labels$fds_filter_rhs
#> [1] "Matching"
#> 
#> $MC$labels$fds_mutate_column
#> [1] "New or current column to alter"
#> 
#> $MC$labels$fds_mutate_rhs
#> [1] "New Value"
#> 
#> $MC$labels$fds_rename_column
#> [1] "Curret Column"
#> 
#> $MC$labels$fds_rename_rhs
#> [1] "New Name"
#> 
#> $MC$labels$fds_group_column
#> [1] "Columns to group"
#> 
#> $MC$labels$fds_longer_column
#> [1] "Columns to make longer"
#> 
#> $MC$labels$fds_longer_names
#> [1] "New column to contain names"
#> 
#> $MC$labels$fds_longer_values
#> [1] "New column to contain values"
#> 
#> $MC$labels$fds_wider_names
#> [1] "Column with names"
#> 
#> $MC$labels$fds_wider_values
#> [1] "Column with values"
#> 
#> $MC$labels$new_dw
#> [1] "New"
#> 
#> $MC$labels$del_dw
#> [1] "Delete"
#> 
#> $MC$labels$save_dw
#> [1] "Save"
#> 
#> $MC$labels$copy_dw
#> [1] "Copy"
#> 
#> $MC$labels$clip_dw
#> [1] "Code"
#> 
#> $MC$labels$ungroup_data
#> [1] "Removing current grouping"
#> 
#> $MC$labels$keep_onerow
#> [1] "Keep first row in current grouping"
#> 
#> $MC$labels$no_dataset
#> [1] "No dataset available to wrangle. You need to load a dataset to use this module."
#> 
#> $MC$labels$no_dw_elements
#> [1] "No data wrangling elements defined yet"
#> 
#> $MC$labels$busy
#> $MC$labels$busy$dv_update
#> [1] "Updating data views. Please be patient..."
#> 
#> 
#> 
#> $MC$errors
#> $MC$errors$fds_filter_column
#> [1] "Filter column not specified."
#> 
#> $MC$errors$fds_filter_operator
#> [1] "Filter condition not specified."
#> 
#> $MC$errors$fds_filter_rhs
#> [1] "Filter RHS not specified."
#> 
#> $MC$errors$fds_mutate_column
#> [1] "Mutate column not specified."
#> 
#> $MC$errors$fds_mutate_rhs
#> [1] "Mutate RHS not specified."
#> 
#> $MC$errors$fds_rename_column
#> [1] "Column to rename not specified."
#> 
#> $MC$errors$fds_rename_rhs
#> [1] "New column name not specified."
#> 
#> $MC$errors$fds_group_column
#> [1] "No columns to group specified."
#> 
#> $MC$errors$fds_select_column
#> [1] "You must select at least one column to keep."
#> 
#> $MC$errors$fds_longer_column
#> [1] "You must select at least one column to make longer."
#> 
#> $MC$errors$fds_longer_names
#> [1] "You must supply a column for the names of the new longer dataset"
#> 
#> $MC$errors$fds_longer_values
#> [1] "You must supply a column for the values of the new longer dataset"
#> 
#> $MC$errors$fds_wider_names
#> [1] "You must select a column with the names."
#> 
#> $MC$errors$`fds_wider_value s`
#> [1] "You must select a column with the values."
#> 
#> $MC$errors$current_key
#> [1] "Short description"
#> 
#> $MC$errors$current_key_empty
#> [1] "The data view description field cannot be empty"
#> 
#> $MC$errors$unknown_action
#> [1] "The selection action is not supported."
#> 
#> $MC$errors$no_rows
#> [1] "There are no rows in the data set."
#> 
#> $MC$errors$element_not_added
#> [1] "Unable to add the data wrangling element:"
#> 
#> $MC$errors$fix_bad_element
#> [1] "There is a bad element in the data wrangling chain,\nyou need to fix that before adding new elements"
#> 
#> 
#> $MC$operators
#> $MC$operators[[1]]
#> $MC$operators[[1]]$type
#> [1] "factor"
#> 
#> $MC$operators[[1]]$text
#> [1] "is one of"
#> 
#> $MC$operators[[1]]$rop
#> [1] "%in%"
#> 
#> 
#> $MC$operators[[2]]
#> $MC$operators[[2]]$type
#> [1] "factor"
#> 
#> $MC$operators[[2]]$text
#> [1] "is not"
#> 
#> $MC$operators[[2]]$rop
#> [1] "!%in%"
#> 
#> 
#> $MC$operators[[3]]
#> $MC$operators[[3]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[3]]$text
#> [1] "is within the range"
#> 
#> $MC$operators[[3]]$rop
#> [1] "within"
#> 
#> 
#> $MC$operators[[4]]
#> $MC$operators[[4]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[4]]$text
#> [1] "is equal to"
#> 
#> $MC$operators[[4]]$rop
#> [1] "=="
#> 
#> 
#> $MC$operators[[5]]
#> $MC$operators[[5]]$type
#> [1] "not_factor"
#> 
#> $MC$operators[[5]]$text
#> [1] "is not equal to"
#> 
#> $MC$operators[[5]]$rop
#> [1] "!="
#> 
#> 
#> 
#> $MC$actions
#> $MC$actions$filter
#> $MC$actions$filter$choice
#> [1] "Filter"
#> 
#> $MC$actions$filter$subtext
#> [1] "Filter data"
#> 
#> 
#> $MC$actions$mutate
#> $MC$actions$mutate$choice
#> [1] "Alter"
#> 
#> $MC$actions$mutate$subtext
#> [1] "Create new column or change existing column"
#> 
#> 
#> $MC$actions$rename
#> $MC$actions$rename$choice
#> [1] "Rename"
#> 
#> $MC$actions$rename$subtext
#> [1] "Rename column"
#> 
#> 
#> $MC$actions$group
#> $MC$actions$group$choice
#> [1] "Group"
#> 
#> $MC$actions$group$subtext
#> [1] "Group subsequent operations by columns"
#> 
#> 
#> $MC$actions$ungroup
#> $MC$actions$ungroup$choice
#> [1] "Ungroup"
#> 
#> $MC$actions$ungroup$subtext
#> [1] "Removing groping"
#> 
#> 
#> $MC$actions$longer
#> $MC$actions$longer$choice
#> [1] "Elongate"
#> 
#> $MC$actions$longer$subtext
#> [1] "Make Longer: rows to columns"
#> 
#> 
#> $MC$actions$wider
#> $MC$actions$wider$choice
#> [1] "Widen"
#> 
#> $MC$actions$wider$subtext
#> [1] "Make Wider: columns to rows"
#> 
#> 
#> $MC$actions$onerow
#> $MC$actions$onerow$choice
#> [1] "First Row"
#> 
#> $MC$actions$onerow$subtext
#> [1] "Keep only first row of grouped data"
#> 
#> 
#> $MC$actions$select
#> $MC$actions$select$choice
#> [1] "Select"
#> 
#> $MC$actions$select$subtext
#> [1] "Choose columns to keep"
#> 
#> 
#> 
#> $MC$formatting
#> $MC$formatting$preview
#> $MC$formatting$preview$width
#> [1] 800
#> 
#> $MC$formatting$preview$height
#> [1] 400
#> 
#> 
#> $MC$formatting$dw_elements
#> $MC$formatting$dw_elements$width
#> [1] 600
#> 
#> $MC$formatting$dw_elements$height
#> [1] 200
#> 
#> 
#> $MC$formatting$current_key
#> $MC$formatting$current_key$width
#> [1] 300
#> 
#> 
#> $MC$formatting$code
#> $MC$formatting$code$width
#> [1] 800
#> 
#> $MC$formatting$code$height
#> [1] 300
#> 
#> 
#> $MC$formatting$button_dw_del
#> $MC$formatting$button_dw_del$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_del$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_del$tooltip
#> [1] "Delete the current data view."
#> 
#> $MC$formatting$button_dw_del$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_clip
#> $MC$formatting$button_dw_clip$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_clip$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_clip$tooltip
#> [1] "Copy code to generate current data view to the clipboard."
#> 
#> $MC$formatting$button_dw_clip$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_save
#> $MC$formatting$button_dw_save$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_save$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_save$tooltip
#> [1] "Save changes to data view name"
#> 
#> $MC$formatting$button_dw_save$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_copy
#> $MC$formatting$button_dw_copy$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_copy$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_copy$tooltip
#> [1] "Copy current data view to a new data view."
#> 
#> $MC$formatting$button_dw_copy$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_new
#> $MC$formatting$button_dw_new$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_new$block
#> [1] TRUE
#> 
#> $MC$formatting$button_dw_new$tooltip
#> [1] "Create a new empty data view."
#> 
#> $MC$formatting$button_dw_new$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_dw_add_element
#> $MC$formatting$button_dw_add_element$size
#> [1] "sm"
#> 
#> $MC$formatting$button_dw_add_element$block
#> [1] TRUE
#> 
#> 
#> 
#> $MC$code
#> $MC$code$readOnly
#> [1] TRUE
#> 
#> $MC$code$mode
#> [1] "r"
#> 
#> $MC$code$packages
#> [1] "dplyr" "tidyr"
#> 
#> $MC$code$preamble
#> [1] "# Data wrangling ----------------------------------------------------------"
#> 
#> 
#> $MC$ds_object_name
#> [1] "DW_myDS"
#> 
#> 
#> $DW
#> $DW$button_counters
#> $DW$button_counters$button_dw_add_element
#> [1] 0
#> 
#> $DW$button_counters$button_dw_new
#> [1] 0
#> 
#> $DW$button_counters$button_dw_del
#> [1] 0
#> 
#> $DW$button_counters$button_dw_save
#> [1] 0
#> 
#> $DW$button_counters$button_dw_copy
#> [1] 0
#> 
#> 
#> $DW$ui_hold
#> $DW$ui_hold$hot_dw_elements
#> [1] FALSE
#> 
#> $DW$ui_hold$current_key
#> [1] FALSE
#> 
#> $DW$ui_hold$select_dw_views
#> [1] FALSE
#> 
#> 
#> $DW$ui_ids
#>  [1] "hot_dw_elements"            "button_dw_add_element"     
#>  [3] "button_dw_new"              "button_dw_del"             
#>  [5] "button_dw_copy"             "button_dw_save"            
#>  [7] "select_fds_filter_column"   "select_fds_filter_operator"
#>  [9] "fds_filter_rhs"             "select_fds_mutate_column"  
#> [11] "select_fds_mutate_rhs"      "select_fds_rename_column"  
#> [13] "fds_rename_rhs"             "current_key"               
#> [15] "select_fds_group_column"    "select_fds_select_column"  
#> [17] "select_fds_longer_column"   "select_fds_longer_values"  
#> [19] "select_fds_longer_names"    "select_fds_wider_values"   
#> [21] "select_fds_wider_names"     "select_dw_views"           
#> [23] "select_dw_element"         
#> 
#> $DW$isgood
#> [1] TRUE
#> 
#> $DW$op_choices
#> $DW$op_choices$factor
#> is one of    is not 
#>    "%in%"   "!%in%" 
#> 
#> $DW$op_choices$not_factor
#> is within the range         is equal to     is not equal to 
#>            "within"                "=="                "!=" 
#> 
#> 
#> $DW$view_cntr
#> [1] 1
#> 
#> $DW$UD
#> $DW$UD$isgood
#> [1] TRUE
#> 
#> $DW$UD$checksum
#> [1] "48559d397570afc9fba8ad9b252fa75a"
#> 
#> $DW$UD$contents
#> # A tibble: 19,080 × 33
#>       ID TIME_DY TIME_HR NTIME…¹ NTIME…²  TIME   AMT    DV RECDESC  DOSE DOSE_…³
#>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl> <dbl> <chr>   <dbl> <chr>  
#>  1     1  0         0     0         0     0      3e6    0  dose        3 3 mg   
#>  2     1  0         0     0         0     0      0   1250  BM_ng_…     3 3 mg   
#>  3     1  0         0     0         0     0      0      0  C_ng_ml     3 3 mg   
#>  4     1  0.0104    0.25  0.0104    0.25  0.25   0   1255  BM_ng_…     3 3 mg   
#>  5     1  0.0104    0.25  0.0104    0.25  0.25   0    816  C_ng_ml     3 3 mg   
#>  6     1  0.0208    0.5   0.0208    0.5   0.5    0   1259  BM_ng_…     3 3 mg   
#>  7     1  0.0208    0.5   0.0208    0.5   0.5    0    814. C_ng_ml     3 3 mg   
#>  8     1  0.0417    1     0.0417    1     1      0   1269  BM_ng_…     3 3 mg   
#>  9     1  0.0417    1     0.0417    1     1      0    812. C_ng_ml     3 3 mg   
#> 10     1  0.0833    2     0.0833    2     2      0   1287  BM_ng_…     3 3 mg   
#> # … with 19,070 more rows, 22 more variables: Cohort <chr>, ROUTE <chr>,
#> #   WT <dbl>, SUBTYPE <chr>, SEX <chr>, F1 <dbl>, ka <dbl>, CL <dbl>, Vc <dbl>,
#> #   Vp <dbl>, Q <dbl>, BM_IC <dbl>, kdeg_BM <dbl>, Emax <dbl>, EC50 <dbl>,
#> #   CMT <chr>, DUR <dbl>, EVID <dbl>, NDose <chr>, DOSE_NUM <dbl>,
#> #   SUBTYPE_ID <dbl>, SEX_ID <dbl>, and abbreviated variable names ¹​NTIME_DY,
#> #   ²​NTIME_HR, ³​DOSE_STR
#> 
#> $DW$UD$object_name
#> [1] "myDS"
#> 
#> 
#> $DW$code_previous
#> [1] "myDS = readxl::read_excel(path=\"TEST_DATA.xlsx\", sheet=\"DATA\")\nmyDS = janitor::clean_names(myDS, case=\"none\")"
#> 
#> $DW$views
#> $DW$views$view_1
#> $DW$views$view_1$isgood
#> [1] TRUE
#> 
#> $DW$views$view_1$id
#> [1] "view_1"
#> 
#> $DW$views$view_1$idx
#> [1] 1
#> 
#> $DW$views$view_1$view_ds_object_name
#> [1] "DW_myDS_1"
#> 
#> $DW$views$view_1$code_previous
#> [1] "DW_myDS_1 = myDS"
#> 
#> $DW$views$view_1$key
#> [1] "data_view_1"
#> 
#> $DW$views$view_1$WDS
#> # A tibble: 19,080 × 33
#>       ID TIME_DY TIME_HR NTIME…¹ NTIME…²  TIME   AMT    DV RECDESC  DOSE DOSE_…³
#>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl> <dbl> <chr>   <dbl> <chr>  
#>  1     1  0         0     0         0     0      3e6    0  dose        3 3 mg   
#>  2     1  0         0     0         0     0      0   1250  BM_ng_…     3 3 mg   
#>  3     1  0         0     0         0     0      0      0  C_ng_ml     3 3 mg   
#>  4     1  0.0104    0.25  0.0104    0.25  0.25   0   1255  BM_ng_…     3 3 mg   
#>  5     1  0.0104    0.25  0.0104    0.25  0.25   0    816  C_ng_ml     3 3 mg   
#>  6     1  0.0208    0.5   0.0208    0.5   0.5    0   1259  BM_ng_…     3 3 mg   
#>  7     1  0.0208    0.5   0.0208    0.5   0.5    0    814. C_ng_ml     3 3 mg   
#>  8     1  0.0417    1     0.0417    1     1      0   1269  BM_ng_…     3 3 mg   
#>  9     1  0.0417    1     0.0417    1     1      0    812. C_ng_ml     3 3 mg   
#> 10     1  0.0833    2     0.0833    2     2      0   1287  BM_ng_…     3 3 mg   
#> # … with 19,070 more rows, 22 more variables: Cohort <chr>, ROUTE <chr>,
#> #   WT <dbl>, SUBTYPE <chr>, SEX <chr>, F1 <dbl>, ka <dbl>, CL <dbl>, Vc <dbl>,
#> #   Vp <dbl>, Q <dbl>, BM_IC <dbl>, kdeg_BM <dbl>, Emax <dbl>, EC50 <dbl>,
#> #   CMT <chr>, DUR <dbl>, EVID <dbl>, NDose <chr>, DOSE_NUM <dbl>,
#> #   SUBTYPE_ID <dbl>, SEX_ID <dbl>, and abbreviated variable names ¹​NTIME_DY,
#> #   ²​NTIME_HR, ³​DOSE_STR
#> 
#> $DW$views$view_1$elements_table
#> NULL
#> 
#> $DW$views$view_1$checksum
#> [1] "48559d397570afc9fba8ad9b252fa75a"
#> 
#> $DW$views$view_1$code
#> NULL
#> 
#> $DW$views$view_1$code_dw_only
#> NULL
#> 
#> 
#> 
#> $DW$current_view
#> [1] "view_1"
#> 
#> $DW$checksum
#> [1] "5a6adb8be080914162e848e04fb61f85"
#> 
#> 
#> $MOD_TYPE
#> [1] "DW"
#> 
#> $id
#> [1] "DW"
#> 
#> $dep_mod_ids
#> [1] "UD"
#> 
#> $FM_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/Rtmpu9dCGc/temp_libpath681843653169/formods/templates/formods.yaml"
#> 
#> $MOD_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/Rtmpu9dCGc/temp_libpath681843653169/formods/templates/DW.yaml"
#>