Evaluates Figure Generation Code

Takes the current state and rebuilds the active figure. If the elements table has a row flagged for deletion, it will be deleted. If the cmd input is not NULL it will attempt to append that element to the figure.

Usage

FG_build(
  state,
  del_row = NULL,
  cmd = NULL,
  element = "unknown",
  desc = "unknown"
)

Arguments

state

FG state from FG_fetch_state()

del_row

Row number to be deleted (NULL if no rows need to be deleted)

cmd

String containing the plotting command. Set to NULL to initialize a new figure or force a rebuild after a dataset update.

element

Short name for the figure element being performed, eg. point

desc

Verbose description for the action being performed

Value

list with the following elements

isgood: Return status of the function.

msgs: Messages to be passed back to the user.

pages: List with each element containing a ggplot object (p) and the code to generate that object (code)

Examples

library(formods)
# Within shiny both session and input variables will exist, 
# this creates examples here for testing purposes:
sess_res = FG_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:631af84e28478adc1628dd2c9fb22662
#> → DW: State initialized
#> → DW: module checksum updated:f5866ed2a560dbc564dec9145314b519
#> → DW: module checksum updated:4911af9c225aac6361acea8f26bbfac9
#> → DW: module checksum updated:c9b0caeab52fb59e0dcbb88ee042729a
#> → DW: module checksum updated:b7f3e08819cabbd242f5bebe1290714b
#> → DW: module checksum updated:8d8944be2daf4ea8aa3b1c1c707cee45
#> → DW: module checksum updated:a6367d7e4e002eca11050774144105f8
#> → DW: module checksum updated:4a381cdf6502fe13dca9c009a25b4418
#> → DW: module checksum updated:1050496cc0eca17d51102e051220ec49
#> → DW: module checksum updated:df684a1d4fbbd8bd33749607f9b982c0
#> → DW: module checksum updated:e3054989800f1e8ca24386371b52accc
#> → DW: module checksum updated:3b43dbda29ded3a3a3ef514ccf0c0578
#> → DW: module checksum updated:550d10d9e424d549878ab4ba5343896e
#> → DW: module checksum updated:5d2a75ed77a22f4ff8cc2a0d2bf2e3e5
#> → DW: module checksum updated:0272b5040b80ff35adf3500fdb60d8c7
#> → DW: module checksum updated:173ecf7d96c0bbb0962c7846df0d4bca
#> → FG: State initialized
#> → FG: module checksum updated:91329389512102f6fab3f9986c1e5703
#> → FG: module checksum updated:650ec75853f0bdfe72fb3bb495cf27e7
#> → FG: module checksum updated:3642eec74d244876113148f1b95ab9a4
#> → FG: module checksum updated:2e78d97154fe0e788bdfd2ab8fdbde9b
#> → FG: module checksum updated:eb5c510f4628e2444c2119ac29e2b231
#> → FG: module checksum updated:3ec2b07f9e289a008ab2571f893de2bd
#> → FG: module checksum updated:a622eb068af64c1556badb1646072727
#> → FG: module checksum updated:7d4cd1c496f8d80cea0493c33561ac02
#> → FG: module checksum updated:3b116eaf31ed1cb097b4941153fa20f3
#> → FG: module checksum updated:0c3ef8f3a59595879587f3c2d06ec8cc
#> → FG: module checksum updated:a00934fb1419efda8ed646c2925898cb
#> → FG: module checksum updated:1306c9a57ff37875dbc4193181a9739e
#> → FG: module checksum updated:d96e85760b1d1c65a7a58f8d6edb4d08
session = sess_res$session
input   = sess_res$input

# This will create a populated FG state object:
state   = sess_res$state

# This sets the current active figure to Fig_1
state[["FG"]][["current_fig"]]  =  "Fig_1"

# This is a paginated figure, and we can access a specific
# figure using the following:
pg_1 = FG_extract_page(state, 1)
pg_2 = FG_extract_page(state, 2)

# This will give you access to the current figure directly:
current_fig = FG_fetch_current_fig(state)

# For example this will set the key for that figure:
current_fig$key = "Individual profiles by cohort (multiple pages)"

# Once you're done you can put it back into the state:
state = FG_set_current_fig(state, current_fig)

# If you made any changes to the actual figure, this will 
# force a rebuild of the current figure:
state = FG_build( state=state, del_row = NULL, cmd = NULL)
#> → FG: module checksum updated:5d82401f6428b8b2cb4e1d17d3e5e6ff

# To create a new empty figure you can do this:
state = FG_new_fig(state)
#> → FG: module checksum updated:061ee359c4d7cf6a36712262294622f1