Evaluates Figure Generation Code

Takes the current state and rebuilds the active figure. If the elements table has a row flagged for deletion, it will be deleted. If the cmd input is not NULL it will attempt to append that element to the figure.

Usage

FG_build(
  state,
  del_row = NULL,
  cmd = NULL,
  element = "unknown",
  desc = "unknown"
)

Arguments

state

FG state from FG_fetch_state()

del_row

Row number to be deleted (NULL if no rows need to be deleted)

cmd

String containing the plotting command. Set to NULL to initialize a new figure or force a rebuild after a dataset update.

element

Short name for the figure element being performed, eg. point

desc

Verbose description for the action being performed

Value

list with the following elements

isgood: Return status of the function.

msgs: Messages to be passed back to the user.

pages: List with each element containing a ggplot object (p) and the code to generate that object (code)

Examples

library(formods)
# Within shiny both session and input variables will exist, 
# this creates examples here for testing purposes:
sess_res = FG_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:48559d397570afc9fba8ad9b252fa75a
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85
#> → DW: module checksum updated:a5d1d3b8a436befcfa119b5d72d3697b
#> → DW: module checksum updated:48d9e35c3f568c7a6f72914538d4ecd4
#> → DW: module checksum updated:680bde3fe602684945b93dda69abc020
#> → DW: module checksum updated:1f571c6eabb372eff6c1ef806344b2e4
#> → DW: module checksum updated:a1632ef431da47f98334a9281270a3fd
#> → DW: module checksum updated:518e56af743d086ea4b1b6070bf70472
#> → DW: module checksum updated:460fd6e3a4a46f56f17bde8ff4ef5c01
#> → DW: module checksum updated:12335ca5ced95462c2ed1df919a72d15
#> → DW: module checksum updated:0794933dac491313b32ca4bbfd996641
#> → DW: module checksum updated:75abe1a1cfc787fc68049b40c9e5749e
#> → DW: module checksum updated:070868f0125fd16b3e648dc05be84cd0
#> → DW: module checksum updated:1da012ee43c59bf8bedc7376511d179e
#> → DW: module checksum updated:6dfd06414929a762cc4f07d6d428289b
#> → DW: module checksum updated:83f7106b33c98e7650432749c16d7a36
#> → DW: module checksum updated:32652abd1996af932cbdf304cc509bd0
#> → DW: module checksum updated:dfa171056875bc474e03e0c911ace2b6
#> → DW: module checksum updated:5d4dfb7affd116c593a856034804aa04
#> → DW: module checksum updated:f47016aa2630055e8a1ae44ab79cabd2
#> → DW: module checksum updated:46cc2c9503359a39be98c3c0b889851a
#> → DW: module checksum updated:58ed3ad9e797e77eba05a2576b78e50f
#> → DW: module checksum updated:aaf6ebcd2be24f05cd9b4776ef042986
#> → DW: module checksum updated:afbbf99a8ed5f01191217b8370603a9d
#> → DW: module checksum updated:086f760681a2b47a9a1dfd572f876fcd
#> → DW: module checksum updated:8c045cb5e0fe9bd74e84d87d3c6725b5
#> → DW: module checksum updated:779a7d3c81e58ea4c8a756c2e422fee0
#> → DW: module checksum updated:104438a337e6617c4504477506e90863
#> → DW: module checksum updated:1bd57823c76715ccceab0be02ed7a79e
#> → DW: module checksum updated:8b901fc11d897dd27c774accf0eb7097
#> → DW: module checksum updated:7d55b640df1207dcdc56a10c9b8d8b87
#> → DW: module checksum updated:8d91f192ea994a133c8e2fbb7ce7c11a
#> → DW: module checksum updated:d5c81022154ffb7347d7a30d66a38f60
#> → FG: State initialized
#> → FG: module checksum updated:91329389512102f6fab3f9986c1e5703
#> → FG: module checksum updated:d079a8d1f3893c64d341c4a91571281d
#> → FG: module checksum updated:8e6f14d226d3fac229aca43a77024339
#> → FG: module checksum updated:fd0a4dcd2a9af7e220b777f621870016
#> → FG: module checksum updated:98d5951ad7f09afa3b022157d1eab303
#> → FG: module checksum updated:83ffca8a8bf4b2a51cc6e6f7e087928b
#> → FG: module checksum updated:05ab337785bd096c5d2f296bc3ac50d3
#> → FG: module checksum updated:43e8133b6728327b9ae90b63f8eb5232
#> → FG: module checksum updated:db83fe32cf0a0ba3f84c48eaaeede2f8
#> → FG: module checksum updated:69e4c1ab50ffc66a9d80ddf613fff42f
#> → FG: module checksum updated:78c49daac15d177905b0ee0a331d0a89
#> → FG: module checksum updated:6d9bcdb48081f12578a99c685ecc6c1c
#> → FG: module checksum updated:b52e043be81f00d8597107924f1caa0d
session = sess_res$session
input   = sess_res$input

# This will create a populated FG state object:
state   = sess_res$state

# This sets the current active figure to Fig_1
state[["FG"]][["current_fig"]]  =  "Fig_1"

# This is a paginated figure, and we can access a specific
# figure using the following:
pg_1 = FG_extract_page(state, 1)
pg_2 = FG_extract_page(state, 2)

# This will give you access to the current figure directly:
current_fig = FG_fetch_current_fig(state)

# For example this will set the key for that figure:
current_fig$key = "Individual profiles by cohort (multiple pages)"

# Once you're done you can put it back into the state:
state = FG_set_current_fig(state, current_fig)

# If you made any changes to the actual figure, this will 
# force a rebuild of the current figure:
state = FG_build( state=state, del_row = NULL, cmd = NULL)
#> → FG: module checksum updated:28cb9363189f4a501854e499d1735376

# To create a new empty figure you can do this:
state = FG_new_fig(state)
#> → FG: module checksum updated:c9afe4089eac9ce34fae02047ed90bcd