Initialize New Figure

Creates a new figure in a FG module

Usage

FG_new_fig(state)

Arguments

state

FG state from FG_fetch_state()

Value

FG state object containing a new empty figure and that figure set as the current active figure

Examples

library(formods)
# Within shiny both session and input variables will exist, 
# this creates examples here for testing purposes:
sess_res = FG_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:631af84e28478adc1628dd2c9fb22662
#> → DW: State initialized
#> → DW: module checksum updated:f5866ed2a560dbc564dec9145314b519
#> → DW: module checksum updated:4911af9c225aac6361acea8f26bbfac9
#> → DW: module checksum updated:c9b0caeab52fb59e0dcbb88ee042729a
#> → DW: module checksum updated:b7f3e08819cabbd242f5bebe1290714b
#> → DW: module checksum updated:8d8944be2daf4ea8aa3b1c1c707cee45
#> → DW: module checksum updated:a6367d7e4e002eca11050774144105f8
#> → DW: module checksum updated:4a381cdf6502fe13dca9c009a25b4418
#> → DW: module checksum updated:1050496cc0eca17d51102e051220ec49
#> → DW: module checksum updated:df684a1d4fbbd8bd33749607f9b982c0
#> → DW: module checksum updated:e3054989800f1e8ca24386371b52accc
#> → DW: module checksum updated:3b43dbda29ded3a3a3ef514ccf0c0578
#> → DW: module checksum updated:550d10d9e424d549878ab4ba5343896e
#> → DW: module checksum updated:5d2a75ed77a22f4ff8cc2a0d2bf2e3e5
#> → DW: module checksum updated:0272b5040b80ff35adf3500fdb60d8c7
#> → DW: module checksum updated:173ecf7d96c0bbb0962c7846df0d4bca
#> → FG: State initialized
#> → FG: module checksum updated:91329389512102f6fab3f9986c1e5703
#> → FG: module checksum updated:a8e26fb292c21566f0343379777d6fca
#> → FG: module checksum updated:e419082193bfa9a394ebbcdddbfd336b
#> → FG: module checksum updated:b28b0e93f96a30f0d165fdef138870f2
#> → FG: module checksum updated:f21f03041100b957a58794bcd87d6bdf
#> → FG: module checksum updated:48b86c40c94cc9e33a1776f13e211d9f
#> → FG: module checksum updated:a553c9a12860d27ac014f65e6d4d0652
#> → FG: module checksum updated:0b84fb2e15f14fb0e779755dc2de8f5e
#> → FG: module checksum updated:c652df639b6bd9590a4d21ba4cecf2ac
#> → FG: module checksum updated:0325e830201dea017dc9c954e779eab4
#> → FG: module checksum updated:a8b2e9e684108340fed6334e057938b7
#> → FG: module checksum updated:13b39577e3358040cd6b750827b91826
#> → FG: module checksum updated:e79cd97430680da53621f0ebbe037bd3
session = sess_res$session
input   = sess_res$input

# This will create a populated FG state object:
state   = sess_res$state

# This sets the current active figure to Fig_1
state[["FG"]][["current_fig"]]  =  "Fig_1"

# This is a paginated figure, and we can access a specific
# figure using the following:
pg_1 = FG_extract_page(state, 1)
pg_2 = FG_extract_page(state, 2)

# This will give you access to the current figure directly:
current_fig = FG_fetch_current_fig(state)

# For example this will set the key for that figure:
current_fig$key = "Individual profiles by cohort (multiple pages)"

# Once you're done you can put it back into the state:
state = FG_set_current_fig(state, current_fig)

# If you made any changes to the actual figure, this will 
# force a rebuild of the current figure:
state = FG_build( state=state, del_row = NULL, cmd = NULL)
#> → FG: module checksum updated:63dfb0cf174887e12aeec9d15ec46cfe

# To create a new empty figure you can do this:
state = FG_new_fig(state)
#> → FG: module checksum updated:0e47de304c79112f8b2ea3812398602e