Fetches Informaiton About the App

Returns diagnostic information about the app

Usage

FM_fetch_app_info(session)

Arguments

session

Shiny session variable.

Value

List with information about the app with the following structure

uiele: All system information as UI elements to be used in shiny apps.

uiele_packages: UI element for installed packages to be used in shiny apps.

uiele_options: UI element for current options.

uiele_modules: UI element for loaded formods modules to be used in shiny apps.

msgs: System information as text to be used in a report/terminal.

si_packages Dataframe with currently used packages.

si_options Dataframe with current options

Examples

# We need a Shiny session object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
session = sess_res$session
app_info  = FM_fetch_app_info(session)
app_info$msgs
#>   [1] "Modules"                                                                                                                                 
#>   [2] "ID: UD"                                                                                                                                  
#>   [3] "type: UD"                                                                                                                                
#>   [4] "FM_yaml_file: /private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpoDixuG/temp_libpath689117e8dee/formods/templates/formods.yaml"
#>   [5] "MOD_yaml_file: /private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpoDixuG/temp_libpath689117e8dee/formods/templates/UD.yaml"    
#>   [6] "User files: /var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T//RtmpfQZzxD/non_shiny/FM"                                                   
#>   [7] "Log file: /var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T//RtmpfQZzxD/non_shiny/FM/formods_log.txt"                                     
#>   [8] "Package dependencies: janitor, readr, readxl"                                                                                            
#>   [9] " package           * version  date (UTC) lib source"                                                                                     
#>  [10] " abind               1.4-5    2016-07-21 [2] CRAN (R 4.2.0)"                                                                             
#>  [11] " askpass             1.1      2019-01-13 [2] CRAN (R 4.2.0)"                                                                             
#>  [12] " backports           1.4.1    2021-12-13 [2] CRAN (R 4.2.0)"                                                                             
#>  [13] " broom               1.0.4    2023-03-11 [2] CRAN (R 4.2.0)"                                                                             
#>  [14] " bslib               0.4.2    2022-12-16 [2] CRAN (R 4.2.0)"                                                                             
#>  [15] " cachem              1.0.7    2023-02-24 [2] CRAN (R 4.2.0)"                                                                             
#>  [16] " callr               3.7.3    2022-11-02 [2] CRAN (R 4.2.0)"                                                                             
#>  [17] " car                 3.1-2    2023-03-30 [2] CRAN (R 4.2.1)"                                                                             
#>  [18] " carData             3.0-5    2022-01-06 [2] CRAN (R 4.2.0)"                                                                             
#>  [19] " cellranger          1.1.0    2016-07-27 [2] CRAN (R 4.2.0)"                                                                             
#>  [20] " cli                 3.6.1    2023-03-23 [2] CRAN (R 4.2.0)"                                                                             
#>  [21] " clipr               0.8.0    2022-02-22 [2] CRAN (R 4.2.0)"                                                                             
#>  [22] " colorspace          2.1-0    2023-01-23 [2] CRAN (R 4.2.0)"                                                                             
#>  [23] " crayon              1.5.2    2022-09-29 [2] CRAN (R 4.2.0)"                                                                             
#>  [24] " crul                1.3      2022-09-03 [2] CRAN (R 4.2.0)"                                                                             
#>  [25] " curl                5.0.0    2023-01-12 [2] CRAN (R 4.2.0)"                                                                             
#>  [26] " data.table          1.14.8   2023-02-17 [2] CRAN (R 4.2.0)"                                                                             
#>  [27] " desc                1.4.2    2022-09-08 [2] CRAN (R 4.2.0)"                                                                             
#>  [28] " devtools            2.4.5    2022-10-11 [2] CRAN (R 4.2.0)"                                                                             
#>  [29] " digest              0.6.31   2022-12-11 [2] CRAN (R 4.2.0)"                                                                             
#>  [30] " downlit             0.4.2    2022-07-05 [2] CRAN (R 4.2.0)"                                                                             
#>  [31] " dplyr               1.1.1    2023-03-22 [2] CRAN (R 4.2.0)"                                                                             
#>  [32] " ellipsis            0.3.2    2021-04-29 [2] CRAN (R 4.2.0)"                                                                             
#>  [33] " evaluate            0.20     2023-01-17 [2] CRAN (R 4.2.0)"                                                                             
#>  [34] " fansi               1.0.4    2023-01-22 [2] CRAN (R 4.2.0)"                                                                             
#>  [35] " farver              2.1.1    2022-07-06 [2] CRAN (R 4.2.0)"                                                                             
#>  [36] " fastmap             1.1.1    2023-02-24 [2] CRAN (R 4.2.0)"                                                                             
#>  [37] " flextable           0.9.1    2023-04-02 [2] CRAN (R 4.2.1)"                                                                             
#>  [38] " fontBitstreamVera   0.1.1    2017-02-01 [2] CRAN (R 4.2.0)"                                                                             
#>  [39] " fontLiberation      0.1.0    2016-10-15 [2] CRAN (R 4.2.0)"                                                                             
#>  [40] " fontquiver          0.2.1    2017-02-01 [2] CRAN (R 4.2.0)"                                                                             
#>  [41] " formods           * 0.1.3    2023-12-03 [1] local"                                                                                      
#>  [42] " fs                  1.6.1    2023-02-06 [2] CRAN (R 4.2.0)"                                                                             
#>  [43] " gdtools             0.3.3    2023-03-27 [2] CRAN (R 4.2.1)"                                                                             
#>  [44] " generics            0.1.3    2022-07-05 [2] CRAN (R 4.2.0)"                                                                             
#>  [45] " gfonts              0.2.0    2023-01-08 [2] CRAN (R 4.2.0)"                                                                             
#>  [46] " ggforce             0.4.1    2022-10-04 [2] CRAN (R 4.2.0)"                                                                             
#>  [47] " ggplot2             3.4.2    2023-04-03 [2] CRAN (R 4.2.0)"                                                                             
#>  [48] " ggpubr              0.6.0    2023-02-10 [2] CRAN (R 4.2.0)"                                                                             
#>  [49] " ggsignif            0.6.4    2022-10-13 [2] CRAN (R 4.2.0)"                                                                             
#>  [50] " glue                1.6.2    2022-02-24 [2] CRAN (R 4.2.0)"                                                                             
#>  [51] " gtable              0.3.3    2023-03-21 [2] CRAN (R 4.2.0)"                                                                             
#>  [52] " hms                 1.1.3    2023-03-21 [2] CRAN (R 4.2.0)"                                                                             
#>  [53] " htmltools           0.5.5    2023-03-23 [2] CRAN (R 4.2.0)"                                                                             
#>  [54] " htmlwidgets         1.6.2    2023-03-17 [2] CRAN (R 4.2.0)"                                                                             
#>  [55] " httpcode            0.3.0    2020-04-10 [2] CRAN (R 4.2.0)"                                                                             
#>  [56] " httpuv              1.6.9    2023-02-14 [2] CRAN (R 4.2.0)"                                                                             
#>  [57] " httr                1.4.5    2023-02-24 [2] CRAN (R 4.2.0)"                                                                             
#>  [58] " janitor             2.2.0    2023-02-02 [2] CRAN (R 4.2.0)"                                                                             
#>  [59] " jquerylib           0.1.4    2021-04-26 [2] CRAN (R 4.2.0)"                                                                             
#>  [60] " jsonlite            1.8.4    2022-12-06 [2] CRAN (R 4.2.0)"                                                                             
#>  [61] " knitr               1.42     2023-01-25 [2] CRAN (R 4.2.0)"                                                                             
#>  [62] " labeling            0.4.2    2020-10-20 [2] CRAN (R 4.2.0)"                                                                             
#>  [63] " later               1.3.0    2021-08-18 [2] CRAN (R 4.2.0)"                                                                             
#>  [64] " lazyeval            0.2.2    2019-03-15 [2] CRAN (R 4.2.0)"                                                                             
#>  [65] " lifecycle           1.0.3    2022-10-07 [2] CRAN (R 4.2.0)"                                                                             
#>  [66] " lubridate           1.9.2    2023-02-10 [2] CRAN (R 4.2.0)"                                                                             
#>  [67] " magrittr            2.0.3    2022-03-30 [2] CRAN (R 4.2.0)"                                                                             
#>  [68] " MASS                7.3-58.2 2023-01-23 [2] CRAN (R 4.2.0)"                                                                             
#>  [69] " memoise             2.0.1    2021-11-26 [2] CRAN (R 4.2.0)"                                                                             
#>  [70] " mime                0.12     2021-09-28 [2] CRAN (R 4.2.0)"                                                                             
#>  [71] " miniUI              0.1.1.1  2018-05-18 [2] CRAN (R 4.2.0)"                                                                             
#>  [72] " munsell             0.5.0    2018-06-12 [2] CRAN (R 4.2.0)"                                                                             
#>  [73] " officer             0.6.2    2023-03-28 [2] CRAN (R 4.2.1)"                                                                             
#>  [74] " onbrand             1.0.3    2023-07-16 [2] CRAN (R 4.2.0)"                                                                             
#>  [75] " openssl             2.0.6    2023-03-09 [2] CRAN (R 4.2.0)"                                                                             
#>  [76] " pillar              1.9.0    2023-03-22 [2] CRAN (R 4.2.0)"                                                                             
#>  [77] " pkgbuild            1.4.0    2022-11-27 [2] CRAN (R 4.2.0)"                                                                             
#>  [78] " pkgconfig           2.0.3    2019-09-22 [2] CRAN (R 4.2.0)"                                                                             
#>  [79] " pkgdown             2.0.7    2022-12-14 [2] CRAN (R 4.2.0)"                                                                             
#>  [80] " pkgload             1.3.2    2022-11-16 [2] CRAN (R 4.2.0)"                                                                             
#>  [81] " plotly              4.10.1   2022-11-07 [2] CRAN (R 4.2.0)"                                                                             
#>  [82] " polyclip            1.10-4   2022-10-20 [2] CRAN (R 4.2.1)"                                                                             
#>  [83] " prettyunits         1.1.1    2020-01-24 [2] CRAN (R 4.2.0)"                                                                             
#>  [84] " processx            3.8.0    2022-10-26 [2] CRAN (R 4.2.0)"                                                                             
#>  [85] " profvis             0.3.7    2020-11-02 [2] CRAN (R 4.2.0)"                                                                             
#>  [86] " promises            1.2.0.1  2021-02-11 [2] CRAN (R 4.2.0)"                                                                             
#>  [87] " prompter            1.1.0    2021-10-07 [2] CRAN (R 4.2.0)"                                                                             
#>  [88] " ps                  1.7.4    2023-04-02 [2] CRAN (R 4.2.1)"                                                                             
#>  [89] " purrr               1.0.1    2023-01-10 [2] CRAN (R 4.2.0)"                                                                             
#>  [90] " R6                  2.5.1    2021-08-19 [2] CRAN (R 4.2.0)"                                                                             
#>  [91] " ragg                1.2.5    2023-01-12 [2] CRAN (R 4.2.0)"                                                                             
#>  [92] " Rcpp                1.0.10   2023-01-22 [2] CRAN (R 4.2.0)"                                                                             
#>  [93] " readr               2.1.4    2023-02-10 [2] CRAN (R 4.2.0)"                                                                             
#>  [94] " readxl              1.4.2    2023-02-09 [2] CRAN (R 4.2.0)"                                                                             
#>  [95] " remotes             2.4.2    2021-11-30 [2] CRAN (R 4.2.0)"                                                                             
#>  [96] " rhandsontable       0.3.8    2021-05-27 [2] CRAN (R 4.2.0)"                                                                             
#>  [97] " rlang               1.1.0    2023-03-14 [2] CRAN (R 4.2.0)"                                                                             
#>  [98] " rmarkdown           2.21     2023-03-26 [2] CRAN (R 4.2.1)"                                                                             
#>  [99] " rprojroot           2.0.3    2022-04-02 [2] CRAN (R 4.2.0)"                                                                             
#> [100] " rstatix             0.7.2    2023-02-01 [2] CRAN (R 4.2.0)"                                                                             
#> [101] " rstudioapi          0.14     2022-08-22 [2] CRAN (R 4.2.0)"                                                                             
#> [102] " sass                0.4.5    2023-01-24 [2] CRAN (R 4.2.0)"                                                                             
#> [103] " scales              1.2.1    2022-08-20 [2] CRAN (R 4.2.0)"                                                                             
#> [104] " sessioninfo         1.2.2    2021-12-06 [2] CRAN (R 4.2.0)"                                                                             
#> [105] " shiny               1.7.4    2022-12-15 [2] CRAN (R 4.2.0)"                                                                             
#> [106] " shinyAce            0.4.2    2022-05-06 [2] CRAN (R 4.2.0)"                                                                             
#> [107] " shinybusy           0.3.1    2022-05-10 [2] CRAN (R 4.2.0)"                                                                             
#> [108] " shinydashboard      0.7.2    2021-09-30 [2] CRAN (R 4.2.0)"                                                                             
#> [109] " shinyWidgets        0.8.0    2023-08-30 [2] CRAN (R 4.2.0)"                                                                             
#> [110] " snakecase           0.11.0   2019-05-25 [2] CRAN (R 4.2.0)"                                                                             
#> [111] " stringi             1.7.12   2023-01-11 [2] CRAN (R 4.2.0)"                                                                             
#> [112] " stringr             1.5.0    2022-12-02 [2] CRAN (R 4.2.0)"                                                                             
#> [113] " systemfonts         1.0.4    2022-02-11 [2] CRAN (R 4.2.0)"                                                                             
#> [114] " textshaping         0.3.6    2021-10-13 [2] CRAN (R 4.2.0)"                                                                             
#> [115] " tibble              3.2.1    2023-03-20 [2] CRAN (R 4.2.0)"                                                                             
#> [116] " tidyr               1.3.0    2023-01-24 [2] CRAN (R 4.2.0)"                                                                             
#> [117] " tidyselect          1.2.0    2022-10-10 [2] CRAN (R 4.2.0)"                                                                             
#> [118] " timechange          0.2.0    2023-01-11 [2] CRAN (R 4.2.0)"                                                                             
#> [119] " tweenr              2.0.2    2022-09-06 [2] CRAN (R 4.2.0)"                                                                             
#> [120] " tzdb                0.3.0    2022-03-28 [2] CRAN (R 4.2.0)"                                                                             
#> [121] " urlchecker          1.0.1    2021-11-30 [2] CRAN (R 4.2.0)"                                                                             
#> [122] " usethis             2.1.6    2022-05-25 [2] CRAN (R 4.2.0)"                                                                             
#> [123] " utf8                1.2.3    2023-01-31 [2] CRAN (R 4.2.0)"                                                                             
#> [124] " uuid                1.1-0    2022-04-19 [2] CRAN (R 4.2.0)"                                                                             
#> [125] " vctrs               0.6.1    2023-03-22 [2] CRAN (R 4.2.0)"                                                                             
#> [126] " viridisLite         0.4.1    2022-08-22 [2] CRAN (R 4.2.0)"                                                                             
#> [127] " whisker             0.4.1    2022-12-05 [2] CRAN (R 4.2.0)"                                                                             
#> [128] " withr               2.5.0    2022-03-03 [2] CRAN (R 4.2.0)"                                                                             
#> [129] " writexl             1.4.2    2023-01-06 [2] CRAN (R 4.2.0)"                                                                             
#> [130] " xfun                0.38     2023-03-24 [2] CRAN (R 4.2.0)"                                                                             
#> [131] " xml2                1.3.3    2021-11-30 [2] CRAN (R 4.2.0)"                                                                             
#> [132] " xtable              1.8-4    2019-04-21 [2] CRAN (R 4.2.0)"                                                                             
#> [133] " yaml                2.3.7    2023-01-23 [2] CRAN (R 4.2.0)"                                                                             
#> [134] " zip                 2.2.2    2022-10-26 [2] CRAN (R 4.2.0)"                                                                             
#> [135] ""                                                                                                                                        
#> [136] " [1] /private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpoDixuG/temp_libpath689117e8dee"                                        
#> [137] " [2] /Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library"