Fetches Informaiton About the App

Returns diagnostic information about the app

Usage

FM_fetch_app_info(session)

Arguments

session

Shiny session variable.

Value

List with information about the app with the following structure

uiele: All system information as UI elements to be used in shiny apps.

uiele_packages: UI element for installed packages to be used in shiny apps.

uiele_options: UI element for current options.

uiele_modules: UI element for loaded formods modules to be used in shiny apps.

modules: List with formods module IDs used in the app as the element names. Each contains details about that module.

msgs: System information as text to be used in a report/terminal.

si_packages Dataframe with currently used packages.

si_options Dataframe with current options

Examples

# We need a Shiny session object to use this function:
sess_res = UD_test_mksession()
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: setting word placeholders: 
#> → ASM:   -> setting docx ph: HEADERLEFT = left header
#> → ASM:   -> setting docx ph: HEADERRIGHT = right header
#> → ASM:   -> setting docx ph: FOOTERLEFT = left footer
#> → ASM: module isgood: TRUE
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
#> → UD: module isgood: TRUE
#> → # preload_complete is good: TRUE ------------------------------------------
session = sess_res$session
app_info  = FM_fetch_app_info(session)
app_info$msgs
#>   [1] "Modules"                                                                                                                                  
#>   [2] "ID: ASM"                                                                                                                                  
#>   [3] "type: ASM"                                                                                                                                
#>   [4] "FM_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpdrB2c1/temp_libpath18a62a764952/formods/templates/formods.yaml"
#>   [5] "MOD_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpdrB2c1/temp_libpath18a62a764952/formods/templates/ASM.yaml"   
#>   [6] "User files: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmp2CtIqP/non_shiny/FM"                                                    
#>   [7] "Log file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmp2CtIqP/non_shiny/FM/formods_log.txt"                                      
#>   [8] "ID: UD"                                                                                                                                   
#>   [9] "type: UD"                                                                                                                                 
#>  [10] "FM_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpdrB2c1/temp_libpath18a62a764952/formods/templates/formods.yaml"
#>  [11] "MOD_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpdrB2c1/temp_libpath18a62a764952/formods/templates/UD.yaml"    
#>  [12] "User files: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmp2CtIqP/non_shiny/FM"                                                    
#>  [13] "Log file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmp2CtIqP/non_shiny/FM/formods_log.txt"                                      
#>  [14] "Package dependencies: janitor, readxl, rio"                                                                                               
#>  [15] " package           * version date (UTC) lib source"                                                                                       
#>  [16] " abind               1.4-8   2024-09-12 [2] CRAN (R 4.5.0)"                                                                               
#>  [17] " askpass             1.2.1   2024-10-04 [2] CRAN (R 4.5.0)"                                                                               
#>  [18] " backports           1.5.0   2024-05-23 [2] CRAN (R 4.5.0)"                                                                               
#>  [19] " brio                1.1.5   2024-04-24 [2] CRAN (R 4.5.0)"                                                                               
#>  [20] " broom               1.0.10  2025-09-13 [2] CRAN (R 4.5.0)"                                                                               
#>  [21] " bslib               0.9.0   2025-01-30 [2] CRAN (R 4.5.0)"                                                                               
#>  [22] " cachem              1.1.0   2024-05-16 [2] CRAN (R 4.5.0)"                                                                               
#>  [23] " car                 3.1-3   2024-09-27 [2] CRAN (R 4.5.0)"                                                                               
#>  [24] " carData             3.0-5   2022-01-06 [2] CRAN (R 4.5.0)"                                                                               
#>  [25] " cellranger          1.1.0   2016-07-27 [2] CRAN (R 4.5.0)"                                                                               
#>  [26] " cli                 3.6.5   2025-04-23 [2] CRAN (R 4.5.0)"                                                                               
#>  [27] " clipr               0.8.0   2022-02-22 [2] CRAN (R 4.5.0)"                                                                               
#>  [28] " covr                3.6.4   2023-11-09 [2] CRAN (R 4.5.0)"                                                                               
#>  [29] " crosstalk           1.2.2   2025-08-26 [2] CRAN (R 4.5.0)"                                                                               
#>  [30] " curl                7.0.0   2025-08-19 [2] CRAN (R 4.5.0)"                                                                               
#>  [31] " data.table          1.17.8  2025-07-10 [2] CRAN (R 4.5.0)"                                                                               
#>  [32] " desc                1.4.3   2023-12-10 [2] CRAN (R 4.5.0)"                                                                               
#>  [33] " devtools            2.4.6   2025-10-03 [2] CRAN (R 4.5.0)"                                                                               
#>  [34] " digest              0.6.39  2025-11-19 [2] CRAN (R 4.5.2)"                                                                               
#>  [35] " downlit             0.4.4   2024-06-10 [2] CRAN (R 4.5.0)"                                                                               
#>  [36] " dplyr               1.1.4   2023-11-17 [2] CRAN (R 4.5.0)"                                                                               
#>  [37] " DT                  0.34.0  2025-09-02 [2] CRAN (R 4.5.0)"                                                                               
#>  [38] " ellipsis            0.3.2   2021-04-29 [2] CRAN (R 4.5.0)"                                                                               
#>  [39] " evaluate            1.0.5   2025-08-27 [2] CRAN (R 4.5.0)"                                                                               
#>  [40] " fansi               1.0.6   2023-12-08 [2] CRAN (R 4.5.0)"                                                                               
#>  [41] " farver              2.1.2   2024-05-13 [2] CRAN (R 4.5.0)"                                                                               
#>  [42] " fastmap             1.2.0   2024-05-15 [2] CRAN (R 4.5.0)"                                                                               
#>  [43] " flextable           0.9.10  2025-08-24 [2] CRAN (R 4.5.0)"                                                                               
#>  [44] " fontawesome         0.5.3   2024-11-16 [2] CRAN (R 4.5.0)"                                                                               
#>  [45] " fontBitstreamVera   0.1.1   2017-02-01 [2] CRAN (R 4.5.0)"                                                                               
#>  [46] " fontLiberation      0.1.0   2016-10-15 [2] CRAN (R 4.5.0)"                                                                               
#>  [47] " fontquiver          0.2.1   2017-02-01 [2] CRAN (R 4.5.0)"                                                                               
#>  [48] " formods           * 0.2.1   2025-12-10 [1] local"                                                                                        
#>  [49] " Formula             1.2-5   2023-02-24 [2] CRAN (R 4.5.0)"                                                                               
#>  [50] " fs                  1.6.6   2025-04-12 [2] CRAN (R 4.5.0)"                                                                               
#>  [51] " gdtools             0.4.4   2025-10-06 [2] CRAN (R 4.5.0)"                                                                               
#>  [52] " generics            0.1.4   2025-05-09 [2] CRAN (R 4.5.0)"                                                                               
#>  [53] " ggforce             0.5.0   2025-06-18 [2] CRAN (R 4.5.0)"                                                                               
#>  [54] " ggplot2             4.0.1   2025-11-14 [2] CRAN (R 4.5.2)"                                                                               
#>  [55] " ggpubr              0.6.2   2025-10-17 [2] CRAN (R 4.5.0)"                                                                               
#>  [56] " ggsignif            0.6.4   2022-10-13 [2] CRAN (R 4.5.0)"                                                                               
#>  [57] " glue                1.8.0   2024-09-30 [2] CRAN (R 4.5.0)"                                                                               
#>  [58] " gtable              0.3.6   2024-10-25 [2] CRAN (R 4.5.0)"                                                                               
#>  [59] " gtools              3.9.5   2023-11-20 [2] CRAN (R 4.5.0)"                                                                               
#>  [60] " here                1.0.2   2025-09-15 [2] CRAN (R 4.5.0)"                                                                               
#>  [61] " htmltools           0.5.8.1 2024-04-04 [2] CRAN (R 4.5.0)"                                                                               
#>  [62] " htmlwidgets         1.6.4   2023-12-06 [2] CRAN (R 4.5.0)"                                                                               
#>  [63] " httpuv              1.6.16  2025-04-16 [2] CRAN (R 4.5.0)"                                                                               
#>  [64] " httr                1.4.7   2023-08-15 [2] CRAN (R 4.5.0)"                                                                               
#>  [65] " httr2               1.2.1   2025-07-22 [2] CRAN (R 4.5.0)"                                                                               
#>  [66] " janitor             2.2.1   2024-12-22 [2] CRAN (R 4.5.0)"                                                                               
#>  [67] " jquerylib           0.1.4   2021-04-26 [2] CRAN (R 4.5.0)"                                                                               
#>  [68] " jsonlite            2.0.0   2025-03-27 [2] CRAN (R 4.5.0)"                                                                               
#>  [69] " knitr               1.50    2025-03-16 [2] CRAN (R 4.5.0)"                                                                               
#>  [70] " labeling            0.4.3   2023-08-29 [2] CRAN (R 4.5.0)"                                                                               
#>  [71] " later               1.4.4   2025-08-27 [2] CRAN (R 4.5.0)"                                                                               
#>  [72] " lattice             0.22-7  2025-04-02 [2] CRAN (R 4.5.1)"                                                                               
#>  [73] " lazyeval            0.2.2   2019-03-15 [2] CRAN (R 4.5.0)"                                                                               
#>  [74] " lifecycle           1.0.4   2023-11-07 [2] CRAN (R 4.5.0)"                                                                               
#>  [75] " lubridate           1.9.4   2024-12-08 [2] CRAN (R 4.5.0)"                                                                               
#>  [76] " magrittr            2.0.4   2025-09-12 [2] CRAN (R 4.5.0)"                                                                               
#>  [77] " MASS                7.3-65  2025-02-28 [2] CRAN (R 4.5.1)"                                                                               
#>  [78] " memoise             2.0.1   2021-11-26 [2] CRAN (R 4.5.0)"                                                                               
#>  [79] " mime                0.13    2025-03-17 [2] CRAN (R 4.5.0)"                                                                               
#>  [80] " officer             0.7.1   2025-11-14 [2] CRAN (R 4.5.2)"                                                                               
#>  [81] " onbrand             1.0.7   2025-07-22 [2] CRAN (R 4.5.0)"                                                                               
#>  [82] " openssl             2.3.4   2025-09-30 [2] CRAN (R 4.5.0)"                                                                               
#>  [83] " otel                0.2.0   2025-08-29 [2] CRAN (R 4.5.0)"                                                                               
#>  [84] " pillar              1.11.1  2025-09-17 [2] CRAN (R 4.5.0)"                                                                               
#>  [85] " pkgbuild            1.4.8   2025-05-26 [2] CRAN (R 4.5.0)"                                                                               
#>  [86] " pkgconfig           2.0.3   2019-09-22 [2] CRAN (R 4.5.0)"                                                                               
#>  [87] " pkgdown             2.1.3   2025-05-25 [2] CRAN (R 4.5.0)"                                                                               
#>  [88] " pkgload             1.4.1   2025-09-23 [2] CRAN (R 4.5.0)"                                                                               
#>  [89] " plotly              4.11.0  2025-06-19 [2] CRAN (R 4.5.0)"                                                                               
#>  [90] " polyclip            1.10-7  2024-07-23 [2] CRAN (R 4.5.0)"                                                                               
#>  [91] " promises            1.5.0   2025-11-01 [2] CRAN (R 4.5.0)"                                                                               
#>  [92] " prompter            1.2.1   2025-07-01 [2] CRAN (R 4.5.0)"                                                                               
#>  [93] " purrr               1.2.0   2025-11-04 [2] CRAN (R 4.5.0)"                                                                               
#>  [94] " R.methodsS3         1.8.2   2022-06-13 [2] CRAN (R 4.5.0)"                                                                               
#>  [95] " R.oo                1.27.1  2025-05-02 [2] CRAN (R 4.5.0)"                                                                               
#>  [96] " R.utils             2.13.0  2025-02-24 [2] CRAN (R 4.5.0)"                                                                               
#>  [97] " R6                  2.6.1   2025-02-15 [2] CRAN (R 4.5.0)"                                                                               
#>  [98] " ragg                1.5.0   2025-09-02 [2] CRAN (R 4.5.0)"                                                                               
#>  [99] " rappdirs            0.3.3   2021-01-31 [2] CRAN (R 4.5.0)"                                                                               
#> [100] " RColorBrewer        1.1-3   2022-04-03 [2] CRAN (R 4.5.0)"                                                                               
#> [101] " Rcpp                1.1.0   2025-07-02 [2] CRAN (R 4.5.0)"                                                                               
#> [102] " readxl              1.4.5   2025-03-07 [2] CRAN (R 4.5.0)"                                                                               
#> [103] " remotes             2.5.0   2024-03-17 [2] CRAN (R 4.5.0)"                                                                               
#> [104] " rex                 1.2.1   2021-11-26 [2] CRAN (R 4.5.0)"                                                                               
#> [105] " rhandsontable       0.3.8   2021-05-27 [2] CRAN (R 4.5.0)"                                                                               
#> [106] " rio                 1.2.4   2025-09-26 [2] CRAN (R 4.5.0)"                                                                               
#> [107] " rlang               1.1.6   2025-04-11 [2] CRAN (R 4.5.0)"                                                                               
#> [108] " rmarkdown           2.30    2025-09-28 [2] CRAN (R 4.5.0)"                                                                               
#> [109] " rprojroot           2.1.1   2025-08-26 [2] CRAN (R 4.5.0)"                                                                               
#> [110] " rstatix             0.7.3   2025-10-18 [2] CRAN (R 4.5.0)"                                                                               
#> [111] " rstudioapi          0.17.1  2024-10-22 [2] CRAN (R 4.5.0)"                                                                               
#> [112] " S7                  0.2.1   2025-11-14 [2] CRAN (R 4.5.2)"                                                                               
#> [113] " sass                0.4.10  2025-04-11 [2] CRAN (R 4.5.0)"                                                                               
#> [114] " scales              1.4.0   2025-04-24 [2] CRAN (R 4.5.0)"                                                                               
#> [115] " sessioninfo         1.2.3   2025-02-05 [2] CRAN (R 4.5.0)"                                                                               
#> [116] " shiny               1.11.1  2025-07-03 [2] CRAN (R 4.5.0)"                                                                               
#> [117] " shinyAce            0.4.4   2025-02-03 [2] CRAN (R 4.5.0)"                                                                               
#> [118] " shinybusy           0.3.3   2024-03-09 [2] CRAN (R 4.5.0)"                                                                               
#> [119] " shinydashboard      0.7.3   2025-04-21 [2] CRAN (R 4.5.0)"                                                                               
#> [120] " shinyWidgets        0.9.0   2025-02-21 [2] CRAN (R 4.5.0)"                                                                               
#> [121] " snakecase           0.11.1  2023-08-27 [2] CRAN (R 4.5.0)"                                                                               
#> [122] " stringi             1.8.7   2025-03-27 [2] CRAN (R 4.5.0)"                                                                               
#> [123] " stringr             1.6.0   2025-11-04 [2] CRAN (R 4.5.0)"                                                                               
#> [124] " systemfonts         1.3.1   2025-10-01 [2] CRAN (R 4.5.0)"                                                                               
#> [125] " testthat            3.2.3   2025-01-13 [2] CRAN (R 4.5.0)"                                                                               
#> [126] " textshaping         1.0.4   2025-10-10 [2] CRAN (R 4.5.0)"                                                                               
#> [127] " tibble              3.3.0   2025-06-08 [2] CRAN (R 4.5.0)"                                                                               
#> [128] " tidyr               1.3.1   2024-01-24 [2] CRAN (R 4.5.0)"                                                                               
#> [129] " tidyselect          1.2.1   2024-03-11 [2] CRAN (R 4.5.0)"                                                                               
#> [130] " timechange          0.3.0   2024-01-18 [2] CRAN (R 4.5.0)"                                                                               
#> [131] " tweenr              2.0.3   2024-02-26 [2] CRAN (R 4.5.0)"                                                                               
#> [132] " usethis             3.2.1   2025-09-06 [2] CRAN (R 4.5.0)"                                                                               
#> [133] " uuid                1.2-1   2024-07-29 [2] CRAN (R 4.5.0)"                                                                               
#> [134] " vctrs               0.6.5   2023-12-01 [2] CRAN (R 4.5.0)"                                                                               
#> [135] " viridisLite         0.4.2   2023-05-02 [2] CRAN (R 4.5.0)"                                                                               
#> [136] " whisker             0.4.1   2022-12-05 [2] CRAN (R 4.5.0)"                                                                               
#> [137] " withr               3.0.2   2024-10-28 [2] CRAN (R 4.5.0)"                                                                               
#> [138] " xfun                0.54    2025-10-30 [2] CRAN (R 4.5.0)"                                                                               
#> [139] " xml2                1.5.0   2025-11-17 [2] CRAN (R 4.5.2)"                                                                               
#> [140] " xtable              1.8-4   2019-04-21 [2] CRAN (R 4.5.0)"                                                                               
#> [141] " yaml                2.3.10  2024-07-26 [2] CRAN (R 4.5.0)"                                                                               
#> [142] " zip                 2.3.3   2025-05-13 [2] CRAN (R 4.5.0)"                                                                               
#> [143] " zoo                 1.8-14  2025-04-10 [2] CRAN (R 4.5.0)"                                                                               
#> [144] ""                                                                                                                                         
#> [145] " [1] /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmpdrB2c1/temp_libpath18a62a764952"                                        
#> [146] " [2] /Library/Frameworks/R.framework/Versions/4.5-arm64/Resources/library"                                                                
#> [147] " * ── Packages attached to the search path."