Fetches Informaiton About the App

Returns diagnostic information about the app

Usage

FM_fetch_app_info(session)

Arguments

session

Shiny session variable.

Value

List with information about the app with the following structure

uiele: All system information as UI elements to be used in shiny apps.

uiele_packages: UI element for installed packages to be used in shiny apps.

uiele_options: UI element for current options.

uiele_modules: UI element for loaded formods modules to be used in shiny apps.

msgs: System information as text to be used in a report/terminal.

si_packages Dataframe with currently used packages.

si_options Dataframe with current options

Examples

# We need a Shiny session object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
session = sess_res$session
app_info  = FM_fetch_app_info(session)
app_info$msgs
#>   [1] "Modules"                                                                                                                                   
#>   [2] "ID: UD"                                                                                                                                    
#>   [3] "type: UD"                                                                                                                                  
#>   [4] "FM_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmprxYYpe/temp_libpath168697ccb096d/formods/templates/formods.yaml"
#>   [5] "MOD_yaml_file: /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmprxYYpe/temp_libpath168697ccb096d/formods/templates/UD.yaml"    
#>   [6] "User files: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//RtmpyVeJZn/non_shiny/FM"                                                     
#>   [7] "Log file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//RtmpyVeJZn/non_shiny/FM/formods_log.txt"                                       
#>   [8] "Package dependencies: janitor, readr, readxl"                                                                                              
#>   [9] " package           * version  date (UTC) lib source"                                                                                       
#>  [10] " abind               1.4-5    2016-07-21 [2] CRAN (R 4.4.0)"                                                                               
#>  [11] " askpass             1.2.0    2023-09-03 [2] CRAN (R 4.4.0)"                                                                               
#>  [12] " backports           1.5.0    2024-05-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [13] " brio                1.1.5    2024-04-24 [2] CRAN (R 4.4.0)"                                                                               
#>  [14] " broom               1.0.6    2024-05-17 [2] CRAN (R 4.4.0)"                                                                               
#>  [15] " bslib               0.8.0    2024-07-29 [2] CRAN (R 4.4.0)"                                                                               
#>  [16] " cachem              1.1.0    2024-05-16 [2] CRAN (R 4.4.0)"                                                                               
#>  [17] " car                 3.1-2    2023-03-30 [2] CRAN (R 4.4.0)"                                                                               
#>  [18] " carData             3.0-5    2022-01-06 [2] CRAN (R 4.4.0)"                                                                               
#>  [19] " cellranger          1.1.0    2016-07-27 [2] CRAN (R 4.4.0)"                                                                               
#>  [20] " cli                 3.6.3    2024-06-21 [2] CRAN (R 4.4.0)"                                                                               
#>  [21] " clipr               0.8.0    2022-02-22 [2] CRAN (R 4.4.0)"                                                                               
#>  [22] " colorspace          2.1-1    2024-07-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [23] " covr                3.6.4    2023-11-09 [2] CRAN (R 4.4.0)"                                                                               
#>  [24] " crayon              1.5.3    2024-06-20 [2] CRAN (R 4.4.0)"                                                                               
#>  [25] " crul                1.4.2    2024-04-09 [2] CRAN (R 4.4.0)"                                                                               
#>  [26] " curl                5.2.1    2024-03-01 [2] CRAN (R 4.4.0)"                                                                               
#>  [27] " data.table          1.15.4   2024-03-30 [2] CRAN (R 4.4.0)"                                                                               
#>  [28] " desc                1.4.3    2023-12-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [29] " devtools            2.4.5    2022-10-11 [2] CRAN (R 4.4.0)"                                                                               
#>  [30] " digest              0.6.36   2024-06-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [31] " downlit             0.4.4    2024-06-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [32] " dplyr               1.1.4    2023-11-17 [2] CRAN (R 4.4.0)"                                                                               
#>  [33] " DT                  0.33     2024-04-04 [2] CRAN (R 4.4.0)"                                                                               
#>  [34] " ellipsis            0.3.2    2021-04-29 [2] CRAN (R 4.4.0)"                                                                               
#>  [35] " evaluate            0.24.0   2024-06-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [36] " fansi               1.0.6    2023-12-08 [2] CRAN (R 4.4.0)"                                                                               
#>  [37] " farver              2.1.2    2024-05-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [38] " fastmap             1.2.0    2024-05-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [39] " flextable           0.9.6    2024-05-05 [2] CRAN (R 4.4.0)"                                                                               
#>  [40] " fontBitstreamVera   0.1.1    2017-02-01 [2] CRAN (R 4.4.0)"                                                                               
#>  [41] " fontLiberation      0.1.0    2016-10-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [42] " fontquiver          0.2.1    2017-02-01 [2] CRAN (R 4.4.0)"                                                                               
#>  [43] " formods           * 0.1.7    2024-09-01 [1] local"                                                                                        
#>  [44] " fs                  1.6.4    2024-04-25 [2] CRAN (R 4.4.0)"                                                                               
#>  [45] " gdtools             0.3.7    2024-03-05 [2] CRAN (R 4.4.0)"                                                                               
#>  [46] " generics            0.1.3    2022-07-05 [2] CRAN (R 4.4.0)"                                                                               
#>  [47] " gfonts              0.2.0    2023-01-08 [2] CRAN (R 4.4.0)"                                                                               
#>  [48] " ggforce             0.4.2    2024-02-19 [2] CRAN (R 4.4.0)"                                                                               
#>  [49] " ggplot2             3.5.1    2024-04-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [50] " ggpubr              0.6.0    2023-02-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [51] " ggsignif            0.6.4    2022-10-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [52] " glue                1.7.0    2024-01-09 [2] CRAN (R 4.4.0)"                                                                               
#>  [53] " gtable              0.3.5    2024-04-22 [2] CRAN (R 4.4.0)"                                                                               
#>  [54] " gtools              3.9.5    2023-11-20 [2] CRAN (R 4.4.0)"                                                                               
#>  [55] " here                1.0.1    2020-12-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [56] " hms                 1.1.3    2023-03-21 [2] CRAN (R 4.4.0)"                                                                               
#>  [57] " htmltools           0.5.8.1  2024-04-04 [2] CRAN (R 4.4.0)"                                                                               
#>  [58] " htmlwidgets         1.6.4    2023-12-06 [2] CRAN (R 4.4.0)"                                                                               
#>  [59] " httpcode            0.3.0    2020-04-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [60] " httpuv              1.6.15   2024-03-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [61] " httr                1.4.7    2023-08-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [62] " janitor             2.2.0    2023-02-02 [2] CRAN (R 4.4.0)"                                                                               
#>  [63] " jquerylib           0.1.4    2021-04-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [64] " jsonlite            1.8.8    2023-12-04 [2] CRAN (R 4.4.0)"                                                                               
#>  [65] " knitr               1.48     2024-07-07 [2] CRAN (R 4.4.0)"                                                                               
#>  [66] " labeling            0.4.3    2023-08-29 [2] CRAN (R 4.4.0)"                                                                               
#>  [67] " later               1.3.2    2023-12-06 [2] CRAN (R 4.4.0)"                                                                               
#>  [68] " lazyeval            0.2.2    2019-03-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [69] " lifecycle           1.0.4    2023-11-07 [2] CRAN (R 4.4.0)"                                                                               
#>  [70] " lubridate           1.9.3    2023-09-27 [2] CRAN (R 4.4.0)"                                                                               
#>  [71] " magrittr            2.0.3    2022-03-30 [2] CRAN (R 4.4.0)"                                                                               
#>  [72] " MASS                7.3-60.2 2024-04-24 [2] local"                                                                                        
#>  [73] " memoise             2.0.1    2021-11-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [74] " mime                0.12     2021-09-28 [2] CRAN (R 4.4.0)"                                                                               
#>  [75] " miniUI              0.1.1.1  2018-05-18 [2] CRAN (R 4.4.0)"                                                                               
#>  [76] " munsell             0.5.1    2024-04-01 [2] CRAN (R 4.4.0)"                                                                               
#>  [77] " officer             0.6.6    2024-05-05 [2] CRAN (R 4.4.0)"                                                                               
#>  [78] " onbrand             1.0.5    2024-02-05 [2] CRAN (R 4.4.0)"                                                                               
#>  [79] " openssl             2.2.0    2024-05-16 [2] CRAN (R 4.4.0)"                                                                               
#>  [80] " pillar              1.9.0    2023-03-22 [2] CRAN (R 4.4.0)"                                                                               
#>  [81] " pkgbuild            1.4.4    2024-03-17 [2] CRAN (R 4.4.0)"                                                                               
#>  [82] " pkgconfig           2.0.3    2019-09-22 [2] CRAN (R 4.4.0)"                                                                               
#>  [83] " pkgdown             2.0.9    2024-04-18 [2] CRAN (R 4.4.0)"                                                                               
#>  [84] " pkgload             1.4.0    2024-06-28 [2] CRAN (R 4.4.0)"                                                                               
#>  [85] " plotly              4.10.4   2024-01-13 [2] CRAN (R 4.4.0)"                                                                               
#>  [86] " polyclip            1.10-7   2024-07-23 [2] CRAN (R 4.4.0)"                                                                               
#>  [87] " profvis             0.3.8    2023-05-02 [2] CRAN (R 4.4.0)"                                                                               
#>  [88] " promises            1.3.0    2024-04-05 [2] CRAN (R 4.4.0)"                                                                               
#>  [89] " prompter            1.2.0    2024-04-08 [2] CRAN (R 4.4.0)"                                                                               
#>  [90] " purrr               1.0.2    2023-08-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [91] " R6                  2.5.1    2021-08-19 [2] CRAN (R 4.4.0)"                                                                               
#>  [92] " ragg                1.3.2    2024-05-15 [2] CRAN (R 4.4.0)"                                                                               
#>  [93] " Rcpp                1.0.13   2024-07-17 [2] CRAN (R 4.4.0)"                                                                               
#>  [94] " readr               2.1.5    2024-01-10 [2] CRAN (R 4.4.0)"                                                                               
#>  [95] " readxl              1.4.3    2023-07-06 [2] CRAN (R 4.4.0)"                                                                               
#>  [96] " remotes             2.5.0    2024-03-17 [2] CRAN (R 4.4.0)"                                                                               
#>  [97] " rex                 1.2.1    2021-11-26 [2] CRAN (R 4.4.0)"                                                                               
#>  [98] " rhandsontable       0.3.8    2021-05-27 [2] CRAN (R 4.4.0)"                                                                               
#>  [99] " rlang               1.1.4    2024-06-04 [2] CRAN (R 4.4.0)"                                                                               
#> [100] " rmarkdown           2.27     2024-05-17 [2] CRAN (R 4.4.0)"                                                                               
#> [101] " rprojroot           2.0.4    2023-11-05 [2] CRAN (R 4.4.0)"                                                                               
#> [102] " rstatix             0.7.2    2023-02-01 [2] CRAN (R 4.4.0)"                                                                               
#> [103] " rstudioapi          0.16.0   2024-03-24 [2] CRAN (R 4.4.0)"                                                                               
#> [104] " sass                0.4.9    2024-03-15 [2] CRAN (R 4.4.0)"                                                                               
#> [105] " scales              1.3.0    2023-11-28 [2] CRAN (R 4.4.0)"                                                                               
#> [106] " sessioninfo         1.2.2    2021-12-06 [2] CRAN (R 4.4.0)"                                                                               
#> [107] " shiny               1.8.1.1  2024-04-02 [2] CRAN (R 4.4.0)"                                                                               
#> [108] " shinyAce            0.4.2    2022-05-06 [2] CRAN (R 4.4.0)"                                                                               
#> [109] " shinybusy           0.3.3    2024-03-09 [2] CRAN (R 4.4.0)"                                                                               
#> [110] " shinydashboard      0.7.2    2021-09-30 [2] CRAN (R 4.4.0)"                                                                               
#> [111] " shinyWidgets        0.8.6    2024-04-24 [2] CRAN (R 4.4.0)"                                                                               
#> [112] " snakecase           0.11.1   2023-08-27 [2] CRAN (R 4.4.0)"                                                                               
#> [113] " stringi             1.8.4    2024-05-06 [2] CRAN (R 4.4.0)"                                                                               
#> [114] " stringr             1.5.1    2023-11-14 [2] CRAN (R 4.4.0)"                                                                               
#> [115] " systemfonts         1.1.0    2024-05-15 [2] CRAN (R 4.4.0)"                                                                               
#> [116] " testthat            3.2.1.1  2024-04-14 [2] CRAN (R 4.4.0)"                                                                               
#> [117] " textshaping         0.4.0    2024-05-24 [2] CRAN (R 4.4.0)"                                                                               
#> [118] " tibble              3.2.1    2023-03-20 [2] CRAN (R 4.4.0)"                                                                               
#> [119] " tidyr               1.3.1    2024-01-24 [2] CRAN (R 4.4.0)"                                                                               
#> [120] " tidyselect          1.2.1    2024-03-11 [2] CRAN (R 4.4.0)"                                                                               
#> [121] " timechange          0.3.0    2024-01-18 [2] CRAN (R 4.4.0)"                                                                               
#> [122] " tweenr              2.0.3    2024-02-26 [2] CRAN (R 4.4.0)"                                                                               
#> [123] " tzdb                0.4.0    2023-05-12 [2] CRAN (R 4.4.0)"                                                                               
#> [124] " urlchecker          1.0.1    2021-11-30 [2] CRAN (R 4.4.0)"                                                                               
#> [125] " usethis             2.2.3    2024-02-19 [2] CRAN (R 4.4.0)"                                                                               
#> [126] " utf8                1.2.4    2023-10-22 [2] CRAN (R 4.4.0)"                                                                               
#> [127] " uuid                1.2-0    2024-01-14 [2] CRAN (R 4.4.0)"                                                                               
#> [128] " vctrs               0.6.5    2023-12-01 [2] CRAN (R 4.4.0)"                                                                               
#> [129] " viridisLite         0.4.2    2023-05-02 [2] CRAN (R 4.4.0)"                                                                               
#> [130] " whisker             0.4.1    2022-12-05 [2] CRAN (R 4.4.0)"                                                                               
#> [131] " withr               3.0.1    2024-07-31 [2] CRAN (R 4.4.0)"                                                                               
#> [132] " writexl             1.5.0    2024-02-09 [2] CRAN (R 4.4.0)"                                                                               
#> [133] " xfun                0.46     2024-07-18 [2] CRAN (R 4.4.0)"                                                                               
#> [134] " xml2                1.3.6    2023-12-04 [2] CRAN (R 4.4.0)"                                                                               
#> [135] " xtable              1.8-4    2019-04-21 [2] CRAN (R 4.4.0)"                                                                               
#> [136] " yaml                2.3.10   2024-07-26 [2] CRAN (R 4.4.0)"                                                                               
#> [137] " zip                 2.3.1    2024-01-27 [2] CRAN (R 4.4.0)"                                                                               
#> [138] ""                                                                                                                                          
#> [139] " [1] /private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/RtmprxYYpe/temp_libpath168697ccb096d"                                        
#> [140] " [2] /Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library"