Fetches Informaiton About the App

Returns diagnostic information about the app

Usage

FM_fetch_app_info(session)

Arguments

session

Shiny session variable.

Value

List with information about the app with the following structure

uiele: All system information as UI elements to be used in shiny apps.

uiele_packages: UI element for installed packages to be used in shiny apps.

uiele_options: UI element for current options.

uiele_modules: UI element for loaded formods modules to be used in shiny apps.

msgs: System information as text to be used in a report/terminal.

si_packages Dataframe with currently used packages.

si_options Dataframe with current options

Examples

# We need a Shiny session object to use this function:
id="UD"
sess_res = UD_test_mksession(session=list(), id=id)
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
session = sess_res$session
app_info  = FM_fetch_app_info(session)
app_info$msgs
#>   [1] "Modules"                                                                                                                                  
#>   [2] "ID: UD"                                                                                                                                   
#>   [3] "type: UD"                                                                                                                                 
#>   [4] "FM_yaml_file: /private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpBlu1TP/temp_libpathcd561a88cca2/formods/templates/formods.yaml"
#>   [5] "MOD_yaml_file: /private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpBlu1TP/temp_libpathcd561a88cca2/formods/templates/UD.yaml"    
#>   [6] "User files: /var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T//Rtmp4cKLQ6/non_shiny/FM"                                                    
#>   [7] "Log file: /var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T//Rtmp4cKLQ6/non_shiny/FM/formods_log.txt"                                      
#>   [8] "Package dependencies: janitor, readr, readxl"                                                                                             
#>   [9] " package           * version date (UTC) lib source"                                                                                       
#>  [10] " abind               1.4-5   2016-07-21 [2] CRAN (R 4.2.0)"                                                                               
#>  [11] " askpass             1.2.0   2023-09-03 [2] CRAN (R 4.2.0)"                                                                               
#>  [12] " backports           1.4.1   2021-12-13 [2] CRAN (R 4.2.0)"                                                                               
#>  [13] " broom               1.0.5   2023-06-09 [2] CRAN (R 4.2.0)"                                                                               
#>  [14] " bslib               0.6.1   2023-11-28 [2] CRAN (R 4.2.3)"                                                                               
#>  [15] " cachem              1.0.8   2023-05-01 [2] CRAN (R 4.2.0)"                                                                               
#>  [16] " car                 3.1-2   2023-03-30 [2] CRAN (R 4.2.1)"                                                                               
#>  [17] " carData             3.0-5   2022-01-06 [2] CRAN (R 4.2.0)"                                                                               
#>  [18] " cellranger          1.1.0   2016-07-27 [2] CRAN (R 4.2.0)"                                                                               
#>  [19] " cli                 3.6.2   2023-12-11 [2] CRAN (R 4.2.3)"                                                                               
#>  [20] " clipr               0.8.0   2022-02-22 [2] CRAN (R 4.2.0)"                                                                               
#>  [21] " colorspace          2.1-0   2023-01-23 [2] CRAN (R 4.2.0)"                                                                               
#>  [22] " crayon              1.5.2   2022-09-29 [2] CRAN (R 4.2.0)"                                                                               
#>  [23] " crul                1.4.0   2023-05-17 [2] CRAN (R 4.2.0)"                                                                               
#>  [24] " curl                5.2.0   2023-12-08 [2] CRAN (R 4.2.3)"                                                                               
#>  [25] " data.table          1.14.10 2023-12-08 [2] CRAN (R 4.2.3)"                                                                               
#>  [26] " desc                1.4.3   2023-12-10 [2] CRAN (R 4.2.3)"                                                                               
#>  [27] " devtools            2.4.5   2022-10-11 [2] CRAN (R 4.2.0)"                                                                               
#>  [28] " digest              0.6.33  2023-07-07 [2] CRAN (R 4.2.0)"                                                                               
#>  [29] " downlit             0.4.3   2023-06-29 [2] CRAN (R 4.2.0)"                                                                               
#>  [30] " dplyr               1.1.4   2023-11-17 [2] CRAN (R 4.2.3)"                                                                               
#>  [31] " DT                  0.31    2023-12-09 [2] CRAN (R 4.2.3)"                                                                               
#>  [32] " ellipsis            0.3.2   2021-04-29 [2] CRAN (R 4.2.0)"                                                                               
#>  [33] " evaluate            0.23    2023-11-01 [2] CRAN (R 4.2.0)"                                                                               
#>  [34] " fansi               1.0.6   2023-12-08 [2] CRAN (R 4.2.3)"                                                                               
#>  [35] " farver              2.1.1   2022-07-06 [2] CRAN (R 4.2.0)"                                                                               
#>  [36] " fastmap             1.1.1   2023-02-24 [2] CRAN (R 4.2.0)"                                                                               
#>  [37] " flextable           0.9.4   2023-10-22 [2] CRAN (R 4.2.0)"                                                                               
#>  [38] " fontBitstreamVera   0.1.1   2017-02-01 [2] CRAN (R 4.2.0)"                                                                               
#>  [39] " fontLiberation      0.1.0   2016-10-15 [2] CRAN (R 4.2.0)"                                                                               
#>  [40] " fontquiver          0.2.1   2017-02-01 [2] CRAN (R 4.2.0)"                                                                               
#>  [41] " formods           * 0.1.4   2023-12-31 [1] local"                                                                                        
#>  [42] " fs                  1.6.3   2023-07-20 [2] CRAN (R 4.2.0)"                                                                               
#>  [43] " gdtools             0.3.5   2023-12-09 [2] CRAN (R 4.2.3)"                                                                               
#>  [44] " generics            0.1.3   2022-07-05 [2] CRAN (R 4.2.0)"                                                                               
#>  [45] " gfonts              0.2.0   2023-01-08 [2] CRAN (R 4.2.0)"                                                                               
#>  [46] " ggforce             0.4.1   2022-10-04 [2] CRAN (R 4.2.0)"                                                                               
#>  [47] " ggplot2             3.4.4   2023-10-12 [2] CRAN (R 4.2.0)"                                                                               
#>  [48] " ggpubr              0.6.0   2023-02-10 [2] CRAN (R 4.2.0)"                                                                               
#>  [49] " ggsignif            0.6.4   2022-10-13 [2] CRAN (R 4.2.0)"                                                                               
#>  [50] " glue                1.6.2   2022-02-24 [2] CRAN (R 4.2.0)"                                                                               
#>  [51] " gtable              0.3.4   2023-08-21 [2] CRAN (R 4.2.0)"                                                                               
#>  [52] " gtools              3.9.5   2023-11-20 [2] CRAN (R 4.2.3)"                                                                               
#>  [53] " here                1.0.1   2020-12-13 [2] CRAN (R 4.2.0)"                                                                               
#>  [54] " hms                 1.1.3   2023-03-21 [2] CRAN (R 4.2.0)"                                                                               
#>  [55] " htmltools           0.5.7   2023-11-03 [2] CRAN (R 4.2.0)"                                                                               
#>  [56] " htmlwidgets         1.6.4   2023-12-06 [2] CRAN (R 4.2.3)"                                                                               
#>  [57] " httpcode            0.3.0   2020-04-10 [2] CRAN (R 4.2.0)"                                                                               
#>  [58] " httpuv              1.6.13  2023-12-06 [2] CRAN (R 4.2.3)"                                                                               
#>  [59] " httr                1.4.7   2023-08-15 [2] CRAN (R 4.2.0)"                                                                               
#>  [60] " janitor             2.2.0   2023-02-02 [2] CRAN (R 4.2.0)"                                                                               
#>  [61] " jquerylib           0.1.4   2021-04-26 [2] CRAN (R 4.2.0)"                                                                               
#>  [62] " jsonlite            1.8.8   2023-12-04 [2] CRAN (R 4.2.3)"                                                                               
#>  [63] " knitr               1.45    2023-10-30 [2] CRAN (R 4.2.0)"                                                                               
#>  [64] " labeling            0.4.3   2023-08-29 [2] CRAN (R 4.2.0)"                                                                               
#>  [65] " later               1.3.2   2023-12-06 [2] CRAN (R 4.2.3)"                                                                               
#>  [66] " lazyeval            0.2.2   2019-03-15 [2] CRAN (R 4.2.0)"                                                                               
#>  [67] " lifecycle           1.0.4   2023-11-07 [2] CRAN (R 4.2.3)"                                                                               
#>  [68] " lubridate           1.9.3   2023-09-27 [2] CRAN (R 4.2.0)"                                                                               
#>  [69] " magrittr            2.0.3   2022-03-30 [2] CRAN (R 4.2.0)"                                                                               
#>  [70] " MASS                7.3-60  2023-05-04 [2] CRAN (R 4.2.0)"                                                                               
#>  [71] " memoise             2.0.1   2021-11-26 [2] CRAN (R 4.2.0)"                                                                               
#>  [72] " mime                0.12    2021-09-28 [2] CRAN (R 4.2.0)"                                                                               
#>  [73] " miniUI              0.1.1.1 2018-05-18 [2] CRAN (R 4.2.0)"                                                                               
#>  [74] " munsell             0.5.0   2018-06-12 [2] CRAN (R 4.2.0)"                                                                               
#>  [75] " officer             0.6.3   2023-10-22 [2] CRAN (R 4.2.0)"                                                                               
#>  [76] " onbrand             1.0.4   2023-12-06 [2] local"                                                                                        
#>  [77] " openssl             2.1.1   2023-09-25 [2] CRAN (R 4.2.0)"                                                                               
#>  [78] " pillar              1.9.0   2023-03-22 [2] CRAN (R 4.2.0)"                                                                               
#>  [79] " pkgbuild            1.4.3   2023-12-10 [2] CRAN (R 4.2.3)"                                                                               
#>  [80] " pkgconfig           2.0.3   2019-09-22 [2] CRAN (R 4.2.0)"                                                                               
#>  [81] " pkgdown             2.0.7   2022-12-14 [2] CRAN (R 4.2.0)"                                                                               
#>  [82] " pkgload             1.3.3   2023-09-22 [2] CRAN (R 4.2.0)"                                                                               
#>  [83] " plotly              4.10.3  2023-10-21 [2] CRAN (R 4.2.0)"                                                                               
#>  [84] " polyclip            1.10-6  2023-09-27 [2] CRAN (R 4.2.0)"                                                                               
#>  [85] " profvis             0.3.8   2023-05-02 [2] CRAN (R 4.2.0)"                                                                               
#>  [86] " promises            1.2.1   2023-08-10 [2] CRAN (R 4.2.0)"                                                                               
#>  [87] " prompter            1.1.0   2021-10-07 [2] CRAN (R 4.2.0)"                                                                               
#>  [88] " purrr               1.0.2   2023-08-10 [2] CRAN (R 4.2.0)"                                                                               
#>  [89] " R6                  2.5.1   2021-08-19 [2] CRAN (R 4.2.0)"                                                                               
#>  [90] " ragg                1.2.7   2023-12-11 [2] CRAN (R 4.2.3)"                                                                               
#>  [91] " Rcpp                1.0.11  2023-07-06 [2] CRAN (R 4.2.0)"                                                                               
#>  [92] " readr               2.1.4   2023-02-10 [2] CRAN (R 4.2.0)"                                                                               
#>  [93] " readxl              1.4.3   2023-07-06 [2] CRAN (R 4.2.0)"                                                                               
#>  [94] " remotes             2.4.2.1 2023-07-18 [2] CRAN (R 4.2.0)"                                                                               
#>  [95] " rhandsontable       0.3.8   2021-05-27 [2] CRAN (R 4.2.0)"                                                                               
#>  [96] " rlang               1.1.2   2023-11-04 [2] CRAN (R 4.2.0)"                                                                               
#>  [97] " rmarkdown           2.25    2023-09-18 [2] CRAN (R 4.2.0)"                                                                               
#>  [98] " rprojroot           2.0.4   2023-11-05 [2] CRAN (R 4.2.0)"                                                                               
#>  [99] " rstatix             0.7.2   2023-02-01 [2] CRAN (R 4.2.0)"                                                                               
#> [100] " rstudioapi          0.15.0  2023-07-07 [2] CRAN (R 4.2.0)"                                                                               
#> [101] " sass                0.4.8   2023-12-06 [2] CRAN (R 4.2.3)"                                                                               
#> [102] " scales              1.3.0   2023-11-28 [2] CRAN (R 4.2.3)"                                                                               
#> [103] " sessioninfo         1.2.2   2021-12-06 [2] CRAN (R 4.2.0)"                                                                               
#> [104] " shiny               1.8.0   2023-11-17 [2] CRAN (R 4.2.3)"                                                                               
#> [105] " shinyAce            0.4.2   2022-05-06 [2] CRAN (R 4.2.0)"                                                                               
#> [106] " shinybusy           0.3.2   2023-11-23 [2] CRAN (R 4.2.3)"                                                                               
#> [107] " shinydashboard      0.7.2   2021-09-30 [2] CRAN (R 4.2.0)"                                                                               
#> [108] " shinyWidgets        0.8.0   2023-08-30 [2] CRAN (R 4.2.0)"                                                                               
#> [109] " snakecase           0.11.1  2023-08-27 [2] CRAN (R 4.2.0)"                                                                               
#> [110] " stringi             1.8.3   2023-12-11 [2] CRAN (R 4.2.3)"                                                                               
#> [111] " stringr             1.5.1   2023-11-14 [2] CRAN (R 4.2.3)"                                                                               
#> [112] " systemfonts         1.0.5   2023-10-09 [2] CRAN (R 4.2.0)"                                                                               
#> [113] " textshaping         0.3.7   2023-10-09 [2] CRAN (R 4.2.0)"                                                                               
#> [114] " tibble              3.2.1   2023-03-20 [2] CRAN (R 4.2.0)"                                                                               
#> [115] " tidyr               1.3.0   2023-01-24 [2] CRAN (R 4.2.0)"                                                                               
#> [116] " tidyselect          1.2.0   2022-10-10 [2] CRAN (R 4.2.0)"                                                                               
#> [117] " timechange          0.2.0   2023-01-11 [2] CRAN (R 4.2.0)"                                                                               
#> [118] " tweenr              2.0.2   2022-09-06 [2] CRAN (R 4.2.0)"                                                                               
#> [119] " tzdb                0.4.0   2023-05-12 [2] CRAN (R 4.2.0)"                                                                               
#> [120] " urlchecker          1.0.1   2021-11-30 [2] CRAN (R 4.2.0)"                                                                               
#> [121] " usethis             2.2.2   2023-07-06 [2] CRAN (R 4.2.0)"                                                                               
#> [122] " utf8                1.2.4   2023-10-22 [2] CRAN (R 4.2.0)"                                                                               
#> [123] " uuid                1.1-1   2023-08-17 [2] CRAN (R 4.2.0)"                                                                               
#> [124] " vctrs               0.6.5   2023-12-01 [2] CRAN (R 4.2.3)"                                                                               
#> [125] " viridisLite         0.4.2   2023-05-02 [2] CRAN (R 4.2.0)"                                                                               
#> [126] " whisker             0.4.1   2022-12-05 [2] CRAN (R 4.2.0)"                                                                               
#> [127] " withr               2.5.2   2023-10-30 [2] CRAN (R 4.2.0)"                                                                               
#> [128] " writexl             1.4.2   2023-01-06 [2] CRAN (R 4.2.0)"                                                                               
#> [129] " xfun                0.41    2023-11-01 [2] CRAN (R 4.2.0)"                                                                               
#> [130] " xml2                1.3.6   2023-12-04 [2] CRAN (R 4.2.3)"                                                                               
#> [131] " xtable              1.8-4   2019-04-21 [2] CRAN (R 4.2.0)"                                                                               
#> [132] " yaml                2.3.8   2023-12-11 [2] CRAN (R 4.2.3)"                                                                               
#> [133] " zip                 2.3.0   2023-04-17 [2] CRAN (R 4.2.0)"                                                                               
#> [134] ""                                                                                                                                         
#> [135] " [1] /private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpBlu1TP/temp_libpathcd561a88cca2"                                        
#> [136] " [2] /Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library"